### 2023-08-02 695193a
inst/tdc: convert raw events.
- change conf / param name 'file' to 'base_configs' and 'base_config'
