+++ mahos/inst/tdc.py
@@ -16,7 +16,6 @@ import os
 import numpy as np
 
 from .instrument import Instrument
-from ..msgs.inst_tdc_msgs import RawEvents
 
 
 def c_str(s: str) -> C.c_char_p:
@@ -28,8 +27,18 @@ def c_str(s: str) -> C.c_char_p:
 class MCS(Instrument):
     """Wrapper Class of DMCSX.dll for Fast ComTec MCS6 / MCS8 Series.
 
-    :param file: mapping from file label to actual file name. used in configure_load_range_bin().
-    :type file: dict[str, str]
+    :param base_configs: Mapping from base config name to actual file name.
+        Used in configure_base_range_bin() etc.
+    :type base_configs: dict[str, str]
+    :param home: (default: "C:\\mcs8x64") Home directory for mcs8 software.
+    :type home: str
+    :param remove_lst: (default: True) Remove *.lst file after loading it.
+    :type remove_lst: bool
+    :param lst_channels: (default: [8, 9]) Collected channels for lst file.
+        Under a setting, default value [8, 9] corresponds to STOP1 and STOP2.
+        As correspondence is unclear after reading the manual (maybe dependent on the setting),
+        it is recommended to inspect the output lst file first.
+    :type lst_channels: list[int]
 
     DLL Modifications
     =================
@@ -139,9 +148,12 @@ class MCS(Instrument):
         self.dll = C.windll.LoadLibrary(fn)
         self.logger.info(f"Loaded {fn}")
 
-        self._file = self.conf.get("file", {})
+        self._base_configs = self.conf.get("base_configs", {})
         self._home = self.conf.get("home", "C:\\mcs8x64")
-        self.logger.debug(f"available set or ctl files: {self._file}")
+        self._save_file_name = None
+        self._remove_lst = self.conf.get("remove_lst", True)
+        self._lst_channels = self.conf.get("lst_channels", [8, 9])
+        self.logger.debug(f"available base config files: {self._base_configs}")
 
     def run_command(self, cmd: str) -> bool:
         """Run a command in MCS Server. Return True on success."""
@@ -232,9 +244,13 @@ class MCS(Instrument):
         self.dll.NewSetting(self.nDev)
 
     def set_save_file_name(self, name: str) -> bool:
+        if not name.endswith(".mpa"):
+            name += ".mpa"
+
         setting = self.get_data_setting()
         setting.filename = name.encode()
         self.set_data_setting(setting)
+        self._save_file_name = name
         return True
 
     def remove_saved_file(self, name: str) -> bool:
@@ -297,10 +313,42 @@ class MCS(Instrument):
         self.dll.StoreMCSSetting(C.byref(setting), self.nDev)
         self.dll.NewSetting(self.nDev)
 
-    def load_lst_file(self, file_name: str) -> RawEvents | None:
-        """Load the *.lst file to get RawEvents data."""
+    def load_raw_events(self) -> list[np.ndarray] | None:
+        if not self._save_file_name:
+            self.logger.error("save file name has not been set.")
+            return None
 
+        file_name = os.path.splitext(self._save_file_name)[0] + ".lst"
         file_name = os.path.join(self._home, file_name)
+
+        ret = self.load_lst_file(file_name)
+
+        if self._remove_lst:
+            self.remove_saved_file(file_name)
+
+        if ret is None:
+            return None
+
+        _, format_info, data = ret
+        return self.convert_raw_events(format_info, data)
+
+    def convert_raw_events(self, format_info, data) -> list[np.ndarray]:
+        cl, ch = format_info["channel"]
+        tl, th = format_info["timedata"]
+        ch_data = (data >> cl) & 2 ** (ch - cl + 1) - 1
+
[…]
