### 2023-08-03 8de6ca5
inst/tdc: add ref clock conf
