### 2023-08-08 716cdd4
confocal: make defaults of params configurable
