+++ mahos/gui/confocal.py
@@ -51,6 +51,7 @@ from ..util.stat import simple_moving_average
 from ..util.timer import FPSCounter
 from .gui_node import GUINode
 from .dialog import save_dialog, load_dialog, export_dialog
+from .param import apply_widgets
 from .common_widget import ClientWidget
 
 
@@ -105,7 +106,7 @@ class scanDialog(QtWidgets.QDialog, Ui_scanDialog):
         zbound,
         roibound,
         pos,
-        scan_capability,
+        params,
         default_path,
         parent=None,
     ):
@@ -157,11 +158,22 @@ class scanDialog(QtWidgets.QDialog, Ui_scanDialog):
             w.valueChanged.connect(self.update_ystep_size)
 
         self.modeBox.clear()
-        for mode in scan_capability:
+        for mode in params["mode"].options():
             self.modeBox.addItem(mode_to_str(mode))
 
         self.modeBox.currentIndexChanged.connect(self.update_modal_boxes)
         self.modeBox.setCurrentIndex(0)
+
+        apply_widgets(
+            params,
+            [
+                ("delay", self.delayBox, 1e3),
+                ("time_window", self.timeBox, 1e3),
+                ("dummy_samples", self.dummysampleBox),
+                ("oversample", self.oversampleBox),
+                ("poll_samples", self.pollsampleBox),
+            ],
+        )
 
         self.update_modal_boxes()
         self.update_xstep_size()
@@ -350,7 +362,7 @@ class scanDialog(QtWidgets.QDialog, Ui_scanDialog):
 class trackDialog(QtWidgets.QDialog, Ui_trackDialog):
     """Dialog for Tracking function."""
 
-    def __init__(self, xbound, ybound, zbound, scan_capability, parent=None):
+    def __init__(self, xbound, ybound, zbound, params, parent=None):
         QtWidgets.QDialog.__init__(self, parent)
         self.setupUi(self)
 
@@ -372,11 +384,22 @@ class trackDialog(QtWidgets.QDialog, Ui_trackDialog):
             w.valueChanged.connect(self.update_yzzstep)
 
         self.modeBox.clear()
-        for mode in scan_capability:
+        for mode in params["mode"].options():
             self.modeBox.addItem(mode_to_str(mode))
 
         self.modeBox.currentIndexChanged.connect(self.update_modal_boxes)
         self.modeBox.setCurrentIndex(0)
+
+        apply_widgets(
+            params,
+            [
+                ("delay", self.delayBox, 1e3),
+                ("time_window", self.timeBox, 1e3),
+                ("dummy_samples", self.dummysampleBox),
+                ("oversample", self.oversampleBox),
+                ("poll_samples", self.pollsampleBox),
+            ],
+        )
 
         self.update_modal_boxes()
 
@@ -1084,7 +1107,7 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
         self._invert = invert
 
         self._finalizing = False
-        self._scan_capability = None
+        self._scan_param_dict = None
         self._prev_pos = None
         self._move_interval_ms = move_interval_ms
 
@@ -1135,8 +1158,8 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
         # update initial GUI state
         self.update_state(status.state, last_state=ConfocalState.IDLE)
 
-        # try to store scan_capability
-        self.get_scan_capability()
+        # try to store scan param dict
+        self.get_scan_param_dict()
 
         self.cli.stateUpdated.connect(self.update_state)
         self.cli.xyImageUpdated.connect(self.XY.update_image)
@@ -1625,8 +1648,8 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
[…]
