### 2023-08-09 6a86e57
confocal: add xmin, xmax etc. in param_dict
