### 2023-08-09 013aed5
gui/confocal: use pyqtgraph.SpinBox in imageDialog
