+++ mahos/gui/ui/imageDialog.py
@@ -1,6 +1,6 @@
 # Form implementation generated from reading ui file 'imageDialog.ui'
 #
-# Created by: PyQt6 UI code generator 6.4.2
+# Created by: PyQt6 UI code generator 6.5.2
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic6 is
 # run again.  Do not edit this file unless you know what you are doing.
@@ -45,9 +45,7 @@ class Ui_imageDialog(object):
         self.sizeyBox.setProperty("value", 6.0)
         self.sizeyBox.setObjectName("sizeyBox")
         self.gridLayout.addWidget(self.sizeyBox, 1, 2, 1, 1)
-        self.minBox = QtWidgets.QDoubleSpinBox(parent=imageDialog)
-        self.minBox.setDecimals(1)
-        self.minBox.setMaximum(100000000.0)
+        self.minBox = SpinBox(parent=imageDialog)
         self.minBox.setObjectName("minBox")
         self.gridLayout.addWidget(self.minBox, 0, 1, 1, 1)
         self.posBox = QtWidgets.QCheckBox(parent=imageDialog)
@@ -80,10 +78,7 @@ class Ui_imageDialog(object):
         self.aspectBox.setChecked(True)
         self.aspectBox.setObjectName("aspectBox")
         self.gridLayout.addWidget(self.aspectBox, 5, 1, 1, 1)
-        self.maxBox = QtWidgets.QDoubleSpinBox(parent=imageDialog)
-        self.maxBox.setDecimals(1)
-        self.maxBox.setMaximum(100000000.0)
-        self.maxBox.setProperty("value", 1000.0)
+        self.maxBox = SpinBox(parent=imageDialog)
         self.maxBox.setObjectName("maxBox")
         self.gridLayout.addWidget(self.maxBox, 0, 2, 1, 1)
         self.label_2 = QtWidgets.QLabel(parent=imageDialog)
@@ -151,3 +146,4 @@ class Ui_imageDialog(object):
         self.label_2.setText(_translate("imageDialog", "Color map"))
         self.label_4.setText(_translate("imageDialog", "Fontsize"))
         self.onlyBox.setText(_translate("imageDialog", "Image only"))
+from pyqtgraph import SpinBox
