+++ mahos/gui/ui/trackDialog.py
@@ -1,6 +1,6 @@
 # Form implementation generated from reading ui file 'trackDialog.ui'
 #
-# Created by: PyQt6 UI code generator 6.5.1
+# Created by: PyQt6 UI code generator 6.5.2
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic6 is
 # run again.  Do not edit this file unless you know what you are doing.
