### 2023-08-09 bfe7f1c
odmr: make 'every' param of PD configurable
- every = True and high pd_rate results in callback calls at
  overwhelming rate, that results in error stopping DAQ.
- Default every is True for the time being,
  but it would be worth considering to change to False.
