+++ mahos/meas/odmr_worker.py
@@ -351,7 +351,7 @@ class Sweeper(Worker):
             "samples": samples,
             "rate": rate,
             "finite": False,
-            "every": True,
+            "every": self._conf.get("every", True),
             "drop_first": drop,
             "gate": True,
             "time_window": time_window,
@@ -381,7 +381,7 @@ class Sweeper(Worker):
             "samples": samples,
             "rate": rate,
             "finite": False,
-            "every": True,
+            "every": self._conf.get("every", True),
             "drop_first": drop,
             "clock_mode": True,
             "oversample": oversamp,
