+++ mahos/gui/ui/odmr.py
@@ -1,6 +1,6 @@
 # Form implementation generated from reading ui file 'odmr.ui'
 #
-# Created by: PyQt6 UI code generator 6.5.1
+# Created by: PyQt6 UI code generator 6.5.2
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic6 is
 # run again.  Do not edit this file unless you know what you are doing.
@@ -153,9 +153,14 @@ class Ui_ODMR(object):
         self.gridLayout_2.addWidget(self.label, 2, 0, 1, 1)
         self.pdrateBox = QtWidgets.QSpinBox(parent=self.measTab)
         self.pdrateBox.setMinimum(1)
-        self.pdrateBox.setMaximum(100)
+        self.pdrateBox.setMaximum(10000)
         self.pdrateBox.setObjectName("pdrateBox")
-        self.gridLayout_2.addWidget(self.pdrateBox, 2, 2, 1, 1)
+        self.gridLayout_2.addWidget(self.pdrateBox, 2, 3, 1, 1)
+        self.bgdelayBox = QtWidgets.QDoubleSpinBox(parent=self.measTab)
+        self.bgdelayBox.setDecimals(4)
+        self.bgdelayBox.setMaximum(1000.0)
+        self.bgdelayBox.setObjectName("bgdelayBox")
+        self.gridLayout_2.addWidget(self.bgdelayBox, 2, 2, 1, 1)
         self.verticalLayout.addLayout(self.gridLayout_2)
         self.pulseLabel = QtWidgets.QLabel(parent=self.measTab)
         self.pulseLabel.setObjectName("pulseLabel")
@@ -188,11 +193,12 @@ class Ui_ODMR(object):
         ODMR.setTabOrder(self.numBox, self.sweepsBox)
         ODMR.setTabOrder(self.sweepsBox, self.mwcontBox)
         ODMR.setTabOrder(self.mwcontBox, self.backgroundBox)
-        ODMR.setTabOrder(self.backgroundBox, self.pdrateBox)
+        ODMR.setTabOrder(self.backgroundBox, self.bgdelayBox)
+        ODMR.setTabOrder(self.bgdelayBox, self.pdrateBox)
         ODMR.setTabOrder(self.pdrateBox, self.cwButton)
-        ODMR.setTabOrder(self.cwButton, self.windowBox)
-        ODMR.setTabOrder(self.windowBox, self.pulseButton)
-        ODMR.setTabOrder(self.pulseButton, self.laserdelayBox)
+        ODMR.setTabOrder(self.cwButton, self.pulseButton)
+        ODMR.setTabOrder(self.pulseButton, self.windowBox)
+        ODMR.setTabOrder(self.windowBox, self.laserdelayBox)
         ODMR.setTabOrder(self.laserdelayBox, self.laserwidthBox)
         ODMR.setTabOrder(self.laserwidthBox, self.mwdelayBox)
         ODMR.setTabOrder(self.mwdelayBox, self.mwwidthBox)
@@ -238,6 +244,8 @@ class Ui_ODMR(object):
         self.label.setText(_translate("ODMR", "Detection"))
         self.pdrateBox.setSuffix(_translate("ODMR", " kHz"))
         self.pdrateBox.setPrefix(_translate("ODMR", "pd rate: "))
+        self.bgdelayBox.setPrefix(_translate("ODMR", "bg delay: "))
+        self.bgdelayBox.setSuffix(_translate("ODMR", " ms"))
         self.pulseLabel.setText(_translate("ODMR", "total window: 0.00 single window: 0.00 (rate: 0.00)"))
[…]
