### 2023-08-09 4521c80
inst/daq: catch DAQException
