### 2023-08-09 41b4b14
inst/dtg: truncate on printing long block
