### 2023-08-09 985b669
gui/odmr: fix bug enabling bgdelayBox
