### 2023-08-10 03d8c1f
inst/pd: add AnalogPD class
