### 2023-08-10 7eafa3c
meas/odmr: add post processing method 'aux'
- the conventional (default) is 'sum' post processing,
  that takes sum for all the PDs.
- 'aux' is (primarily) prepared for lock-in technique to store
  quadrature output for reference.
  For that purpose, pd[0] (pd[1] should be set in-phase (quadrature).
  pd[0] is assigned as ODMRData.data (main data) and pd[1] is
  assigned as ODMRData.aux_data.
