### 2023-08-10 1110d22
add new measurement Qdyne
