### 2023-08-10 f8fc749
qdyne: fix bugs and add remove_raw_data conf
