### 2023-08-10 9d5712b
qdyne: clean up a bit and add docstring
