+++ mahos/meas/qdyne.py
@@ -65,6 +65,13 @@ class Qdyne(BasicMeasNode):
         :type pulser.minimum_block_length: int
         :param pulser.block_base: block base granularity (default: 4)
         :type pulser.block_base: int
+        :param pulser.start_delay: (sec.) delay time before starting PG output. (default: 0.0)
+        :type pulser.start_delay: float
+
+        :param pulser.home_raw_events: (default: "") Home directory to load RawEvents file.
+        :type pulser.home_raw_events: str
+        :param pulser.remove_raw_events: (default: True) Remove RawEvents file after loading it.
+        :type pulser.remove_raw_events: bool
 
         """
 
