+++ mahos/meas/qdyne_worker.py
@@ -16,6 +16,7 @@ import numpy as np
 
 from ..util.timer import IntervalTimer, StopWatch
 from ..util.io import load_h5
+from ..util.conv import real_fft
 from ..msgs import param_msgs as P
 from ..msgs.pulse_msgs import PulsePattern
 from ..msgs.inst_pg_msgs import Block, Blocks
@@ -181,6 +182,14 @@ class QdyneAnalyzer(object):
         data.xdata = np.arange(0, N * T, T)
         data.data = np.zeros(N, dtype=np.uint64)
         C.analyze(data.raw_data, data.xdata, data.data, signal_head, signal_tail)
+        return True
+
+    def fft(self, data: QdyneData) -> bool:
+        if not data.has_data():
+            return False
+        x = data.get_xdata() * data.get_bin()
+        y = data.get_ydata()
+        data.fft_xdata, data.fft_ydata = real_fft(x, y)
         return True
 
 
@@ -366,6 +375,7 @@ class Pulser(Worker):
 
         self.logger.info("Analyzing fetched raw data.")
         self.analyzer.analyze(self.data)
+        self.analyzer.fft(self.data)
         self.logger.info("Finished analyzing raw data.")
 
         if self.data.params.get("remove_raw_data", True):
