### 2023-08-11 afca06a
inst/dtg: repeat start process in case of failure
- DTG sometimes fails to start the sequencer.
  Start() checks if the sequencer actually starts,
  and retry start command on failure
