+++ mahos/inst/dtg/dtg.py
@@ -170,7 +170,24 @@ def gen_sequence(label, subname, Nrep, goto, jumpto, trigger):
 
 
 class DTG5000(VisaInstrument):
-    """Base Class for DTG5000 series."""
+    """Base Class for DTG5000 series.
+
+    :param root_local: Root directory in local computer.
+    :type root_local: str
+    :param root_remote: Root directory in remote DTG.
+    :type root_remote: str
+
+    :param start_delay_sec: (default: 1.0) Delay between output relay on and sequencer start.
+        This is to wait the output relay to stabilize.
+    :type start_delay_sec: float
+    :param start_query_delay_sec: (default: 0.1) Delay for OPC? command after start command.
+    :type start_query_delay_sec: float
+    :param start_loop_delay_sec: (default: 1.0) Delay for each loop of the start loop.
+    :type start_loop_delay_sec: float
+    :param start_loop_num: (default: 20) Max. number of start command repeats in the start loop.
+    :type start_loop_num: int
+
+    """
 
     MAX_BLOCK_NUM = 8000
 
@@ -190,7 +207,10 @@ class DTG5000(VisaInstrument):
         self.CHANNELS = {"0": 0, "1": 1, "2": 2, "3": 3, "4": 4, "5": 5, "6": 6, "7": 7}
         if "channels" in conf:
             self.CHANNELS.update(conf["channels"])
-        self._start_delay_sec = conf.get("start_delay_sec", 0.1)
+        self._start_delay_sec = conf.get("start_delay_sec", 1.0)
+        self._start_query_delay_sec = conf.get("start_query_sec", 0.1)
+        self._start_loop_delay_sec = conf.get("start_loop_delay_sec", 1.0)
+        self._start_loop_num = conf.get("start_loop_num", 20)
 
         # last total block length and offsets.
         self.length = 0
@@ -634,24 +654,6 @@ class DTG5000(VisaInstrument):
         self.logger.info(f"DTG tree prepared. Total length: {self.length}")
         return self._write_load_setup(tree)
 
-    def query_start(self, delay=0.1, maxloop=20) -> bool:
-        """query start until actually run.
-
-        if loop count reaches maxloop, return False.
-
-        """
-
-        start_loop_counter = 0
-        while self.inst.query("TBAS:RST?") != "RUN":
-            self.logger.debug(f"query start DTG #{start_loop_counter}")
-            # TODO: really? call start() iteratively?
-            self.start()
-            self.query_opc(delay=delay)
-            start_loop_counter += 1
-            if start_loop_counter > maxloop:
-                return False
-        return True
-
     def min_block_len(self, freq):
         if not hasattr(self, "MIN_BLOCK_LENGTH"):
             raise NotImplementedError("MIN_BLOCK_LENGTH is not defined")
@@ -672,24 +674,43 @@ class DTG5000(VisaInstrument):
         else:
             return 1
 
-    # Standard API
-
-    def start(self) -> bool:
-        """Set all outputs on and start sequencer.
-
-        Delay can be inserted for waiting output relay to stabilize.
-
-        """
-
+    def start_once(self):
         self.logger.info(
             "Setting output all-on and starting sequencer. Delay {:.1f} sec.".format(
                 self._start_delay_sec
             )
         )
+
         self.inst.write("OUTP:STAT:ALL ON")
         time.sleep(self._start_delay_sec)
         self.inst.write("TBAS:RUN ON")
-        return True
+
+    def get_sequencer_status(self) -> str:
+        return self.inst.query("TBAS:RST?")
+
+    def start_loop(self) -> bool:
+        """repeat start command until it actually starts."""
+
+        for i in range(self._start_loop_num):
+            self.start_once()
[…]
