### 2023-08-11 12fe2a8
tdc/MCS: accept *.ctl for configuration
- loadcnf *.set causes hung-ups in some cases.
- run *.ctl can avoid this.
