### 2023-08-11 0b5ce18
CI: narrow python version (only 3.10)
