### 2023-08-11 4b92c1c
CI: add annotation import and revert py version
