### 2023-08-21 a16e412
qdyne: add tests for QdyneAnalyzer, clean C++ code
