### 2023-08-22 5aef421
docs: add Instrument docs etc.
