+++ mahos/meas/iodmr_io.py
@@ -47,17 +47,21 @@ class IODMRIO(object):
 
         self._executor.shutdown(wait=True)
 
-    def save_data(self, fn, data: IODMRData, params: dict | None = None, note: str = "") -> bool:
+    def save_data(
+        self, filename: str, data: IODMRData, params: dict | None = None, note: str = ""
+    ) -> bool:
         if params is None:
             params = {}
 
-        if isinstance(fn, str) and (fn.endswith(".pkl") or fn.endswith(".pkl.bz2")):
+        if isinstance(filename, str) and (
+            filename.endswith(".pkl") or filename.endswith(".pkl.bz2")
+        ):
             default_compression = None
         else:
             default_compression = "lzf"
 
         return save_pickle_or_h5(
-            fn,
+            filename,
             data,
             IODMRData,
             self.logger,
@@ -67,43 +71,49 @@ class IODMRIO(object):
         )
 
     def save_data_async(
-        self, fn, data: IODMRData, params: dict | None = None, note: str = ""
+        self, filename: str, data: IODMRData, params: dict | None = None, note: str = ""
     ) -> bool:
         """save data asynchronously using ThreadPoolExecutor."""
 
-        self.logger.info(f"Saving (async) {fn}.")
-        self._executor.submit(self.save_data, fn, data, params=params, note=note)
+        self.logger.info(f"Saving (async) {filename}.")
+        self._executor.submit(self.save_data, filename, data, params=params, note=note)
         return True
 
-    def load_data(self, fn) -> IODMRData | None:
-        if isinstance(fn, str) and (fn.endswith(".pkl") or fn.endswith(".pkl.bz2")):
+    def load_data(self, filename: str) -> IODMRData | None:
+        if isinstance(filename, str) and (
+            filename.endswith(".pkl") or filename.endswith(".pkl.bz2")
+        ):
             # try to load twice (bz2 and no compression) for backward compatibility
             # because old files are named *.pkl with bz2 compression.
-            d = load_pickle(fn, IODMRData, self.logger, compression="bz2")
+            d = load_pickle(filename, IODMRData, self.logger, compression="bz2")
             if d is None:
-                d = load_pickle(fn, IODMRData, self.logger, compression=None)
+                d = load_pickle(filename, IODMRData, self.logger, compression=None)
             return d
         else:
-            return load_h5(fn, IODMRData, self.logger)
+            return load_h5(filename, IODMRData, self.logger)
 
     def fit_data(self, data: IODMRData, params: dict) -> IODMRFitResult | None:
         fitter = IODMRFitter(data, self.logger)
         return fitter.fit(params)
 
-    def fit_save_data(self, fn: str, data: IODMRData, params: dict) -> IODMRFitResult | None:
+    def fit_save_data(self, filename: str, data: IODMRData, params: dict) -> IODMRFitResult | None:
         res = self.fit_data(data, params)
         if res is not None:
-            res.save(fn)
+            res.save(filename)
         return res
 
-    def load_fit(self, fn) -> IODMRFitResult | None:
-        res = IODMRFitResult.load(fn)
-        self.logger.info(f"Loaded {fn}.")
+    def load_fit(self, filename: str) -> IODMRFitResult | None:
+        """Load IODMRFitResult from filename. return None if load is failed."""
+
+        res = IODMRFitResult.load(filename)
+        self.logger.info(f"Loaded {filename}.")
         return res
 
     def export_data(
-        self, fn, data: IODMRData | list[IODMRData], params: dict | None = None
+        self, filename: str, data: IODMRData | list[IODMRData], params: dict | None = None
     ) -> bool:
+        """Export an IODMRData."""
+
         if params is None:
             params = {}
 
@@ -117,20 +127,22 @@ class IODMRIO(object):
             )
[…]
