+++ docs/_sources/inst/generated/mahos.inst.daq.AnalogIn.rst.txt
@@ -1,0 +1,36 @@
+﻿mahos.inst.daq.AnalogIn
+=======================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.inst.daq
+
+.. autoclass:: AnalogIn
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~AnalogIn.__init__
+      ~AnalogIn.configure
+      ~AnalogIn.configure_clock
+      ~AnalogIn.configure_on_demand
+      ~AnalogIn.get
+      ~AnalogIn.get_buffer_size
+      ~AnalogIn.get_max_rate
+      ~AnalogIn.pop_all_block
+      ~AnalogIn.pop_all_opt
+      ~AnalogIn.pop_block
+      ~AnalogIn.pop_opt
+      ~AnalogIn.read_on_demand
+   
+   
+
+   
+   
+   
