+++ docs/_sources/inst/generated/mahos.inst.daq_interface.AnalogOutInterface.rst.txt
@@ -1,0 +1,27 @@
+﻿mahos.inst.daq\_interface.AnalogOutInterface
+============================================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.inst.daq_interface
+
+.. autoclass:: AnalogOutInterface
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~AnalogOutInterface.configure_clock
+      ~AnalogOutInterface.configure_command
+      ~AnalogOutInterface.set_output_once
+   
+   
+
+   
+   
+   
