+++ docs/_sources/inst/pg.rst.txt
@@ -1,0 +1,39 @@
+mahos.inst.pg
+=============
+
+Interfaces
+----------
+
+.. automodule:: mahos.inst.pg_interface
+
+.. currentmodule:: mahos.inst.pg_interface
+
+.. autosummary::
+   :toctree: generated
+
+   PGInterface
+
+Instruments
+-----------
+
+.. automodule:: mahos.inst.pg
+
+.. currentmodule:: mahos.inst.pg
+
+.. autosummary::
+   :toctree: generated
+
+   PulseStreamer
+   PulseStreamerDAQTrigger
+
+.. automodule:: mahos.inst.dtg
+
+.. currentmodule:: mahos.inst.dtg
+
+.. autosummary::
+   :toctree: generated
+
+   DTG5000
+   DTG5078
+   DTG5274
+   DTG5334
