+++ docs/_sources/inst/generated/mahos.inst.overlay.confocal_scanner_interface.ConfocalScannerInterface.rst.txt
@@ -1,0 +1,29 @@
+﻿mahos.inst.overlay.confocal\_scanner\_interface.ConfocalScannerInterface
+========================================================================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.inst.overlay.confocal_scanner_interface
+
+.. autoclass:: ConfocalScannerInterface
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ConfocalScannerInterface.configure_scan
+      ~ConfocalScannerInterface.get_capability
+      ~ConfocalScannerInterface.get_line
+      ~ConfocalScannerInterface.get_range
+      ~ConfocalScannerInterface.get_unit
+   
+   
+
+   
+   
+   
