+++ docs/_sources/gui/generated/mahos.gui.hbt.HBTGUI.rst.txt
@@ -1,0 +1,25 @@
+﻿mahos.gui.hbt.HBTGUI
+====================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.gui.hbt
+
+.. autoclass:: HBTGUI
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~HBTGUI.init_widget
+   
+   
+
+   
+   
+   

[…]
