### 2023-08-22 f271e93
docs: fix tag in cli and add links in contrib
