### 2023-08-22 7916c7b
inst/spectrometer: try-except import
- make docs buildable without LightField installation
