### 2023-08-24 31ecadb
examples: add cfm examples with real instruments
