+++ examples/cfm/README.rst
@@ -1,0 +1,8 @@
+CFM
+===
+
+Example configuration files for confocal fluorescence microscope (CFM) system
+with optically detected magnetic resonance (ODMR) for color center research.
+
+- Rather featureful (complex) example: `conf.toml`_.
+- Reduced (simpler) example: `conf_simple.toml`_.
