+++ examples/cfm/conf.toml
@@ -1,0 +1,362 @@
+[global]
+req_timeout_ms = 30_000
+poll_timeout_ms = 100
+
+[localhost.log_broker]
+module = "mahos.node.log_broker"
+class = "LogBroker"
+target = { log = "localhost::log_broker" }
+xpub_endpoint = "tcp://127.0.0.1:5755"
+xsub_endpoint = "tcp://127.0.0.1:5756"
+file = true
+
+[localhost.param_server]
+module = "mahos.node.param_server"
+class = "ParamServer"
+target = { log = "localhost::log_broker" }
+rep_endpoint = "tcp://127.0.0.1:5757"
+pub_endpoint = "tcp://127.0.0.1:5758"
+
+[localhost.server]
+module = "mahos.inst.server"
+class = "InstrumentServer"
+target = { log = "localhost::log_broker" }
+log_level = "DEBUG"
+rep_endpoint = "tcp://127.0.0.1:5759"
+pub_endpoint = "tcp://127.0.0.1:5760"
+
+[localhost.server.instrument.clock]
+module = "daq"
+class = "ClockSource"
+conf = { counter = "/Dev1/Ctr3" }
+
+[localhost.server.instrument.switch]
+module = "daq"
+class = "DigitalOut"
+[localhost.server.instrument.switch.conf]
+# RF(Timing), RF(APD0), RF(APD1), FlipMount(Spectrometer), Flipmount (HBT SPF), AOM Sel
+lines = ["/Dev1/port0/line8", "/Dev1/port0/line9", "/Dev1/port0/line10", "/Dev1/port0/line11", "/Dev1/port0/line12", "/Dev1/port0/line13"]
+[localhost.server.instrument.switch.conf.command]
+confocal =     [0, 0, 0, 1, 1, 0]
+odmr =         [0, 0, 0, 1, 1, 1]
+podmr =        [1, 1, 1, 1, 1, 1]
+hbt =          [1, 1, 1, 1, 0, 0]
+spectroscopy = [0, 0, 0, 0, 1, 0]
+
+[localhost.server.instrument.aom_switch]
+module = "daq"
+class = "DigitalOut"
+[localhost.server.instrument.aom_switch.conf]
+lines = ["/Dev1/port0/line14"]
+
+[localhost.server.instrument.piezo]
+module = "piezo"
+class = "E727_3_USB_AO"
+[localhost.server.instrument.piezo.conf]
+lines = ["/Dev1/ao2", "/Dev1/ao3", "/Dev1/ao1"] # (y, z, x)
+limits_um = [ [0.0, 100.0], [0.0, 100.0], [0.0, 100.0]]
+axes_order = [1, 2, 0] # (y, z, x)
+scale_volt_per_um = 0.2
+offset_um = 50.0
+
+[localhost.server.instrument.pd0]
+module = "pd"
+class = "APDCounter"
+[localhost.server.instrument.pd0.conf]
+counter = "/Dev1/Ctr0"
+source = "/Dev1/PFI0"
+module_no = "SPCM-AQRH-14 45930"
+corr_x_kcps = [20.0, 50.4, 126.2, 318.3, 783.2, 1903.5, 2910.8, 4349.8, 6349.3, 9011.6, 10341.6, 12286.8, 14141.6, 16139.2, 18103.0, 20209.2, 21891.8, 24195.7, 26062.8, 27769.8, 29296.6, 30500.1]
[…]
