+++ examples/cfm/conf_simple.toml
@@ -1,0 +1,237 @@
+[global]
+req_timeout_ms = 30_000
+poll_timeout_ms = 100
+
+[localhost.log_broker]
+module = "mahos.node.log_broker"
+class = "LogBroker"
+target = { log = "localhost::log_broker" }
+xpub_endpoint = "tcp://127.0.0.1:5755"
+xsub_endpoint = "tcp://127.0.0.1:5756"
+file = true
+
+[localhost.param_server]
+module = "mahos.node.param_server"
+class = "ParamServer"
+target = { log = "localhost::log_broker" }
+rep_endpoint = "tcp://127.0.0.1:5757"
+pub_endpoint = "tcp://127.0.0.1:5758"
+
+[localhost.server]
+module = "mahos.inst.server"
+class = "InstrumentServer"
+target = { log = "localhost::log_broker" }
+log_level = "DEBUG"
+rep_endpoint = "tcp://127.0.0.1:5759"
+pub_endpoint = "tcp://127.0.0.1:5760"
+
+[localhost.server.instrument.clock]
+module = "daq"
+class = "ClockSource"
+conf = { counter = "/Dev1/Ctr3" }
+
+[localhost.server.instrument.piezo]
+module = "piezo"
+class = "E727_3_USB_AO"
+[localhost.server.instrument.piezo.conf]
+lines = ["/Dev1/ao2", "/Dev1/ao3", "/Dev1/ao1"] # (y, z, x)
+limits_um = [ [0.0, 100.0], [0.0, 100.0], [0.0, 100.0]]
+axes_order = [1, 2, 0] # (y, z, x)
+scale_volt_per_um = 0.2
+offset_um = 50.0
+
+[localhost.server.instrument.pd0]
+module = "pd"
+class = "APDCounter"
+[localhost.server.instrument.pd0.conf]
+counter = "/Dev1/Ctr0"
+source = "/Dev1/PFI0"
+module_no = "SPCM-AQRH-14 45930"
+corr_x_kcps = [20.0, 50.4, 126.2, 318.3, 783.2, 1903.5, 2910.8, 4349.8, 6349.3, 9011.6, 10341.6, 12286.8, 14141.6, 16139.2, 18103.0, 20209.2, 21891.8, 24195.7, 26062.8, 27769.8, 29296.6, 30500.1]
+corr_y = [1.00, 1.00, 1.00, 1.00, 1.02, 1.05, 1.09, 1.16, 1.26, 1.40, 1.54, 1.63, 1.78, 1.97, 2.21, 2.49, 2.89, 3.30, 3.85, 4.55, 5.43, 6.57]
+dark_cps = 50.0
+
+[localhost.server.instrument.pd1]
+module = "pd"
+class = "APDCounter"
+[localhost.server.instrument.pd1.conf]
+counter = "/Dev1/Ctr1"
+source = "/Dev1/PFI12"
+module_no = "SPCM-AQRH-14 45931"
+corr_x_kcps = [20.4, 51.9, 129.3, 326.3, 797.9, 1941.1, 2970.7, 4424.2, 6433.7, 9135.8, 10477.6, 12422.8, 14279.4, 16257.0, 18225.2, 20324.3, 22027.0, 24291.3, 26133.7, 27834.5, 29340.1, 30530.6]
+corr_y = [1.0, 0.99, 0.99, 0.99, 1.02, 1.05, 1.09, 1.16, 1.26, 1.41, 1.55, 1.64, 1.80, 1.99, 2.23, 2.52, 2.93, 3.34, 3.91, 4.62, 5.52, 6.68]
+dark_cps = 42.0
+
+[localhost.server.instrument.sg]
+module = "sg"
+class = "N5182B"
[…]
