+++ examples/cfm/README.rst
@@ -4,5 +4,5 @@ CFM
 Example configuration files for confocal fluorescence microscope (CFM) system
 with optically detected magnetic resonance (ODMR) for color center research.
 
-- Rather featureful (complex) example: `conf.toml`_.
-- Reduced (simpler) example: `conf_simple.toml`_.
+- Rather featureful (complex) example: ``conf.toml``
+- Reduced (simpler) example: ``conf_simple.toml``
