### 2023-08-25 21ec3dd
cli: add new subcommand 'data note'
