### 2023-08-28 8aaf1b1
change conf key names to improve the consistency
- new convention: directory path keys should be suffixed with '_dir'.
