+++ mahos/inst/piezo.py
@@ -10,7 +10,7 @@ available class(es):
 
 """
 
-import typing as T
+from __future__ import annotations
 from itertools import product
 
 import numpy as np
@@ -27,6 +27,14 @@ class E727_3_USB(Instrument):
 
     :param start_servo: (default: True) if True, start servo on init.
     :type start_servo: bool
+    :param limit_um: (default: [[0.0, 100.0], [0.0, 100.0], [0.0, 100.0)]) limits of positions
+        in command coordinate in the following format:
+        [[xmin, xmax], [ymin, ymax], [zmin, zmax]].
+    :type limit_um: list[tuple[float, float]]
+    :param axes_order: (default: [0, 1, 2]) permutation for piezo axes IDs and (x, y, z).
+        When axes_order is [1, 2, 0], the x axis is piezo axis of index 1,
+        and the y (z) axis is index 2 (0).
+    :type axes_order: tuple[int, int, int]
 
     """
 
@@ -49,19 +57,19 @@ class E727_3_USB(Instrument):
         self.logger.info(f"Connected to {devs[0]}.")
 
     def init_param(self):
-        limits = self.conf.get("limits_um", ((0.0, 100.0), (0.0, 100.0), (0.0, 100.0)))
-        for low, high in limits:
-            if low < 0.0 or high > 100.0:
-                raise ValueError("limits_um is invalid")
+        limit = self.conf.get("limit_um", ((0.0, 100.0), (0.0, 100.0), (0.0, 100.0)))
+        for low, high in limit:
+            if low < 0.0 or high > 200.0:
+                raise ValueError("limit_um is invalid")
         axes_order = self.conf.get("axes_order", [0, 1, 2])
 
         ids = self.dev.qSAI()
         self.AXIS_IDs = [ids[i] for i in axes_order]
         self.AXIS_ID_X, self.AXIS_ID_Y, self.AXIS_ID_Z = self.AXIS_IDs
         self.Axis_to_ID = {Axis.X: self.AXIS_ID_X, Axis.Y: self.AXIS_ID_Y, Axis.Z: self.AXIS_ID_Z}
-        self.limits = {ax: r for ax, r in zip(self.AXIS_IDs, limits)}
+        self.limit = {ax: r for ax, r in zip(self.AXIS_IDs, limit)}
         self.target = {ax: 0.0 for ax in self.AXIS_IDs}
-        self._range = [self.limits[ax] for ax in self.AXIS_IDs]
+        self.range = [self.limit[ax] for ax in self.AXIS_IDs]
 
         self.logger.info(f"got AXIS_IDs: {self.AXIS_IDs}")
 
@@ -92,7 +100,7 @@ class E727_3_USB(Instrument):
             if ax not in self.AXIS_IDs:
                 self.logger.error(f"Unknown Axis ID: {ax}")
                 return False
-            mn, mx = self.limits[ax]
+            mn, mx = self.limit[ax]
             if positions[i] < mn:
                 positions[i] = mn
             if positions[i] > mx:
@@ -114,19 +122,19 @@ class E727_3_USB(Instrument):
     def set_target_Z(self, position) -> bool:
         return self.set_target_pos([position], (self.AXIS_ID_Z,))
 
-    def query_servo(self) -> T.List[bool]:
+    def query_servo(self) -> list[bool]:
         r = self.dev.qSVO(self.AXIS_IDs)
         return [r[ax] for ax in self.AXIS_IDs]
 
-    def query_pos(self) -> T.List[float]:
+    def query_pos(self) -> list[float]:
         r = self.dev.qPOS(axes=self.AXIS_IDs)
         return [r[ax] for ax in self.AXIS_IDs]
 
[…]
