### 2023-08-28 6e11def
examples/cfm: fix config to adapt to previous
