### 2023-08-28 829413e
changelog: fix syntax of nested bullet list
