### 2023-08-29 f9d3d7d
odmr_msgs/ODMRData: allow negative normalize_n
- negative normalize_n takes bottom (instead of top) n data points
  to calc the baseline.
