### 2023-08-29 5f3961d
inst/daq: add confs to adapt to USB-6363
- samples_margin: margin 1 for samples looks bad for USB-6363.
  made the margin configurable.
- write_and_start in piezo: with USB-6363, we should write value and
  start task for AnalogOut.
