### 2023-08-30 e7c65d6
odmr: working on multi channel AnalogIn
