+++ mahos/gui/ui/odmr.py
@@ -161,6 +161,9 @@ class Ui_ODMR(object):
         self.bgdelayBox.setMaximum(1000.0)
         self.bgdelayBox.setObjectName("bgdelayBox")
         self.gridLayout_2.addWidget(self.bgdelayBox, 2, 2, 1, 1)
+        self.sgmodBox = QtWidgets.QCheckBox(parent=self.measTab)
+        self.sgmodBox.setObjectName("sgmodBox")
+        self.gridLayout_2.addWidget(self.sgmodBox, 1, 3, 1, 1)
         self.verticalLayout.addLayout(self.gridLayout_2)
         self.pulseLabel = QtWidgets.QLabel(parent=self.measTab)
         self.pulseLabel.setObjectName("pulseLabel")
@@ -191,7 +194,8 @@ class Ui_ODMR(object):
         ODMR.setTabOrder(self.startBox, self.stopBox)
         ODMR.setTabOrder(self.stopBox, self.numBox)
         ODMR.setTabOrder(self.numBox, self.sweepsBox)
-        ODMR.setTabOrder(self.sweepsBox, self.mwcontBox)
+        ODMR.setTabOrder(self.sweepsBox, self.sgmodBox)
+        ODMR.setTabOrder(self.sgmodBox, self.mwcontBox)
         ODMR.setTabOrder(self.mwcontBox, self.backgroundBox)
         ODMR.setTabOrder(self.backgroundBox, self.bgdelayBox)
         ODMR.setTabOrder(self.bgdelayBox, self.pdrateBox)
@@ -246,6 +250,7 @@ class Ui_ODMR(object):
         self.pdrateBox.setPrefix(_translate("ODMR", "pd rate: "))
         self.bgdelayBox.setPrefix(_translate("ODMR", "bg delay: "))
         self.bgdelayBox.setSuffix(_translate("ODMR", " ms"))
+        self.sgmodBox.setText(_translate("ODMR", "IQ modulation"))
         self.pulseLabel.setText(_translate("ODMR", "total window: 0.00 single window: 0.00 (rate: 0.00)"))
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.measTab), _translate("ODMR", "Main"))
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.fiTab), _translate("ODMR", "Fit"))
