### 2023-08-30 e8bab7c
inst/pd: define get() properly
