### 2023-08-30 99a5a89
inst/daq/AnalogIn: fix bug on multiple lines
