### 2023-08-30 a488527
meas/odmr: fix typo-induced bugs
