### 2023-08-31 c4b8954
odmr: add new param delay
