### 2023-08-31 d81dab7
confocal: accept complex data
