+++ mahos/gui/confocal.py
@@ -665,13 +665,13 @@ class subImage(pg.ImageItem):
 class mainImage(pg.ImageItem):
     """pg.ImageItem with a bit of additonal properties to express pseudo-color scan image."""
 
-    def __init__(self, image: Image, **kargs):
+    def __init__(self, image: Image, complex_conv: str, **kargs):
         pg.ImageItem.__init__(self, image=None, **kargs)
 
         self.direction = image.direction
         self.params = image.params
         self.ident = image.ident
-        self.data = image.image
+        self.data = image.get_image(complex_conv)
 
         self.convert_param_types()
 
@@ -737,12 +737,14 @@ class scanItem(QtCore.QObject):
         histo,
         pos,
         style: dict,
+        complex_conv: str,
         roi_size=(20.0, 20.0),
     ):
         QtCore.QObject.__init__(self)
         self.xmin, self.xmax = xbound
         self.ymin, self.ymax = ybound
         self.histo = histo
+        self.complex_conv = complex_conv
 
         self.direction = direction
         self.pi = pi
@@ -803,7 +805,7 @@ class scanItem(QtCore.QObject):
         im.finalize(False)
         im.image = bg_data
 
-        self.img_bg = mainImage(im)
+        self.img_bg = mainImage(im, self.complex_conv)
         self.img_bg.setZValue(-1.0)
 
         self.img = self.img_bg
@@ -908,6 +910,9 @@ class scanItem(QtCore.QObject):
         self.move_crosshair_Y(y)
         self.sigYChanged.emit(y)
 
+    def set_complex_conv(self, conv: str):
+        self.complex_conv = conv
+
     def move_crosshair(self, point):
         if hasattr(point, "x") and hasattr(point, "y"):
             x, y = point.x(), point.y()
@@ -1002,7 +1007,7 @@ class scanItem(QtCore.QObject):
         self._update_image()
 
     def _append(self, image: Image):
-        self.img = mainImage(image)
+        self.img = mainImage(image, self.complex_conv)
 
         self.img_list.append(self.img)
         self.img.setZValue(float(self.img_list.index(self.img)))
@@ -1015,7 +1020,7 @@ class scanItem(QtCore.QObject):
             self._append(image)
             self.fit_roi()
         else:
-            self.img.data = image.image
+            self.img.data = image.get_image(self.complex_conv)
         self._update_image()
 
     def _set_levels(self):
@@ -1224,6 +1229,7 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
             self.histo_XY,
             (x, y),
             self._style,
+            self.complexBox.currentText(),
         )
         self.XZ = scanItem(
             ScanDirection.XZ,
@@ -1234,6 +1240,7 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
             self.histo_XZ,
             (x, z),
             self._style,
+            self.complexBox.currentText(),
         )
         self.YZ = scanItem(
             ScanDirection.YZ,
@@ -1244,6 +1251,7 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
             self.histo_YZ,
             (y, z),
             self._style,
+            self.complexBox.currentText(),
         )
 
     def init_splitter(self):
@@ -1343,6 +1351,10 @@ class ConfocalWidget(ClientWidget, Ui_Confocal):
 
         self.exportviewButton.clicked.connect(self.export_view)
         self.loadButton.clicked.connect(self.load_image)
+
+        self.complexBox.currentTextChanged.connect(self.XY.set_complex_conv)
+        self.complexBox.currentTextChanged.connect(self.XZ.set_complex_conv)
[…]
