+++ mahos/gui/ui/confocal.py
@@ -1,6 +1,6 @@
 # Form implementation generated from reading ui file 'confocal.ui'
 #
-# Created by: PyQt6 UI code generator 6.5.1
+# Created by: PyQt6 UI code generator 6.5.2
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic6 is
 # run again.  Do not edit this file unless you know what you are doing.
@@ -75,6 +75,13 @@ class Ui_Confocal(object):
         self.centerButton = QtWidgets.QPushButton(parent=Confocal)
         self.centerButton.setObjectName("centerButton")
         self.horizontalLayout.addWidget(self.centerButton)
+        self.complexBox = QtWidgets.QComboBox(parent=Confocal)
+        self.complexBox.setObjectName("complexBox")
+        self.complexBox.addItem("")
+        self.complexBox.addItem("")
+        self.complexBox.addItem("")
+        self.complexBox.addItem("")
+        self.horizontalLayout.addWidget(self.complexBox)
         spacerItem1 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Minimum)
         self.horizontalLayout.addItem(spacerItem1)
         self.label_4 = QtWidgets.QLabel(parent=Confocal)
@@ -203,25 +210,33 @@ class Ui_Confocal(object):
         Confocal.setTabOrder(self.xyscanButton, self.xzscanButton)
         Confocal.setTabOrder(self.xzscanButton, self.yzscanButton)
         Confocal.setTabOrder(self.yzscanButton, self.loadButton)
-        Confocal.setTabOrder(self.loadButton, self.sethomeButton)
+        Confocal.setTabOrder(self.loadButton, self.exportviewButton)
+        Confocal.setTabOrder(self.exportviewButton, self.sethomeButton)
         Confocal.setTabOrder(self.sethomeButton, self.gohomeButton)
         Confocal.setTabOrder(self.gohomeButton, self.centerButton)
-        Confocal.setTabOrder(self.centerButton, self.trackstartButton)
+        Confocal.setTabOrder(self.centerButton, self.complexBox)
+        Confocal.setTabOrder(self.complexBox, self.trackstartButton)
         Confocal.setTabOrder(self.trackstartButton, self.trackstopButton)
         Confocal.setTabOrder(self.trackstopButton, self.tracknowButton)
         Confocal.setTabOrder(self.tracknowButton, self.xyfitButton)
         Confocal.setTabOrder(self.xyfitButton, self.xymaxButton)
         Confocal.setTabOrder(self.xymaxButton, self.xypopButton)
         Confocal.setTabOrder(self.xypopButton, self.xyclearButton)
-        Confocal.setTabOrder(self.xyclearButton, self.xzfitButton)
+        Confocal.setTabOrder(self.xyclearButton, self.xysaveButton)
+        Confocal.setTabOrder(self.xysaveButton, self.xyexportButton)
+        Confocal.setTabOrder(self.xyexportButton, self.xzfitButton)
         Confocal.setTabOrder(self.xzfitButton, self.xzmaxButton)
         Confocal.setTabOrder(self.xzmaxButton, self.xzclearButton)
         Confocal.setTabOrder(self.xzclearButton, self.xzpopButton)
-        Confocal.setTabOrder(self.xzpopButton, self.yzfitButton)
+        Confocal.setTabOrder(self.xzpopButton, self.xzsaveButton)
+        Confocal.setTabOrder(self.xzsaveButton, self.xzexportButton)
+        Confocal.setTabOrder(self.xzexportButton, self.yzfitButton)
[…]

[…]
