### 2023-08-31 69c6abd
confocal: debug to accept complex data
