### 2023-09-01 2cfe095
adding new meas/tweaker and inst/lockin
- ParamDict is being applied for inst too.
- This is useful for several instruments involving manual tweaking
  (such as amplifiers with manual gain setting, thormostats,
   magnet for static field generation, etc.)
