+++ mahos/inst/instrument.py
@@ -11,6 +11,7 @@ from __future__ import annotations
 import typing as T
 
 from ..node.log import init_topic_logger
+from ..msgs import param_msgs as P
 
 
 class Instrument(object):
@@ -142,7 +143,7 @@ class Instrument(object):
         self.logger.error("reset() is called but not implemented.")
         return False
 
-    def configure(self, params: dict) -> bool:
+    def configure(self, params: dict, name: str = "", group: str = "") -> bool:
         """Configure the instrument settings. Returns True on success."""
 
         self.logger.error("configure() is called but not implemented.")
@@ -159,3 +160,17 @@ class Instrument(object):
 
         self.logger.error("get() is called but not implemented.")
         return None
+
+    def get_param_dict(
+        self, name: str = "", group: str = ""
+    ) -> P.ParamDict[str, P.PDValue] | None:
+        """Get ParamDict for `name` in `group`."""
+
+        self.logger.error("get_param_dict() is called but not implemented.")
+        return None
+
+    def get_param_dict_names(self, group: str) -> list[str]:
+        """Get list of names of available ParamDicts pertaining to `group`."""
+
+        self.logger.error("get_param_dict_names() is called but not implemented.")
+        return []
