+++ mahos/inst/overlay/confocal_scanner.py
@@ -134,7 +134,7 @@ class ConfocalScannerAnalog(InstrumentOverlay, ConfocalScannerMixin):
     def close(self):
         self.stop()
 
-    def configure(self, params: dict) -> bool:
+    def configure(self, params: dict, name: str = "", group: str = "") -> bool:
         if not self.check_required_params(
             params,
             (
@@ -425,7 +425,7 @@ class ConfocalScannerCommand(InstrumentOverlay, ConfocalScannerMixin):
 
     # Standard API
 
-    def configure(self, params: dict) -> bool:
+    def configure(self, params: dict, name: str = "", group: str = "") -> bool:
         if not self.check_required_params(
             params,
             (
