+++ mahos/inst/overlay/confocal_scanner_mock.py
@@ -37,7 +37,7 @@ class ConfocalScanner_mock(InstrumentOverlay):
 
     # Standard API
 
-    def configure(self, params: dict) -> bool:
+    def configure(self, params: dict, name: str = "", group: str = "") -> bool:
         self.params = params
 
         req_keys = (
