+++ mahos/inst/pg.py
@@ -188,7 +188,7 @@ class PulseStreamer(Instrument):
         self.logger.info("Stop streaming and outputting constant all-zero.")
         return True
 
-    def configure(self, params: dict) -> bool:
+    def configure(self, params: dict, name: str = "", group: str = "") -> bool:
         if "blocks" in params and "freq" in params:
             return self.configure_blocks(
                 params["blocks"],
