+++ mahos/inst/pd.py
@@ -227,8 +227,123 @@ class LUCI10(Instrument):
             return None
 
 
-class OE200(AnalogIn):
-    """FEMTO Messtechnik OE-200 Variable Gain Photoreceiver.
+class LUCI_OE200(LUCI10):
+    """LUCI-10 for FEMTO Messtechnik OE-200 Variable Gain Photoreceiver.
+
+    This class requires LUCI-10 installation (see LUCI10 class).
+
+    :param dll_dir: The directory path containing LUCI-10 DLL.
+    :type dll_dir: str
+    :param index: (default: 1) Index of LUCI-10 device.
+    :type index: int
+
+    :param DC_coupling: (default: True) Init setting. True (False) for DC (AC) coupling.
+    :type DC_coupling: bool
+    :param low_noise: (default: True) Init setting. True (False) for low_noise (high_speed) mode.
+    :type low_noise: bool
+    :param gain_exponent: (default: 3) Init setting. The gain exponent.
+    :type gain_exponent: int
+
+    """
+
+    GAINS = {
+        True: {3: 0, 4: 1, 5: 2, 6: 3, 7: 4, 8: 5, 9: 6},
+        False: {5: 0, 6: 1, 7: 2, 8: 3, 9: 4, 10: 5, 11: 6},
+    }
+
+    def __init__(self, name, conf, prefix=None):
+        LUCI10.__init__(self, name, conf=conf, prefix=prefix)
+
+        self.DC_coupling = self.conf.get("DC_coupling", True)
+        low_noise = self.conf.get("low_noise", True)
+        gain_exponent = self.conf.get("gain_exponent", 3)
+        if not self.set_gain(low_noise, gain_exponent):
+            raise ValueError("Initial setting failed.")
+
+    def _write_settings(self) -> bool:
+        if self.DC_coupling:
+            to_write = self.gain_value | 0b1000
+        else:
+            to_write = self.gain_value
+
+        self.logger.debug(f"Writing LUCI-10 setting: 0b{to_write:b}")
+        return self.write_data(to_write)
+
+    def _update_gain(self, low_noise: bool, gain_exponent: int) -> bool:
+        self.low_noise = low_noise
+        self.gain_exponent = gain_exponent
+
+        try:
+            self.gain_value = self.GAINS[low_noise][gain_exponent]
+        except KeyError:
+            return self.fail_with(f"Invalid arguments: {low_noise}, {gain_exponent}")
+
+        if low_noise:
+            self.gain_value |= 0b1_0000
+
+        self.gain = 10**gain_exponent
+        return True
+
+    def set_gain(self, low_noise: bool, gain_exponent: int) -> bool:
+        return self._update_gain(low_noise, gain_exponent) and self._write_settings()
+
+    def set_coupling(self, DC_coupling: bool) -> bool:
+        self.DC_coupling = DC_coupling
+        return self._write_settings()
+
+    def configure_gain_coupling(
+        self, low_noise: bool, gain_exponent: int, DC_coupling: bool
+    ) -> bool:
+        self.DC_coupling = DC_coupling
+        return self._update_gain(low_noise, gain_exponent) and self._write_settings()
+
+    def set(self, key: str, value=None) -> bool:
+        key = key.lower()
+        if key == "led":
+            return self.set_led(value)
+        elif key == "gain":
+            try:
+                return self.set_gain(value["low_noise"], value["gain_exponent"])
+            except (KeyError, TypeError):
+                return self.fail_with("value must be a dict with low_noise and gain_exponent.")
+        elif key == "coupling":
+            return self.set_coupling(bool(value))
+        else:
+            return self.fail_with(f"Unknown set() key: {key}.")
+
[…]
