### 2023-09-01 1fded62
inst/interface: fix bug not following API change
