### 2023-09-01 a38a913
rename args or attrs
- co-existence of `name` and `pd_name` looked quite confusing
  `name` is `inst`, and `pd_name` is just `name`.
- tweaker: param_dict_names -> param_dict_ids. the latter is
  single-str expression like inst.group.name.
