+++ mahos/inst/server.py
@@ -48,47 +48,47 @@ class OverlayConf(object):
 
 
 class Locks(object):
-    def __init__(self, insts):
+    def __init__(self, insts: list[str]):
         self.insts = insts
         self.locks = {}
         self.overlay_deps = {}
 
-        for name in insts:
-            self.locks[name] = None
+        for inst in insts:
+            self.locks[inst] = None
 
-    def add_overlay(self, name: str, inst_names: tuple):
-        self.overlay_deps[name] = inst_names
+    def add_overlay(self, inst: str, inst_names: tuple):
+        self.overlay_deps[inst] = inst_names
 
-    def is_locked(self, name: str, ident=None, reduce_=any) -> bool:
+    def is_locked(self, inst: str, ident=None, reduce_=any) -> bool:
         def inst_is_locked(n):
             if ident is not None:
                 return self.locks[n] is not None and self.locks[n] != ident
             else:
                 return self.locks[n] is not None
 
-        if name in self.insts:
-            return inst_is_locked(name)
-        else:  # assert name in self.overlay
-            return reduce_([inst_is_locked(n) for n in self.overlay_deps[name]])
+        if inst in self.insts:
+            return inst_is_locked(inst)
+        else:  # assert inst in self.overlay
+            return reduce_([inst_is_locked(n) for n in self.overlay_deps[inst]])
 
-    def locked_by(self, name):
-        if name in self.insts:
-            return self.locks[name]
-        else:  # assert name in self.overlay
-            return tuple((self.locks[n] for n in self.overlay_deps[name]))
+    def locked_by(self, inst: str):
+        if inst in self.insts:
+            return self.locks[inst]
+        else:  # assert inst in self.overlay
+            return tuple((self.locks[n] for n in self.overlay_deps[inst]))
 
-    def lock(self, name: str, ident: Ident):
-        if name in self.insts:
-            self.locks[name] = ident
-        else:  # assert name in self.overlay
-            for n in self.overlay_deps[name]:
+    def lock(self, inst: str, ident: Ident):
+        if inst in self.insts:
+            self.locks[inst] = ident
+        else:  # assert inst in self.overlay
+            for n in self.overlay_deps[inst]:
                 self.locks[n] = ident
 
-    def release(self, name: str, ident: Ident):
-        if name in self.insts:
-            self.locks[name] = None
-        else:  # assert name in self.overlay
-            for n in self.overlay_deps[name]:
+    def release(self, inst: str, ident: Ident):
+        if inst in self.insts:
+            self.locks[inst] = None
+        else:  # assert inst in self.overlay
+            for n in self.overlay_deps[inst]:
                 self.locks[n] = None
 
 
@@ -118,35 +118,35 @@ class InstrumentClient(StatusClient):
         lay = self.conf["instrument_overlay"] if "instrument_overlay" in self.conf else {}
         self.inst_names = tuple(inst.keys()) + tuple(lay.keys())
 
-    def lock(self, name: str) -> bool:
+    def lock(self, inst: str) -> bool:
         """Acquire lock of an instrument."""
 
-        resp = self.req.request(LockReq(self.ident, name))
+        resp = self.req.request(LockReq(self.ident, inst))
 
         if not resp.success:
             self.logger.error(resp.message)
         return resp.success
 
-    def is_locked(self, name: str) -> T.Optional[bool]:
+    def is_locked(self, inst: str) -> T.Optional[bool]:
         """Check if an instrument is locked."""
 
-        resp = self.req.request(CheckLockReq(self.ident, name))
+        resp = self.req.request(CheckLockReq(self.ident, inst))
 
         if not resp.success:
             self.logger.error(resp.message)
             return None
         return resp.ret
 
-    def release(self, name: str) -> bool:
+    def release(self, inst: str) -> bool:
[…]
