### 2023-09-02 a730c85
rename: ParamDict name(s) -> ParamDict label(s)
Since 'name' for param dict sounds quite conflicting (to Node 'name'), it is changed to 'label'.
Now, quite generally, ParamDict is identified by a 'group' (if any), and 'label'.

Signature / name changes:
- configure() of Instrument
- get_param_dict() and get_param_dict_names() of Instrument and meas nodes.
