### 2023-09-04 32d301b
tweaker: fix io and make it usable from other meas
- use h5 instead of pickle, that enables injection of
  tweaker-managed inst_params to measurement Data files.
- added test_tweaker too.
