+++ mahos/meas/tweaker.py
@@ -1,25 +1,28 @@
 #!/usr/bin/env python3
 
 """
-Tweaker for manually-tuned instrument parameters.
+Tweaker for manually-tuned Instrument's ParamDicts.
 
 .. This file is a part of MAHOS project.
 
 """
 
 from __future__ import annotations
-import pickle
+import os
+
+import h5py
 
 from ..msgs.common_msgs import Resp
 from ..msgs import param_msgs as P
 from ..msgs import tweaker_msgs
-from ..msgs.tweaker_msgs import TweakerStatus, ReadReq, ReadAllReq, WriteReq, SaveReq, LoadReq
+from ..msgs.tweaker_msgs import TweakerStatus, ReadReq, ReadAllReq, WriteReq, WriteAllReq
+from ..msgs.tweaker_msgs import SaveReq, LoadReq
 from ..node.node import Node
 from ..node.client import StatusClient
 from ..inst.server import MultiInstrumentClient
 
 
-PARAM_DICT_ID_DELIM = "."
+PARAM_DICT_ID_DELIM = "::"
 
 
 class TweakerClient(StatusClient):
@@ -27,32 +30,37 @@ class TweakerClient(StatusClient):
 
     M = tweaker_msgs
 
-    def read_all(self) -> dict[str, P.ParamDict[str, P.PDValue] | None] | None:
+    def read_all(self) -> dict[str, P.ParamDict[str, P.PDValue] | None]:
         resp = self.req.request(ReadAllReq())
-        if resp.success:
-            return resp.ret
+        return resp.ret
 
     def read(self, param_dict_id: str) -> P.ParamDict[str, P.PDValue] | None:
         resp = self.req.request(ReadReq(param_dict_id))
         if resp.success:
             return resp.ret
 
+    def write_all(self, param_dicts: dict[str, P.ParamDict[str, P.PDValue]]) -> bool:
+        resp = self.req.request(WriteAllReq(param_dicts))
+        return resp.success
+
     def write(self, param_dict_id: str, params: P.ParamDict[str, P.PDValue]) -> bool:
         resp = self.req.request(WriteReq(param_dict_id, params))
         return resp.success
 
-    def save(self, filename: str) -> bool:
-        resp = self.req.request(SaveReq(filename))
+    def save(self, filename: str, group: str = "") -> bool:
+        resp = self.req.request(SaveReq(filename, group))
         return resp.success
 
-    def load(self, filename: str) -> dict[str, P.ParamDict[str, P.PDValue] | None] | None:
-        resp = self.req.request(LoadReq(filename))
+    def load(
+        self, filename: str, group: str = ""
+    ) -> dict[str, P.ParamDict[str, P.PDValue] | None] | None:
+        resp = self.req.request(LoadReq(filename, group))
         if resp.success:
             return resp.ret
 
 
 class Tweaker(Node):
-    """Tweaker for manually-tuned instrument parameters."""
+    """Tweaker for manually-tuned Instrument's ParamDicts."""
 
     CLIENT = TweakerClient
 
@@ -63,67 +71,90 @@ class Tweaker(Node):
             gconf, self.conf["target"]["servers"], context=self.ctx, prefix=self.joined_name()
         )
 
-        self._param_dict_ids = self.conf["param_dicts"]
-        self._param_dicts = {pd: None for pd in self._param_dict_ids}
+        #: dict[param_dict_id, ParamDict | None]
+        self._param_dicts = {pd: None for pd in self.conf["param_dicts"]}
 
         self.add_rep()
         self.status_pub = self.add_pub(b"status")
 
-    def _parse_param_dict_id(self, param_dict_id: str) -> tuple[str, str, str]:
[…]
