+++ mahos/meas/spectroscopy.py
@@ -14,6 +14,7 @@ from ..msgs.param_msgs import GetParamDictReq, GetParamDictLabelsReq
 from ..msgs import spectroscopy_msgs
 from ..msgs.spectroscopy_msgs import SpectroscopyData
 from ..util.timer import IntervalTimer
+from .tweaker import TweakerClient
 from .common_meas import BasicMeasClient, BasicMeasNode
 from .common_worker import DummyWorker, PulseGen_CW, Switch
 from .spectroscopy_worker import Repeater
@@ -43,6 +44,12 @@ class Spectroscopy(BasicMeasNode):
             self.pg = PulseGen_CW(self.cli, self.logger)
         else:
             self.pg = DummyWorker()
+        if "tweaker" in self.conf["target"]:
+            self.tweaker_cli = TweakerClient(
+                gconf, self.conf["target"]["tweaker"], context=self.ctx, prefix=self.joined_name()
+            )
+        else:
+            self.tweaker_cli = None
 
         self.worker = Repeater(self.cli, self.logger, self.conf.get("repeater", {}))
         self.fitter = SpectroscopyFitter(self.logger)
@@ -99,6 +106,8 @@ class Spectroscopy(BasicMeasNode):
 
     def save_data(self, msg: SaveDataReq) -> Resp:
         success = self.io.save_data(msg.file_name, self.worker.data_msg(), msg.note)
+        if success and self.tweaker_cli is not None:
+            success &= self.tweaker_cli.save(msg.file_name, "_inst_params")
         return Resp(success)
 
     def export_data(self, msg: ExportDataReq) -> Resp:
