### 2023-09-05 58a221b
inst/server: allow Overlay that depends on Overlay
- now overlay's can resolve reference including other overlays
- for that, added dependency-resolver (topological sort on DAG) utility in util.graph.
