### 2023-09-05 430835b
test: add test for overlay resolver
