### 2023-09-05 4bc34bc
inst: fix LI5640 and add amplified pd classes
