+++ mahos/inst/overlay/amplified_pd.py
@@ -12,10 +12,88 @@ import numpy as np
 from .overlay import InstrumentOverlay
 from ..msgs import param_msgs as P
 from ..lockin import LI5640
-from ..pd import LUCI_OE200, LockinAnalogPD
+from ..pd import LUCI_OE200, AnalogPD, LockinAnalogPD
 
 
-class OE200_LI5640(InstrumentOverlay):
+class OE200_AI(InstrumentOverlay):
+    """FEMTO Messtechnik OE-200 Variable Gain Photoreceiver with DAQ AnalogIn.
+
+    :param luci: a LUCI_OE200 instance
+    :type luci: LUCI_OE200
+    :param pd: a AnalogPD instance. gain should be 1.
+    :type pd: AnalogPD
+
+    """
+
+    def __init__(self, name, conf, prefix=None):
+        self.luci: LUCI_OE200 = self.conf.get("luci")
+        self.pd: AnalogPD = self.conf.get("pd")
+        if len(self.pd.lines) != 1:
+            raise ValueError("len(lines) of pd must be 1.")
+
+        self.add_instruments(self.luci, self.pd)
+
+    def _convert_data(
+        self, data: np.ndarray | np.float64 | tuple[np.ndarray, float] | None
+    ) -> np.ndarray | np.float64 | tuple[np.ndarray, float] | None:
+        if data is None:
+            return None
+
+        if isinstance(data, tuple):
+            #  stamped clock-mode
+            return data[0] / self.luci.gain, data[1]
+
+        # np.ndarray (non-stamped clock-mode) | np.float64 (read-on-demand)
+        return data / self.luci.gain
+
+    def _convert_all_data(
+        self, data: list[np.ndarray | tuple[np.ndarray, float]] | None
+    ) -> list[np.ndarray | tuple[np.ndarray, float]] | None:
+        if data is None:
+            return None
+        return [self._convert_data(d) for d in data]
+
+    def set(self, key: str, value=None) -> bool:
+        # no set() key for AnalogIn
+        return self.luci.set(key, value)
+
+    def get(self, key: str, args=None):
+        if key == "data":
+            return self._convert_data(self.pd.get(key, args))
+        elif key == "all_data":
+            return self._convert_all_data(self.pd.get(key, args))
+        elif key == "unit":
+            return "W"
+        else:
+            return self.luci.get(key, args)
+
+    def get_param_dict_labels(self, group: str) -> list[str]:
+        return ["luci"]
+
+    def get_param_dict(
+        self, label: str = "", group: str = ""
+    ) -> P.ParamDict[str, P.PDValue] | None:
+        """Get ParamDict for `label` in `group`."""
+
+        if label == "luci":
+            return self.luci.get_param_dict(label, group)
+        else:
+            return self.pd.get_param_dict(label, group)
+
+    def configure(self, params: dict, label: str = "", group: str = "") -> bool:
+        if label == "luci":
+            return self.luci.configure(params, label, group)
+        else:
+            return self.pd.configure(params, label, group)
+
+    def start(self) -> bool:
+        return self.pd.start()
+
+    def stop(self) -> bool:
+        return self.pd.stop()
+
+
+class OE200_LI5640_AI(InstrumentOverlay):
     """FEMTO Messtechnik OE-200 Variable Gain Photoreceiver and with LI5640 Lockin & DAQ AnalogIn.
 
     :param luci: a LUCI_OE200 instance
@@ -35,23 +113,39 @@ class OE200_LI5640(InstrumentOverlay):
         self.add_instruments(self.luci, self.li5640, self.pd)
 
         self.init_lockin()
+        self.update_gain()
+
+        gr, gi = self.gain
[…]
