+++ mahos/inst/lockin.py
@@ -11,7 +11,6 @@ from __future__ import annotations
 import enum
 
 from .visa_instrument import VisaInstrument
-
 from ..msgs import param_msgs as P
 
 
@@ -93,9 +92,9 @@ class LI5640(VisaInstrument):
         VoltageSensitivity.s200uV: 200e-6,
         VoltageSensitivity.s500uV: 500e-6,
         VoltageSensitivity.s1mV: 1e-3,
-        VoltageSensitivity.s2mV: 2e03,
-        VoltageSensitivity.s5mV: 5e03,
-        VoltageSensitivity.s10mV: 10e-03,
+        VoltageSensitivity.s2mV: 2e-3,
+        VoltageSensitivity.s5mV: 5e-3,
+        VoltageSensitivity.s10mV: 10e-3,
         VoltageSensitivity.s20mV: 20e-3,
         VoltageSensitivity.s50mV: 50e-3,
         VoltageSensitivity.s100mV: 100e-3,
@@ -171,6 +170,42 @@ class LI5640(VisaInstrument):
         dB: 1
         PERCENT: 2
 
+    PARAM_NAMES = [
+        "phase_offset",
+        "freq",
+        "amplitude",
+        "harmonics",
+        "ref_source",
+        "ref_edge",
+        "input_source",
+        "coupling",
+        "ground",
+        "line_filter",
+        "line_freq",
+        "LPF_through",
+        "dynamic_reserve",
+        "volt_sensitivity",
+        "current_sensitivity",
+        "time_constant",
+        "sync_filter",
+        "slope",
+        "data1",
+        "data2",
+        "data_normalization",
+        "data_normalization_voltage",
+        "data_normalization_current",
+        "noise_smooth",
+        "Xoffset_enable",
+        "Yoffset_enable",
+        "Xoffset",
+        "Yoffset",
+        "data1_expand",
+        "data2_expand",
+        "ratio_enable",
+        "ratio_k",
+        "lamp",
+    ]
+
     def __init__(self, name, conf, prefix=None):
         if "write_termination" not in conf:
             conf["write_termination"] = "\n"
@@ -178,11 +213,7 @@ class LI5640(VisaInstrument):
             conf["read_termination"] = "\n"
         VisaInstrument.__init__(self, name, conf, prefix=prefix)
 
-        # memoise
-        # TODO: this is ad-hoc.
-        self._data1_expand = self.get_data1_expand()
-        self._data2_expand = self.get_data2_expand()
-        self._volt_sensitivity = self.get_volt_sensitivity_float()
+        self._locks = {p: False for p in self.PARAM_NAMES}
 
     def get_phase_offset(self) -> float:
         """get current phase offset in degrees."""
@@ -200,6 +231,43 @@ class LI5640(VisaInstrument):
         self.inst.write(f"PHAS {phase_deg:.4f}")
         return True
 
+    def get_freq(self) -> float:
+        """get frequency of `internal oscillator`.
+
+        Note that this is not measured frequency of external reference.
+
+        """
+
+        return float(self.inst.query("FREQ?"))
+
+    def set_freq(self, freq: float) -> bool:
+        """set frequency of `internal oscillator`."""
+
+        self.inst.write(f"FREQ {freq:.4f}")
+        return True
+
+    def get_amplitude(self) -> float:
+        """get amplitude of internal oscillator."""
+
+        #  drop range info (after comma)
+        return float(self.inst.query("AMPL?").split(",")[0])
+
+    def set_amplitude(self, ampl: float) -> bool:
+        """set amplitude of internal oscillator."""
+
+        if not (0.0 <= ampl <= 5.0):
[…]
