### 2023-09-05 68ebb8e
inst/lockin,pd: fix bugs using real insts
