### 2023-09-05 f5d8ca3
inst/spectrometer: make force-importable on linux
- without pythonnet installation
