### 2023-09-06 04b9a08
inst/fg: add ParamDict-based interface
