+++ mahos/inst/fg.py
@@ -7,9 +7,8 @@ Function Generator module.
 
 """
 
-import typing as T
-
 from .visa_instrument import VisaInstrument
+from ..msgs import param_msgs as P
 
 
 class DG2000(VisaInstrument):
@@ -30,11 +29,15 @@ class DG2000(VisaInstrument):
 
     TRIG_SOURCE = ("INT", "EXT", "MAN")
     IDLE_LEVEL = ("FPT", "TOP", "CENTER", "BOTTOM")
+    OUTPUT_HighZ = 11_000
 
     def __init__(self, name, conf, prefix=None):
         VisaInstrument.__init__(self, name, conf, prefix=prefix)
-        self.update_amplitude_bounds()
-        self.update_freq_bounds()
+        self._ampl_bounds = {1: None, 2: None}
+        self._freq_bounds = {1: None, 2: None}
+        for ch in (1, 2):
+            self.update_amplitude_bounds(ch)
+            self.update_freq_bounds(ch)
 
         self.ext_ref_clock = self.conf.get("ext_ref_clock", False)
         self.set_reference_clock(bool(self.ext_ref_clock))
@@ -48,18 +51,18 @@ class DG2000(VisaInstrument):
         }
         self.logger.debug(f"gate configuration: {self.gate_conf}")
 
-    def update_amplitude_bounds(self):
-        mn = float(self.inst.query(":VOLT? MIN"))
-        mx = float(self.inst.query(":VOLT? MAX"))
-        self._ampl_bounds = (mn, mx)
+    def update_amplitude_bounds(self, ch: int):
+        mn = float(self.inst.query(f":SOUR{ch}:VOLT? MIN"))
+        mx = float(self.inst.query(f":SOUR{ch}:VOLT? MAX"))
+        self._ampl_bounds[ch] = (mn, mx)
 
-    def update_freq_bounds(self):
-        mn = float(self.inst.query(":FREQ? MIN"))
-        mx = float(self.inst.query(":FREQ? MAX"))
-        self._freq_bounds = (mn, mx)
+    def update_freq_bounds(self, ch: int):
+        mn = float(self.inst.query(f":SOUR{ch}:FREQ? MIN"))
+        mx = float(self.inst.query(f":SOUR{ch}:FREQ? MAX"))
+        self._freq_bounds[ch] = (mn, mx)
 
-    def get_bounds(self):
-        return {"ampl": self._ampl_bounds, "freq": self._freq_bounds}
+    def get_bounds(self, ch: int = 1):
+        return {"ampl": self._ampl_bounds[ch], "freq": self._freq_bounds[ch]}
 
     def _check_channel(self, ch: int) -> bool:
         if ch not in (1, 2):
@@ -77,10 +80,14 @@ class DG2000(VisaInstrument):
         self.logger.info(f"Output{ch} {on_off}")
         return True
 
-    def set_output_impedance(self, imp_ohm: T.Union[int, str], ch: int = 1) -> bool:
+    def get_output(self, ch: int = 1) -> bool:
+        return self.query(f":OUTP{ch}:STAT?") == "ON"
+
+    def set_output_impedance(self, imp_ohm: int | str, ch: int = 1) -> bool:
         """Set output impedance.
 
         :param imp_ohm: Impedance in Ohm. May be str that's one of (INF|MIN|MAX).
+            Value greater than 10e3 is considered HighZ (INF).
         :param ch: Output channel (1 or 2).
 
         """
@@ -90,21 +97,31 @@ class DG2000(VisaInstrument):
         if isinstance(imp_ohm, str):
             imp_ohm = imp_ohm.upper()[:3]
             if imp_ohm not in ("INF", "MIN", "MAX"):
-                self.logger.error(f"Invalid impedance string {imp_ohm}")
[…]
