### 2023-09-06 46a58d8
name change: ParamServer -> GlobalParams
- In config file, target or preferred shorthand is 'gparam'.
- preferred name of LogBroker has been changed from
  'log_broker' to 'log'.
