+++ mahos/node/global_params.py
@@ -1,43 +1,43 @@
 #!/usr/bin/env python3
 
 """
-Global parameter server.
+Global parameter dictionary.
 
 .. This file is a part of MAHOS project.
 
 """
 
 from ..msgs.common_msgs import Resp
-from ..msgs import param_server_msgs
-from ..msgs.param_server_msgs import ParamServerStatus, SetParamReq
+from ..msgs import global_params_msgs
+from ..msgs.global_params_msgs import GlobalParamsStatus, SetParamReq
 from .node import Node
 from .client import StatusClient
 
 
-class ParamClient(StatusClient):
-    """Simple ParamServer Client."""
+class GlobalParamsClient(StatusClient):
+    """Simple GlobalParams Client."""
 
-    M = param_server_msgs
+    M = global_params_msgs
 
-    def get_param(self, name: str, default=None):
-        """Get a parameter. If `name` is not found, return `default`."""
+    def get_param(self, key: str, default=None):
+        """Get a parameter. If `key` is not found, return `default`."""
 
         s = self.get_status()
         if s is None:
             return default
         else:
-            return s.params.get(name, default)
+            return s.params.get(key, default)
 
-    def set_param(self, name: str, value):
+    def set_param(self, key: str, value):
         """Set a parameter. Value can be any pickle-able Python object."""
 
-        resp = self.req.request(SetParamReq(name, value))
+        resp = self.req.request(SetParamReq(key, value))
 
         return resp.success
 
 
-class ParamServer(Node):
-    CLIENT = ParamClient
+class GlobalParams(Node):
+    CLIENT = GlobalParamsClient
 
     def __init__(self, gconf: dict, name, context=None):
         Node.__init__(self, gconf, name, context=context)
@@ -45,10 +45,10 @@ class ParamServer(Node):
         self.add_rep()
         self.status_pub = self.add_pub(b"status")
 
-        self._param_dict = {}
+        self._params = {}
 
     def set_param(self, msg: SetParamReq) -> Resp:
-        self._param_dict[msg.name] = msg.value
+        self._params[msg.key] = msg.value
         return Resp(True, "")
 
     def handle_req(self, msg):
@@ -58,7 +58,7 @@ class ParamServer(Node):
             return Resp(False, "Invalid message type")
 
     def _publish(self):
-        s = ParamServerStatus(params=self._param_dict)
+        s = GlobalParamsStatus(params=self._params)
         self.status_pub.publish(s)
 
     def main(self):

[…]
