### 2023-09-06 bd59583
meas: use add_client() to close it properly
