+++ mahos/meas/spectroscopy.py
@@ -48,6 +48,7 @@ class Spectroscopy(BasicMeasNode):
             self.tweaker_cli = TweakerClient(
                 gconf, self.conf["target"]["tweaker"], context=self.ctx, prefix=self.joined_name()
             )
+            self.add_client(self.tweaker_cli)
         else:
             self.tweaker_cli = None
 
