+++ mahos/meas/camera.py
@@ -43,6 +43,7 @@ class Camera(BasicMeasNode):
             self.tweaker_cli = TweakerClient(
                 gconf, self.conf["target"]["tweaker"], context=self.ctx, prefix=self.joined_name()
             )
+            self.add_client(self.tweaker_cli)
         else:
             self.tweaker_cli = None
 
