+++ mahos/meas/confocal.py
@@ -257,6 +257,7 @@ class Confocal(Node):
             self.tweaker_cli = TweakerClient(
                 gconf, self.conf["target"]["tweaker"], context=self.ctx, prefix=self.joined_name()
             )
+            self.add_client(self.tweaker_cli)
         else:
             self.tweaker_cli = None
 
