+++ mahos/meas/podmr.py
@@ -87,6 +87,7 @@ class PODMR(BasicMeasNode):
             self.tweaker_cli = TweakerClient(
                 gconf, self.conf["target"]["tweaker"], context=self.ctx, prefix=self.joined_name()
             )
+            self.add_client(self.tweaker_cli)
         else:
             self.tweaker_cli = None
 
