### 2023-09-06 23208c5
docs: add docs of recently added things
- Tweaker and get_param_dict in Instrument
- lockin and amplified_pd
