+++ mahos/node/global_params.py
@@ -37,6 +37,13 @@ class GlobalParamsClient(StatusClient):
 
 
 class GlobalParams(Node):
+    """Node to handle the global parameter dictionary (gparams).
+
+    A parameter is set by SetParamReq (set_param() method of the client).
+    Whole dictionary is distributed inside GlobalParamsStatus ('status' topic).
+
+    """
+
     CLIENT = GlobalParamsClient
 
     def __init__(self, gconf: dict, name, context=None):
