+++ docs/_sources/inst/generated/mahos.inst.overlay.amplified_pd.OE200_AI.rst.txt
@@ -1,0 +1,34 @@
+﻿mahos.inst.overlay.amplified\_pd.OE200\_AI
+==========================================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.inst.overlay.amplified_pd
+
+.. autoclass:: OE200_AI
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~OE200_AI.__init__
+      ~OE200_AI.configure
+      ~OE200_AI.get
+      ~OE200_AI.get_max_rate
+      ~OE200_AI.get_param_dict
+      ~OE200_AI.get_param_dict_labels
+      ~OE200_AI.pop_block
+      ~OE200_AI.set
+      ~OE200_AI.start
+      ~OE200_AI.stop
+   
+   
+
+   
+   
+   
