### 2023-09-07 58aa231
inst: debug fg using real instruments
