+++ mahos/inst/fg.py
@@ -7,8 +7,42 @@ Function Generator module.
 
 """
 
+from functools import wraps
+
 from .visa_instrument import VisaInstrument
 from ..msgs import param_msgs as P
+
+
+def ch_setter(channels):
+    def _check_ch(f):
+        @wraps(f)
+        def wrapper(*args, **kwargs):
+            self = args[0]
+            ch = kwargs.get("ch", 1)
+            if ch not in channels:
+                self.logger.error(f"Invalid ch {ch} (available: {channels})")
+                return False
+            return f(*args, **kwargs)
+
+        return wrapper
+
+    return _check_ch
+
+
+def ch_getter(channels):
+    def _check_ch(f):
+        @wraps(f)
+        def wrapper(*args, **kwargs):
+            self = args[0]
+            ch = kwargs.get("ch", 1)
+            if ch not in channels:
+                self.logger.error(f"Invalid ch {ch} (available: {channels})")
+                return None
+            return f(*args, **kwargs)
+
+        return wrapper
+
+    return _check_ch
 
 
 class DG2000(VisaInstrument):
@@ -32,12 +66,16 @@ class DG2000(VisaInstrument):
     OUTPUT_HighZ = 11_000
 
     def __init__(self, name, conf, prefix=None):
+        if "write_termination" not in conf:
+            conf["write_termination"] = "\n"
+        if "read_termination" not in conf:
+            conf["read_termination"] = "\n"
         VisaInstrument.__init__(self, name, conf, prefix=prefix)
-        self._ampl_bounds = {1: None, 2: None}
-        self._freq_bounds = {1: None, 2: None}
-        for ch in (1, 2):
-            self.update_amplitude_bounds(ch)
-            self.update_freq_bounds(ch)
+
+        self._ampl_bounds = {}
+        self._offs_bounds = {}
+        self._freq_bounds = {}
+        self.update_all_bounds()
 
         self.ext_ref_clock = self.conf.get("ext_ref_clock", False)
         self.set_reference_clock(bool(self.ext_ref_clock))
@@ -51,38 +89,47 @@ class DG2000(VisaInstrument):
         }
         self.logger.debug(f"gate configuration: {self.gate_conf}")
 
+    def update_all_bounds(self):
+        for ch in (1, 2):
+            self.update_amplitude_bounds(ch)
+            self.update_offset_bounds(ch)
+            self.update_freq_bounds(ch)
+
     def update_amplitude_bounds(self, ch: int):
         mn = float(self.inst.query(f":SOUR{ch}:VOLT? MIN"))
         mx = float(self.inst.query(f":SOUR{ch}:VOLT? MAX"))
         self._ampl_bounds[ch] = (mn, mx)
+
+    def update_offset_bounds(self, ch: int):
+        mn = float(self.inst.query(f":SOUR{ch}:VOLT:OFFS? MIN"))
+        mx = float(self.inst.query(f":SOUR{ch}:VOLT:OFFS? MAX"))
+        self._offs_bounds[ch] = (mn, mx)
 
     def update_freq_bounds(self, ch: int):
         mn = float(self.inst.query(f":SOUR{ch}:FREQ? MIN"))
         mx = float(self.inst.query(f":SOUR{ch}:FREQ? MAX"))
         self._freq_bounds[ch] = (mn, mx)
 
+    @ch_getter((1, 2))
     def get_bounds(self, ch: int = 1):
-        return {"ampl": self._ampl_bounds[ch], "freq": self._freq_bounds[ch]}
+        return {
+            "ampl": self._ampl_bounds[ch],
+            "offs": self._offs_bounds[ch],
+            "freq": self._freq_bounds[ch],
+        }
 
[…]
