### 2023-09-07 7ac0c68
inst/spectrometer: clean try-import part
