### 2023-09-07 cc7ac60
meas/odmr,spec_fitter: give params for KMeans
- to suppress the warning
