+++ mahos/util/graph.py
@@ -13,14 +13,15 @@ import networkx as nx
 
 
 def sort_dependency(dep_dict: dict) -> list:
-    """topological sort to resolve dependency.
+    """use topological sort to resolve dependency.
 
     :param dep_dict: a dict to represent nodes (any hashable objects) and their dependencies.
         {"a": ("b", "c"), "b": ("c",)} reads as "'a' depends on 'b' & 'c'",
         and "'b' depends on 'c'".
     :type dep_dict: dict[Hashable, tuple[Hashable]]
 
-    :returns: list[Hashable]
+    :returns: list[Hashable] topological-sorted nodes (dep_dict's keys and value elements).
+        ["c", "b", "a"] for the example above.
 
     """
 
@@ -28,5 +29,6 @@ def sort_dependency(dep_dict: dict) -> list:
     G.add_nodes_from(dep_dict.keys())
     for v, deps in dep_dict.items():
         for u in deps:
+            #  new node may be automatically created here if u is not in dep_dict.keys().
             G.add_edge(u, v)
     return list(nx.topological_sort(G))
