+++ CHANGELOG.rst
@@ -8,36 +8,51 @@ Changelog
 Added
 ^^^^^
 
-- meas: added new measurement Qdyne.
-- inst/tdc: added raw event recording features for above.
-- meas,gui: added new generic node Tweaker.
+- meas: new measurement Qdyne.
+- inst.tdc: raw event recording features for above.
+- meas,gui: new generic node Tweaker (for manual instrument parameter tuning).
+- msgs.data_msgs.Data: new reserved attr '_inst_params' for parameter injection by Tweaker.
 - inst: new module lockin and LI5640 class in it.
-- inst/overlay: new module amplified_pd.
+- inst.pd: new class LockinAnalogPD.
+- meas.confocal,odmr: accept complex-value data (of LockinAnalogPD).
+- inst.overlay: new module amplified_pd (to use AnalogPDs with configurable amplifiers).
+- inst.fg: ParamDict-based interface.
+- meas.odmr: new parameters delay and sg_modulation.
+- util: new module graph.
 
 Changed
 ^^^^^^^
 
-- make config key names consistent for several Instruments or measurement Workers.
+- (breaking config) make config key names consistent for several Instruments or measurement Workers.
   Suffix key name with '_dir' for directory path configs.
 
-  - inst/piezo/E727_3: limits_um -> limit_um, range -> range_um
-  - inst/camera/ThorlabsCamera: dll_path -> dll_dir
-  - inst/dtg/DTG: root_local -> local_dir, root_remote -> remote_dir
-  - inst/pd/LUCI10,LUCI_OE200: dll_path -> dll_dir
-  - inst/tdc/MCS: home -> mcs_dir, home_raw_events -> raw_events_dir
-  - meas/qdyne_worker -> home_raw_events -> raw_events_dir
+  - inst.piezo.E727_3: limits_um -> limit_um, range -> range_um.
+  - inst.camera.ThorlabsCamera: dll_path -> dll_dir.
+  - inst.dtg.DTG: root_local -> local_dir, root_remote -> remote_dir.
+  - inst.pd.LUCI10,LUCI_OE200: dll_path -> dll_dir.
+  - inst.tdc.MCS: home -> mcs_dir, home_raw_events -> raw_events_dir.
+  - meas.qdyne_worker -> home_raw_events -> raw_events_dir.
 
-- inst/InstrumentServer, Instrument, InstrumentOverlay
+- (breaking config) node class and preferred name changes (46a58d8).
 
-  - overlay can reference other overlays.
-  - new standard API functions get_param_dict() and get_param_dict_labels().
+  - LogBroker's preferred node name: 'log_broker' -> 'log'.
+    Replace '[localhost.log_broker]' / 'localhost::log_broker' with '[localhost.log]' / 'localhost::log'.
+  - Module and class: mahos.node.param_server.ParamServer -> mahos.node.global_params.GlobalParams.
+    Config target key and preferred name: param_server -> gparams.
+    Replace '[localhost.param_server]' with '[localhost.gparams]'.
+    Replace '[localhost.somenode.target] param_server = "localhost::param_server"' with '[localhost.somenode.target] gparams = "localhost::gparams"'.
+
+- inst.InstrumentServer, Instrument, InstrumentOverlay.
+
+  - overlay can reference other overlays as well as raw Instrument.
+  - add new standard API methods get_param_dict() and get_param_dict_labels().
     Instrument can also use ParamDicts with this.
-  - signature of configure(): new kwargs label and group to make it usable with get_param_dict().
[…]
