### 2023-09-07 c4d34c5
cli: auto-start & auto-exclude of log (LogBroker)
- mahos log: auto-start if log node is local and not up
- mahos launch: auto-exclude if it's up
