### 2023-09-11 3bdb5ea
cli: fix log autostart, add docs of launch & log
