### 2023-09-12 fdd2ff3
gui/main_monitor: make pathEdit editable
