### 2023-09-14 911e212
gui/camera: add frame rate setting widgets
