+++ mahos/gui/ui/camera.py
@@ -1,6 +1,6 @@
 # Form implementation generated from reading ui file 'camera.ui'
 #
-# Created by: PyQt6 UI code generator 6.4.2
+# Created by: PyQt6 UI code generator 6.5.2
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic6 is
 # run again.  Do not edit this file unless you know what you are doing.
@@ -51,6 +51,13 @@ class Ui_Camera(object):
         self.exposuretimeBox.setProperty("value", 10.0)
         self.exposuretimeBox.setObjectName("exposuretimeBox")
         self.horizontalLayout_3.addWidget(self.exposuretimeBox)
+        self.framerateenableBox = QtWidgets.QCheckBox(parent=Camera)
+        self.framerateenableBox.setObjectName("framerateenableBox")
+        self.horizontalLayout_3.addWidget(self.framerateenableBox)
+        self.framerateBox = QtWidgets.QDoubleSpinBox(parent=Camera)
+        self.framerateBox.setDecimals(1)
+        self.framerateBox.setObjectName("framerateBox")
+        self.horizontalLayout_3.addWidget(self.framerateBox)
         self.fpsLabel = QtWidgets.QLabel(parent=Camera)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.Minimum, QtWidgets.QSizePolicy.Policy.Preferred)
         sizePolicy.setHorizontalStretch(0)
@@ -122,6 +129,18 @@ class Ui_Camera(object):
 
         self.retranslateUi(Camera)
         QtCore.QMetaObject.connectSlotsByName(Camera)
+        Camera.setTabOrder(self.startButton, self.stopButton)
+        Camera.setTabOrder(self.stopButton, self.saveButton)
+        Camera.setTabOrder(self.saveButton, self.loadButton)
+        Camera.setTabOrder(self.loadButton, self.exposuretimeBox)
+        Camera.setTabOrder(self.exposuretimeBox, self.framerateenableBox)
+        Camera.setTabOrder(self.framerateenableBox, self.framerateBox)
+        Camera.setTabOrder(self.framerateBox, self.binningBox)
+        Camera.setTabOrder(self.binningBox, self.roiBox)
+        Camera.setTabOrder(self.roiBox, self.woffsetBox)
+        Camera.setTabOrder(self.woffsetBox, self.widthBox)
+        Camera.setTabOrder(self.widthBox, self.hoffsetBox)
+        Camera.setTabOrder(self.hoffsetBox, self.heightBox)
 
     def retranslateUi(self, Camera):
         _translate = QtCore.QCoreApplication.translate
@@ -132,6 +151,8 @@ class Ui_Camera(object):
         self.loadButton.setText(_translate("Camera", "Load"))
         self.exposuretimeBox.setPrefix(_translate("Camera", "Exposure time: "))
         self.exposuretimeBox.setSuffix(_translate("Camera", " ms"))
+        self.framerateenableBox.setText(_translate("Camera", "limit fps"))
+        self.framerateBox.setSuffix(_translate("Camera", " fps"))
         self.fpsLabel.setText(_translate("Camera", " 0.0 fps"))
         self.binningBox.setSuffix(_translate("Camera", " (0 to skip)"))
         self.binningBox.setPrefix(_translate("Camera", "binning: "))
