### 2023-09-14 bae43f1
camera: set frame_rate (fps limit) from gui
