### 2023-09-14 98c14ef
confocal/odmr: set volt. bounds of AnalogIn PDs
