### 2023-09-14 7cee862
inst/daq: fix bug reading parameters (AI range)
