### 2023-09-14 d52f60e
odmr: set pd_bounds properly
