### 2023-09-15 433d4be
inst.camera.ThorlabsCamera: fix frame_rate setting
- seemingly, this setting order matters
