### 2023-09-15 5af9112
ThorlabsCamera: add conf frame_rate_control
- Some cameras doesn't support frame rate control
  and setting the options raises error.
  But the availability-check function is not found.
