### 2023-09-18 b468b15
MultiInstrumentClient: add inst_remap function
