+++ mahos/meas/common_meas.py
@@ -191,7 +191,11 @@ class BasicMeasNode(Node):
         self.state = BinaryState.IDLE
 
         self.cli = MultiInstrumentClient(
-            gconf, self.conf["target"]["servers"], context=self.ctx, prefix=self.joined_name()
+            gconf,
+            self.conf["target"]["servers"],
+            inst_remap=self.conf.get("inst_remap"),
+            context=self.ctx,
+            prefix=self.joined_name(),
         )
         self.add_client(self.cli)
         self.add_rep()
