### 2023-09-18 458aa23
confocal/piezo: don't log error on target pos
