### 2023-09-18 8cefdd3
LockinAnalogPD_LI5640: fix bug on gain
