### 2023-09-18 292c1da
MultiInstrumentClient: fix bug of remap inst
- b468b15 was totally bugged
