### 2023-09-26 af04601
pulse: fix bugs and add Block/Blocks operations
