### 2023-09-26 98259c6
wip: implementing SPODMR
- SPODMR: Pulse ODMR with Slow detectors
