+++ mahos/meas/spodmr.py
@@ -1,0 +1,221 @@
+#!/usr/bin/env python3
+
+"""
+Logic and instrument control part of Pulse ODMR with Slow detectors.
+
+.. This file is a part of MAHOS project.
+
+"""
+
+from __future__ import annotations
+
+from ..msgs.common_msgs import Resp, Request, StateReq, BinaryState, BinaryStatus
+from ..msgs.common_msgs import SaveDataReq, ExportDataReq, LoadDataReq
+from ..msgs.common_meas_msgs import Buffer
+from ..msgs.param_msgs import GetParamDictLabelsReq, GetParamDictReq
+from ..msgs import spodmr_msgs
+from ..msgs.spodmr_msgs import SPODMRData, UpdatePlotParamsReq, ValidateReq
+from ..util.timer import IntervalTimer
+from .tweaker import TweakerClient
+from .common_meas import BasicMeasClient, BasicMeasNode
+from .common_worker import DummyWorker, Switch
+from .podmr_fitter import PODMRFitter
+from .spodmr_worker import Pulser, SPODMRDataOperator
+from .spodmr_io import SPODMRIO
+
+
+class SPODMRClient(BasicMeasClient):
+    """Node Client for SPODMR."""
+
+    #: Message types for SPODMR.
+    M = spodmr_msgs
+
+    # override for annotation
+    def get_data(self) -> SPODMRData:
+        return self._get_data()
+
+    def get_buffer(self) -> Buffer[tuple[str, SPODMRData]]:
+        return self._get_buffer()
+
+    def update_plot_params(self, params: dict) -> bool:
+        resp = self.req.request(UpdatePlotParamsReq(params))
+        return resp.success
+
+    def validate(self, params: dict) -> bool:
+        resp = self.req.request(ValidateReq(params))
+        return resp.success
+
+
+class SPODMR(BasicMeasNode):
+    CLIENT = SPODMRClient
+    DATA = SPODMRData
+
+    def __init__(self, gconf: dict, name, context=None):
+        """Pulse ODMR measurement with Slow detectors.
+
+        :param pulser.freq: pulse generator frequency (default: 2.0E9)
+        :type pulser.freq: float
+        :param pulser.reduce_start_divisor: the divisor on start of reducing frequency (default: 2)
+            reduce is done first by this value, and then repeated by 10.
+        :type pulser.reduce_start_divisor: int
+        :param pulser.split_fraction: fraction factor (F) to split the free period for MW phase
+            modulation (default: 4). the period (T) is split into (T // F, T - T // F) and MW phase
+            is switched at T // F. Thus, larger F results in "quicker start" of the phase
+            modulation (depending on hardware, but its response may be a bit slow).
+        :type pulser.split_fraction: int
+        :param pulser.minimum_block_length: minimum block length in generated blocks
+                                            (default: 1000)
+        :type pulser.minimum_block_length: int
+        :param pulser.block_base: block base granularity (default: 4)
+        :type pulser.block_base: int
+
+        """
+
+        BasicMeasNode.__init__(self, gconf, name, context=context)
+
+        self.pulse_pub = self.add_pub(b"pulse")
+
+        if "switch" in self.conf["target"]["servers"]:
+            self.switch = Switch(self.cli, self.logger, "podmr")
+        else:
+            self.switch = DummyWorker()
+        if "tweaker" in self.conf["target"]:
+            self.tweaker_cli = TweakerClient(
+                gconf, self.conf["target"]["tweaker"], context=self.ctx, prefix=self.joined_name()
+            )
+            self.add_client(self.tweaker_cli)
+        else:
+            self.tweaker_cli = None
+
+        has_fg = "fg" in self.conf["target"]["servers"]
+        self.worker = Pulser(self.cli, self.logger, has_fg, self.conf.get("pulser", {}))
[…]
