+++ mahos/msgs/spodmr_msgs.py
@@ -1,0 +1,357 @@
+#!/usr/bin/env python3
+
+"""
+Message Types for Pulse ODMR with Slow detectors.
+
+.. This file is a part of MAHOS project.
+
+"""
+
+from __future__ import annotations
+
+import numpy as np
+from numpy.typing import NDArray
+
+from ..util.comp import dict_isclose
+from .common_msgs import Request
+from .common_meas_msgs import BasicMeasData
+from .data_msgs import ComplexDataMixin
+
+
+class ValidateReq(Request):
+    """Validate Measurement Params Request"""
+
+    def __init__(self, params: dict):
+        self.params = params
+
+
+class UpdatePlotParamsReq(Request):
+    """Request to update the plot param"""
+
+    def __init__(self, params: dict):
+        self.params = params
+
+
+def is_sweepN(method: str) -> bool:
+    return method in ("cpN", "cpmgN", "xy4N", "xy8N", "xy16N", "ddgateN", "xy8clNflip")
+
+
+def is_CPlike(method: str) -> bool:
+    return method in (
+        "cp",
+        "cpN",
+        "cpmg",
+        "cpmgN",
+        "xy4",
+        "xy4N",
+        "xy8",
+        "xy8N",
+        "xy16",
+        "xy16N",
+        "ddgate",
+        "ddgateN",
+    )
+
+
+def is_correlation(method: str) -> bool:
+    return method in ("xy8cl", "xy8cl1flip", "xy8clNflip")
+
+
+class SPODMRData(BasicMeasData, ComplexDataMixin):
+    def __init__(self, params: dict | None = None):
+        self.set_version(0)
+        self.init_params(params)
+        self.init_attrs()
+
+        self.init_xdata()
+
+        self.data0 = None
+        self.data1 = None
+
+        #: complex_conv used for loading data fit
+        self.fit_complex_conv: str = ""
+
+    def init_xdata(self):
+        if not self.has_params():
+            # dummy xdata
+            self.xdata = np.arange(101) * 1e-9
+            return
+
+        # xdata is N (not tau)
+        if self.is_sweepN():
+            i = np.arange(self.params["Nnum"], dtype=np.int64)
+            self.xdata = self.params["Nstart"] + i * self.params["Nstep"]
+        else:
+            # supersampling
+            if self.is_supersampling():
+                i = np.arange(self.params["num"])
+                pulse_num = {"xy8": 8, "xy16": 16}[self.params["method"]] * self.params["Nconst"]
+                sample = np.linspace(0, 1, num=self.params["supersample"] + 1)
+                i_sub = (sample * pulse_num).astype(np.int64)[:-1] / pulse_num
+
+                i = np.array([i_sub + j for j in i]).reshape(len(i) * len(i_sub))
+                self.xdata = self.params["start"] + i * self.params["step"]
+
+            # xlog
+            elif self.params["log"]:
+                start = np.log10(self.params["start"])
+                stop = np.log10(
+                    self.params["start"] + (self.params["num"] - 1) * self.params["step"]
+                )
+                x = np.logspace(start, stop, self.params["num"])
+                if (self.params["start"] >= 1e-6) and (self.params["step"] >= 1e-6):
+                    self.xdata = (x * 1e9 / 10).astype(np.int64) * 10 / 1e9  # round in [10 ns]
+                else:
+                    self.xdata = (x * 1e9).astype(np.int64) / 1e9  # round in [ns]
+
+            # normal
+            else:
+                i = np.arange(self.params["num"])
+                self.xdata = self.params["start"] + i * self.params["step"]
+
+        if self.params.get("invertsweep", False):
+            self.xdata = self.xdata[::-1]
+
[…]

[…]
