### 2023-09-27 78eb680
SPODMR: store laser_duties for normalization
