+++ mahos/msgs/inst_pg_msgs.py
@@ -32,7 +32,8 @@ Pattern = T.NewType("Pattern", T.List[T.Tuple[Channels, int]])
 AcceptedChannels = T.NewType(
     "AcceptedChannels", T.Union[None, str, int, T.Tuple[str, ...], T.Tuple[int, ...]]
 )
-AcceptedPattern = T.NewType("AcceptedPattern", T.List[T.Tuple[AcceptedChannels, int]])
+AcceptedPatternElement = T.NewType("AcceptedPatternElement", T.Tuple[AcceptedChannels, int])
+AcceptedPattern = T.NewType("AcceptedPattern", T.List[AcceptedPatternElement])
 
 
 class Block(Message):
@@ -62,23 +63,56 @@ class Block(Message):
         else:
             raise TypeError("channel {} has unrecognizable type {}".format(ch, type(ch)))
 
-    def regularize_pattern_elem(self, elem: tuple[AcceptedChannels, int]):
+    def regularize_pattern_elem(self, elem: AcceptedPatternElement):
         # cast duration to builtin int because numpy types (like np.int64) may be incorpolated
         return self.regularize_channel(elem[0]), int(elem[1])
 
     def regularize_pattern(self, pattern: AcceptedPattern) -> Pattern:
         return [self.regularize_pattern_elem(p) for p in pattern]
 
-    def insert(self, index: int, elem: T.Tuple[AcceptedChannels, int]):
+    # Mutating (list-like) operations
+
+    def insert(self, index: int, elem: AcceptedPatternElement):
+        """Insert a pattern element `elem` to `index`."""
+
         self.pattern.insert(index, self.regularize_pattern_elem(elem))
 
-    def append(self, elem: T.Tuple[AcceptedChannels, int]):
+    def append(self, elem: AcceptedPatternElement):
+        """Append a pattern element `elem` to the tail."""
+
         self.pattern.append(self.regularize_pattern_elem(elem))
 
     def extend(self, ptn: Block | AcceptedPattern):
+        """Extend the pattern or pattern in a Block."""
+
         if isinstance(ptn, Block):
             ptn = ptn.total_pattern()
         self.pattern.extend(self.regularize_pattern(ptn))
+
+    # Non-mutating (copying) operations
+
+    def repeat(self, num: int) -> Block:
+        """Return new Block with `num`-times repeat."""
+
+        ret = self.copy()
+        ret.Nrep *= num
+        return ret
+
+    def collapse(self) -> Block:
+        """Return new Block with collapsed pattern.
+
+        The pattern is physically repeated and Nrep is set 1.
+
+        """
+
+        return Block(self.name, self.total_pattern(), Nrep=1, trigger=self.trigger)
+
+    def concatenate(self, other: Block) -> Block:
+        """Generate a concatenated Block of this and `other`."""
+
+        ret = self.collapse()
+        ret.extend(other)
+        return ret
 
     def raw_length(self) -> int:
         """Raw block length without considering Nrep."""
@@ -94,9 +128,6 @@ class Block(Message):
         """Total (repeated) pattern considering Nrep."""
 
         return self.Nrep * self.pattern
-
-    def collapse(self) -> Block:
-        return Block(self.name, self.total_pattern(), Nrep=1, trigger=self.trigger)
 
     def channels(self) -> set[str | int]:
         """Get set of channels included in this Block."""
@@ -127,6 +158,17 @@ class Block(Message):
         return channels, patterns
 
     def union(self, other: Block) -> Block:
+        """Return new united Block of this and `other`.
+
+        union() operation is used to merge two Blocks generated in independent ways.
+        The channels described in this and `other` must be exclusive, i.e.,
[…]

[…]
