### 2023-09-28 04ae06f
add test of SPODMR
- Block.union: make this faster
