### 2023-09-28 669b9a8
inst_pg_msgs: use NamedTuple for PatternElement
