### 2023-09-28 28be882
add gui/spodmr
