+++ mahos/meas/spodmr_worker.py
@@ -137,22 +137,26 @@ class BlocksBuilder(object):
     """Build the PG Blocks for SPODMR from PODMR's Blocks."""
 
     def build_complementary(
-        self, blocks: list[Blocks[Block]], freq: float, half_flip_period: int, params: dict
+        self,
+        blocks: list[Blocks[Block]],
+        accum_window: int,
+        accum_rep: int,
+        pd_period: int,
+        laser_width: int,
     ):
-        pd_period = int(round(freq / params["pd_rate"]))
-        pd_rep = half_flip_period // pd_period
-        residual = half_flip_period % pd_period
+        pd_rep = accum_window // pd_period
+        residual = accum_window % pd_period
         w = pd_period // 2
         gate0_blk = Block(
             "gate",
-            [("sync", residual)] + [(("sync", "gate"), w), ("sync", pd_period - w)] * pd_rep,
+            [("sync", residual)] + [("sync", w), (("sync", "gate"), pd_period - w)] * pd_rep,
         )
         gate1_blk = Block(
-            "gate", [(None, residual)] + [("gate", w), (None, pd_period - w)] * pd_rep
+            "gate", [(None, residual)] + [(None, w), ("gate", pd_period - w)] * pd_rep
         )
 
         extended_blks = Blocks()
-        periods = []
+        laser_duties = []
         # markers just for PulseMonitor visualization
         markers = [0]
 
@@ -163,31 +167,38 @@ class BlocksBuilder(object):
             blks1: Blocks[Block] = blks[2:]
             T = blks0.total_length()
             assert T == blks1.total_length()
-            periods.append(T)
-            rep, residual = half_flip_period // T, half_flip_period % T
+            rep, residual = accum_window // T, accum_window % T
             blk0 = blks0.repeat(rep).collapse()
             blk1 = blks1.repeat(rep).collapse()
             if residual:
-                blk0.insert(0, (None, residual))
-                blk1.insert(0, (None, residual))
-            extended_blks.extend([blk0.union(gate0_blk), blk1.union(gate1_blk)])
+                blk0.insert(0, (blk0.pattern[0].channels, residual))
+                blk1.insert(0, (blk1.pattern[0].channels, residual))
+            extended_blks.extend(
+                [blk0.union(gate0_blk).repeat(accum_rep), blk1.union(gate1_blk).repeat(accum_rep)]
+            )
+            # residual is dark: duty becomes slightly lower than laser_width / T
+            laser_duties.append(laser_width * rep / accum_window)
             markers.append(extended_blks.total_length())
 
-        return extended_blks, periods, markers, pd_rep
+        return extended_blks, laser_duties, markers, pd_rep * accum_rep
 
     def build_partial(
-        self, blocks: list[Blocks[Block]], freq: float, half_flip_period: int, params: dict
+        self,
+        blocks: list[Blocks[Block]],
+        accum_window: int,
+        accum_rep: int,
+        pd_period: int,
+        laser_width: int,
     ):
-        pd_period = int(round(freq / params["pd_rate"]))
-        pd_rep = half_flip_period // pd_period
-        residual = half_flip_period % pd_period
+        pd_rep = accum_window // pd_period
+        residual = accum_window % pd_period
         w = pd_period // 2
         gate_blks = Block(
-            "gates", [(None, residual)] + [("gate", w), (None, pd_period - w)] * pd_rep
+            "gate", [(None, residual)] + [(None, w), ("gate", pd_period - w)] * pd_rep
         )
 
         extended_blks = Blocks()
-        periods = []
+        laser_duties = []
         # markers just for PulseMonitor visualization
         markers = [0]
 
@@ -195,34 +206,38 @@ class BlocksBuilder(object):
[…]
