### 2023-09-28 d74356d
gui/pulse_monitor: add max data points limit
