+++ mahos/msgs/inst_pg_msgs.py
@@ -160,16 +160,21 @@ class Block(Message):
             s.update(ch)
         return s
 
-    def decode(self, channel: str | int) -> list[bool]:
+    def decode(self, channel: str | int, max_len: int | None = None) -> list[bool]:
         """Decode the pulse pattern to bool list."""
 
         ptn = []
         for ch, duration in self.total_pattern():
             elem = [True] if channel in ch else [False]
             ptn.extend(elem * duration)
-        return ptn
+            if max_len is not None and len(ptn) > max_len:
+                break
+        if max_len is None:
+            return ptn
+        else:
+            return ptn[:max_len]
 
-    def decode_all(self) -> tuple[list[str | int], list[list[bool]]]:
+    def decode_all(self, max_len: int | None = None) -> tuple[list[str | int], list[list[bool]]]:
         """Decode the patterns for all included channels."""
 
         channels = self.channels()
@@ -177,7 +182,7 @@ class Block(Message):
             channels = list(sorted(channels))
         except TypeError:
             channels = list(channels)
-        patterns = [self.decode(ch) for ch in channels]
+        patterns = [self.decode(ch, max_len=max_len) for ch in channels]
         return channels, patterns
 
     def union(self, other: Block) -> Block:
@@ -406,7 +411,7 @@ class Blocks(UserList):
         except TypeError:
             return list(channels)
 
-    def decode(self, channel: str | int) -> NDArray:
+    def decode(self, channel: str | int, max_len: int | None = None) -> NDArray:
         """Decode the pattern for given channel.
 
         If channel is not included, all-zero array will be returned.
@@ -415,14 +420,19 @@ class Blocks(UserList):
 
         ptn = []
         for block in self.data:
-            ptn.extend(block.decode(channel))
-        return np.array(ptn, dtype=np.uint8)
+            ptn.extend(block.decode(channel, max_len=max_len))
+            if max_len is not None and len(ptn) > max_len:
+                break
+        if max_len is None:
+            return np.array(ptn, dtype=np.uint8)
+        else:
+            return np.array(ptn[:max_len], dtype=np.uint8)
 
-    def decode_all(self) -> tuple[list[str | int], list[NDArray]]:
+    def decode_all(self, max_len: int | None = None) -> tuple[list[str | int], list[NDArray]]:
         """Decode the patterns for all included channels."""
 
         channels = self.channels()
-        patterns = [self.decode(ch) for ch in channels]
+        patterns = [self.decode(ch, max_len=max_len) for ch in channels]
         return channels, patterns
 
     def equivalent(self, other: Blocks[Block]) -> bool:
