### 2023-09-29 1c6ae24
spodmr: add param lockin_rep
