+++ mahos/gui/ui/spodmr.py
@@ -164,17 +164,17 @@ class Ui_SPODMR(object):
         self.sweepsBox.setMaximum(999999999)
         self.sweepsBox.setProperty("value", 0)
         self.sweepsBox.setObjectName("sweepsBox")
-        self.gridLayout_2.addWidget(self.sweepsBox, 0, 4, 1, 1)
+        self.gridLayout_2.addWidget(self.sweepsBox, 0, 5, 1, 1)
+        self.lockinrepBox = QtWidgets.QSpinBox(parent=self.groupBox_2)
+        self.lockinrepBox.setMinimum(1)
+        self.lockinrepBox.setMaximum(9999)
+        self.lockinrepBox.setObjectName("lockinrepBox")
+        self.gridLayout_2.addWidget(self.lockinrepBox, 0, 4, 1, 1)
         self.verticalLayout.addWidget(self.groupBox_2)
         self.groupBox_5 = QtWidgets.QGroupBox(parent=self.mainTab)
         self.groupBox_5.setObjectName("groupBox_5")
         self.gridLayout_3 = QtWidgets.QGridLayout(self.groupBox_5)
         self.gridLayout_3.setObjectName("gridLayout_3")
-        self.divideblockBox = QtWidgets.QCheckBox(parent=self.groupBox_5)
-        self.divideblockBox.setEnabled(True)
-        self.divideblockBox.setChecked(True)
-        self.divideblockBox.setObjectName("divideblockBox")
-        self.gridLayout_3.addWidget(self.divideblockBox, 2, 6, 1, 1)
         self.NstopLabel = QtWidgets.QLabel(parent=self.groupBox_5)
         self.NstopLabel.setAlignment(QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignLeft|QtCore.Qt.AlignmentFlag.AlignVCenter)
         self.NstopLabel.setObjectName("NstopLabel")
@@ -521,7 +521,8 @@ class Ui_SPODMR(object):
         SPODMR.setTabOrder(self.loadButton, self.tabWidget)
         SPODMR.setTabOrder(self.tabWidget, self.accumwindowBox)
         SPODMR.setTabOrder(self.accumwindowBox, self.accumrepBox)
-        SPODMR.setTabOrder(self.accumrepBox, self.sweepsBox)
+        SPODMR.setTabOrder(self.accumrepBox, self.lockinrepBox)
+        SPODMR.setTabOrder(self.lockinrepBox, self.sweepsBox)
         SPODMR.setTabOrder(self.sweepsBox, self.pdrateBox)
         SPODMR.setTabOrder(self.pdrateBox, self.pd_lbBox)
         SPODMR.setTabOrder(self.pd_lbBox, self.pd_ubBox)
@@ -547,8 +548,7 @@ class Ui_SPODMR(object):
         SPODMR.setTabOrder(self.reinitXBox, self.readYBox)
         SPODMR.setTabOrder(self.readYBox, self.invertinitBox)
         SPODMR.setTabOrder(self.invertinitBox, self.reduceBox)
-        SPODMR.setTabOrder(self.reduceBox, self.divideblockBox)
-        SPODMR.setTabOrder(self.divideblockBox, self.tau2constBox)
+        SPODMR.setTabOrder(self.reduceBox, self.tau2constBox)
         SPODMR.setTabOrder(self.tau2constBox, self.N2constBox)
         SPODMR.setTabOrder(self.N2constBox, self.ddgatephaseEdit)
         SPODMR.setTabOrder(self.ddgatephaseEdit, self.N3constBox)
@@ -566,13 +566,13 @@ class Ui_SPODMR(object):
[…]
