### 2023-09-29 bca821d
debugging spodmr
