### 2023-09-29 5ddebc0
spodmr: tune buffer size
