### 2023-09-29 11419ea
inst.pg.PulseStreamer: add D/A mapping function
