### 2023-10-02 e742d57
spodmr: add DebugPulser
