### 2023-10-02 15a2d33
Block/Blocks: add plottable()
- decode() generates data unnecessarily too large
  for plotting purpose
- gui.pulse_monitor.PulseMonitor uses new plottable() instead.
