### 2023-10-02 c52b94a
gui/spodmr: don't request validate before start
