+++ mahos/gui/ui/spodmr.py
@@ -55,12 +55,21 @@ class Ui_SPODMR(object):
         self.groupBox_2.setObjectName("groupBox_2")
         self.gridLayout_2 = QtWidgets.QGridLayout(self.groupBox_2)
         self.gridLayout_2.setObjectName("gridLayout_2")
-        self.nomwBox = QtWidgets.QCheckBox(parent=self.groupBox_2)
-        self.nomwBox.setObjectName("nomwBox")
-        self.gridLayout_2.addWidget(self.nomwBox, 2, 4, 1, 1)
+        self.label_3 = QtWidgets.QLabel(parent=self.groupBox_2)
+        self.label_3.setObjectName("label_3")
+        self.gridLayout_2.addWidget(self.label_3, 1, 0, 1, 1)
         self.label_5 = QtWidgets.QLabel(parent=self.groupBox_2)
         self.label_5.setObjectName("label_5")
         self.gridLayout_2.addWidget(self.label_5, 3, 0, 1, 1)
+        self.label = QtWidgets.QLabel(parent=self.groupBox_2)
+        self.label.setObjectName("label")
+        self.gridLayout_2.addWidget(self.label, 0, 0, 1, 1)
+        self.pd_ubBox = QtWidgets.QDoubleSpinBox(parent=self.groupBox_2)
+        self.pd_ubBox.setMinimum(-10.0)
+        self.pd_ubBox.setMaximum(10.0)
+        self.pd_ubBox.setProperty("value", 10.0)
+        self.pd_ubBox.setObjectName("pd_ubBox")
+        self.gridLayout_2.addWidget(self.pd_ubBox, 1, 5, 1, 1)
         self.powerBox = QtWidgets.QDoubleSpinBox(parent=self.groupBox_2)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.MinimumExpanding, QtWidgets.QSizePolicy.Policy.Fixed)
         sizePolicy.setHorizontalStretch(0)
@@ -72,6 +81,38 @@ class Ui_SPODMR(object):
         self.powerBox.setProperty("value", 9.0)
         self.powerBox.setObjectName("powerBox")
         self.gridLayout_2.addWidget(self.powerBox, 2, 3, 1, 1)
+        self.accumwindowBox = QtWidgets.QDoubleSpinBox(parent=self.groupBox_2)
+        self.accumwindowBox.setObjectName("accumwindowBox")
+        self.gridLayout_2.addWidget(self.accumwindowBox, 0, 1, 1, 1)
+        self.accumrepBox = QtWidgets.QSpinBox(parent=self.groupBox_2)
+        self.accumrepBox.setMinimum(1)
+        self.accumrepBox.setMaximum(1000)
+        self.accumrepBox.setObjectName("accumrepBox")
+        self.gridLayout_2.addWidget(self.accumrepBox, 0, 3, 1, 1)
+        self.label_6 = QtWidgets.QLabel(parent=self.groupBox_2)
+        self.label_6.setObjectName("label_6")
+        self.gridLayout_2.addWidget(self.label_6, 2, 0, 1, 1)
+        spacerItem1 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Minimum)
+        self.gridLayout_2.addItem(spacerItem1, 0, 9, 1, 1)
+        self.freqBox = QtWidgets.QDoubleSpinBox(parent=self.groupBox_2)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.MinimumExpanding, QtWidgets.QSizePolicy.Policy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
[…]

[…]
