### 2023-10-02 ebfc43a
spodmr: tweak a little
