### 2023-10-02 ffaa137
spodmr: debugging with real instruments
