### 2023-10-03 8bf9fbb
inst.daq.ClockSource: add trigger function
