### 2023-10-03 3f0e5e1
spodmr: sample PD using DAQ-generated clock
- reduces pulse pattern size (PG memory)
- name change accum_drop -> drop_rep
