+++ mahos/meas/spodmr_worker.py
@@ -18,6 +18,7 @@ from ..inst.sg_interface import SGInterface
 from ..inst.pg_interface import PGInterface, Block, Blocks
 from ..inst.pd_interface import PDInterface
 from ..inst.fg_interface import FGInterface
+from ..inst.daq_interface import ClockSourceInterface
 from .common_worker import Worker
 
 from .podmr_generator.generator import make_generators
@@ -136,27 +137,33 @@ class Bounds(object):
 class BlocksBuilder(object):
     """Build the PG Blocks for SPODMR from PODMR's Blocks."""
 
+    def __init__(self, trigger_width: float, trigger_channel: str):
+        self.trigger_width = trigger_width
+        self.trigger_channel = trigger_channel
+
     def build_complementary(
         self,
         blocks: list[Blocks[Block]],
         accum_window: int,
         accum_rep: int,
-        accum_drop: int,
-        pd_period: int,
+        drop_rep: int,
         laser_width: int,
+        trigger_width: int,
     ):
-        pd_rep = accum_window // pd_period
-        residual = accum_window % pd_period
-        w = pd_period // 2
-        drop0_blk = Block("drop", [("sync", accum_window)])
-        gate0_blk = Block(
-            "gate",
-            [("sync", residual)] + [("sync", w), (("sync", "gate"), pd_period - w)] * pd_rep,
+        trigger0_blk = Block(
+            "trigger0",
+            [
+                (("sync", self.trigger_channel), trigger_width),
+                ("sync", accum_window - trigger_width),
+            ],
         )
-        # drop1_blk = Block("drop", [(None, accum_window)])
-        gate1_blk = Block(
-            "gate", [(None, residual)] + [(None, w), ("gate", pd_period - w)] * pd_rep
+        sync0_blk = Block("sync0", [("sync", accum_window)])
+        # drop1_blk = Block("drop1", [(None, accum_window)])
+        trigger1_blk = Block(
+            "trigger1",
+            [(self.trigger_channel, trigger_width), (None, accum_window - trigger_width)],
         )
+        # rest1_blk = Block("rest1", [(None, accum_window)])
 
         extended_blks = Blocks()
         laser_duties = []
@@ -178,32 +185,32 @@ class BlocksBuilder(object):
                 blk1.insert(0, (blk1.pattern[0].channels, residual))
             extended_blks.extend(
                 [
-                    blk0.union(drop0_blk).repeat(accum_drop),
-                    blk0.union(gate0_blk).repeat(accum_rep),
-                    blk1.repeat(accum_drop),  # no need to union(drop1_blk)
-                    blk1.union(gate1_blk).repeat(accum_rep),
+                    blk0.union(sync0_blk).suffix("_d").repeat(drop_rep),
+                    blk0.union(trigger0_blk).suffix("_t"),
+                    blk0.union(sync0_blk).repeat(accum_rep - 1),
+                    blk1.suffix("_d").repeat(drop_rep),  # no need to union(drop1_blk)
+                    blk1.union(trigger1_blk).suffix("_t"),
+                    blk1.repeat(accum_rep - 1),  # no need to union(rest1)
                 ]
             )
             # residual is dark: duty becomes slightly lower than laser_width / T
             laser_duties.append(laser_width * rep / accum_window)
             markers.append(extended_blks.total_length())
 
-        return extended_blks, laser_duties, markers, pd_rep * accum_rep
+        return extended_blks, laser_duties, markers
 
     def build_partial(
         self,
         blocks: list[Blocks[Block]],
         accum_window: int,
         accum_rep: int,
-        accum_drop: int,
-        pd_period: int,
+        drop_rep: int,
         laser_width: int,
+        trigger_width: int,
     ):
-        pd_rep = accum_window // pd_period
[…]
