+++ mahos/gui/spodmr.py
@@ -383,7 +383,7 @@ class SPODMRWidget(ClientWidget, Ui_SPODMR):
             [
                 ("accum_window", self.accumwindowBox, 1e3),  # sec to ms
                 ("accum_rep", self.accumrepBox),
-                ("accum_drop", self.accumdropBox),
+                ("drop_rep", self.droprepBox),
                 ("lockin_rep", self.lockinrepBox),
                 ("pd_rate", self.pdrateBox, 1e-3),  # Hz to kHz
                 ("pd_bounds", [self.pd_lbBox, self.pd_ubBox]),
@@ -576,8 +576,8 @@ class SPODMRWidget(ClientWidget, Ui_SPODMR):
         self.sweepsBox.setValue(p.get("sweeps", 0))
         self.accumwindowBox.setValue(p.get("accum_window", 1e-3) * 1e3)  # [sec] ==> [ms]
         self.accumrepBox.setValue(p.get("accum_rep", 1))
-        if "accum_drop" in p:
-            self.accumdropBox.setValue(p["accum_drop"])
+        if "drop_rep" in p:
+            self.droprepBox.setValue(p["drop_rep"])
         if "lockin_rep" in p:
             self.lockinrepBox.setValue(p["lockin_rep"])
         self.pdrateBox.setValue(round(p.get("pd_rate", 1e5) * 1e-3))
@@ -684,10 +684,9 @@ class SPODMRWidget(ClientWidget, Ui_SPODMR):
         params["sweeps"] = self.sweepsBox.value()
         params["accum_window"] = self.accumwindowBox.value() * 1e-3  # [ms] ==> [sec]
         params["accum_rep"] = self.accumrepBox.value()
+        params["drop_rep"] = self.droprepBox.value()
         if params["partial"] == 2:
             params["lockin_rep"] = self.lockinrepBox.value()
-        else:
-            params["accum_drop"] = self.accumdropBox.value()
         params["pd_rate"] = self.pdrateBox.value() * 1e3  # [kHz] ==> [Hz]
         params["pd_bounds"] = [self.pd_lbBox.value(), self.pd_ubBox.value()]
 
@@ -835,6 +834,7 @@ class SPODMRWidget(ClientWidget, Ui_SPODMR):
             self.sweepsBox,
             self.accumwindowBox,
             self.accumrepBox,
+            self.droprepBox,
             self.pdrateBox,
             self.pd_lbBox,
             self.pd_ubBox,
@@ -846,9 +846,6 @@ class SPODMRWidget(ClientWidget, Ui_SPODMR):
         ):
             w.setEnabled(state == BinaryState.IDLE)
 
-        self.accumdropBox.setEnabled(
-            state == BinaryState.IDLE and self.partialBox.currentIndex() != 3
-        )
         self.lockinrepBox.setEnabled(
             state == BinaryState.IDLE and self.partialBox.currentIndex() == 3
         )
