+++ mahos/gui/ui/spodmr.py
@@ -171,9 +171,9 @@ class Ui_SPODMR(object):
         self.lockinrepBox.setMaximum(9999)
         self.lockinrepBox.setObjectName("lockinrepBox")
         self.gridLayout_2.addWidget(self.lockinrepBox, 0, 7, 1, 1)
-        self.accumdropBox = QtWidgets.QSpinBox(parent=self.groupBox_2)
-        self.accumdropBox.setObjectName("accumdropBox")
-        self.gridLayout_2.addWidget(self.accumdropBox, 0, 5, 1, 1)
+        self.droprepBox = QtWidgets.QSpinBox(parent=self.groupBox_2)
+        self.droprepBox.setObjectName("droprepBox")
+        self.gridLayout_2.addWidget(self.droprepBox, 0, 5, 1, 1)
         self.verticalLayout.addWidget(self.groupBox_2)
         self.groupBox_5 = QtWidgets.QGroupBox(parent=self.mainTab)
         self.groupBox_5.setObjectName("groupBox_5")
@@ -525,8 +525,8 @@ class Ui_SPODMR(object):
         SPODMR.setTabOrder(self.loadButton, self.tabWidget)
         SPODMR.setTabOrder(self.tabWidget, self.accumwindowBox)
         SPODMR.setTabOrder(self.accumwindowBox, self.accumrepBox)
-        SPODMR.setTabOrder(self.accumrepBox, self.accumdropBox)
-        SPODMR.setTabOrder(self.accumdropBox, self.lockinrepBox)
+        SPODMR.setTabOrder(self.accumrepBox, self.droprepBox)
+        SPODMR.setTabOrder(self.droprepBox, self.lockinrepBox)
         SPODMR.setTabOrder(self.lockinrepBox, self.sweepsBox)
         SPODMR.setTabOrder(self.sweepsBox, self.pdrateBox)
         SPODMR.setTabOrder(self.pdrateBox, self.pd_lbBox)
@@ -628,7 +628,7 @@ class Ui_SPODMR(object):
         self.pdrateBox.setSuffix(_translate("SPODMR", " kHz"))
         self.pdrateBox.setPrefix(_translate("SPODMR", "rate: "))
         self.lockinrepBox.setPrefix(_translate("SPODMR", "lockin rep: "))
-        self.accumdropBox.setPrefix(_translate("SPODMR", "accum drop: "))
+        self.droprepBox.setPrefix(_translate("SPODMR", "drop rep: "))
         self.groupBox_5.setTitle(_translate("SPODMR", "Sweep parameter (tau or N)"))
         self.NstopLabel.setText(_translate("SPODMR", "stop:"))
         self.label_9.setText(_translate("SPODMR", "tau"))
