### 2023-10-03 88ab747
spodmr: fix switch command name
