### 2023-10-03 45cadc0
cli: add SPODMR for data ls and data print
