### 2023-10-03 5cfe16e
spodmr: add new plot method normalize1
