+++ mahos/msgs/spodmr_msgs.py
@@ -135,8 +135,10 @@ class SPODMRData(BasicMeasData, ComplexDataMixin):
             elif plotmode == "average":
                 return (i0 + i1) / 2
             elif plotmode == "normalize":
-                return (i0 - i1) / (i0 + i1), None
-            else:  # "diff", "data01", ...
+                return (i0 - i1) / (i0 + i1)
+            elif plotmode == "normalize1":
+                return (i0 - i1) / i1
+            else:  # "diff". fall back "data01" to "diff" too.
                 return i0 - i1
 
     def get_fit_xdata(self) -> NDArray | None:
@@ -297,6 +299,8 @@ class SPODMRData(BasicMeasData, ComplexDataMixin):
             return (s0 + s1) / 2, None
         elif plotmode == "normalize":
             return (s0 - s1) / (s0 + s1), None
+        elif plotmode == "normalize1":
+            return (s0 - s1) / s1, None
         elif plotmode == "concatenate":
             return np.column_stack((s0, s1)).reshape(len(s0) * 2), None
         else:
