+++ mahos/gui/ui/spodmr.py
@@ -380,6 +380,7 @@ class Ui_SPODMR(object):
         self.plotmodeBox.addItem("")
         self.plotmodeBox.addItem("")
         self.plotmodeBox.addItem("")
+        self.plotmodeBox.addItem("")
         self.gridLayout.addWidget(self.plotmodeBox, 1, 1, 1, 1)
         self.plotenableBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
         self.plotenableBox.setObjectName("plotenableBox")
@@ -681,7 +682,8 @@ class Ui_SPODMR(object):
         self.plotmodeBox.setItemText(3, _translate("SPODMR", "diff"))
         self.plotmodeBox.setItemText(4, _translate("SPODMR", "average"))
         self.plotmodeBox.setItemText(5, _translate("SPODMR", "normalize"))
-        self.plotmodeBox.setItemText(6, _translate("SPODMR", "concatenate"))
+        self.plotmodeBox.setItemText(6, _translate("SPODMR", "normalize1"))
+        self.plotmodeBox.setItemText(7, _translate("SPODMR", "concatenate"))
         self.plotenableBox.setText(_translate("SPODMR", "enable edit"))
         self.ylogBox.setText(_translate("SPODMR", "log Y"))
         self.label_7.setText(_translate("SPODMR", "Complex conv"))
