### 2023-10-03 2c25864
cli.plot: add spodmr
