+++ mahos/cli/plot.py
@@ -11,6 +11,7 @@ from ..msgs.fit_msgs import str_to_peak_type
 from ..meas.confocal_io import ConfocalIO
 from ..meas.odmr_io import ODMRIO
 from ..meas.podmr_io import PODMRIO
+from ..meas.spodmr_io import SPODMRIO
 from ..meas.iodmr_io import IODMRIO
 from ..meas.hbt_io import HBTIO
 from ..meas.spectroscopy_io import SpectroscopyIO
@@ -124,6 +125,46 @@ def plot_podmr(args):
         for key, value in zip(("sigdelay", "sigwidth", "refdelay", "refwidth"), args.timings):
             plot_params[key] = value * 1e-9  # ns to sec
     for key in ("plotmode", "taumode", "xlogscale", "ylogscale", "fft", "refmode", "refaverage"):
+        value = getattr(args, key)
+        if value is not None:
+            plot_params[key] = value
+    if args.method or args.fit_params:
+        if args.fit_params:
+            fit_params = toml.load(args.fit_params)
+            if "method" not in fit_params:
+                fit_params["method"] = args.method
+        else:
+            fit_params = {"method": args.method}
+    else:
+        fit_params = {}
+
+    params = {
+        "plot": plot_params,
+        "fit": fit_params,
+        "show_fit": args.show_fit,
+        "color0": args.color0,
+        "color1": args.color1,
+        "color_fit": args.color_fit,
+        "marker0": args.marker0,
+        "marker1": args.marker1,
+        "label": args.label,
+        "offset": args.offset,
+        "linewidth": args.linewidth,
+        "linewidth_fit": args.linewidth_fit,
+    }
+    params.update(make_default_params(args))
+    if not args.one_figure:
+        plot_files(args, io.load_data, lambda fn, d: io.export_data(fn, d, params))
+    else:
+        data = [io.load_data(n) for n in args.names]
+        io.export_data(args.one_figure, data, params)
+
+
+def plot_spodmr(args):
+    io = SPODMRIO()
+
+    plot_params = {}
+    for key in ("plotmode", "taumode", "xlogscale", "ylogscale", "fft", "normalize"):
         value = getattr(args, key)
         if value is not None:
             plot_params[key] = value
@@ -516,6 +557,69 @@ def add_podmr_parser(sub_parsers):
     p.set_defaults(func=plot_podmr)
 
 
+def add_spodmr_parser(sub_parsers):
+    p = sub_parsers.add_parser(
+        "spodmr",
+        help=" ".join(
+            [
+                "Pulse odmr with Slow detectors.",
+                "Replot data if -m (--method) is set.",
+            ]
+        ),
+    )
+
+    p.add_argument(
+        "-p",
+        "--print",
+        action="store_true",
+        help="Print plot parameters stored in data file (no plot performed)",
+    )
+    p.add_argument(
+        "-o", "--one-figure", metavar="FILENAME", type=str, help="Filename for one-figure mode"
+    )
+
+    p.add_argument(
+        "-F", "--no-fit", dest="show_fit", action="store_false", help="Don't show fitting result"
+    )
+    p.add_argument(
+        "-M",
+        "--plotmode",
+        type=str,
+        help="[plot] Plot mode (data01|data0|data1|diff|average|normalize|concatenate|ref)",
+    )
+    p.add_argument("-T", "--taumode", type=str, help="[plot] Tau mode (raw|total|freq|index|head)")
+    p.add_argument("--fft", type=strtobool, help="[plot] FFT mode")
+    p.add_argument("--xlog", dest="xlogscale", type=strtobool, help="[plot] logscale X axis")
[…]
