### 2023-10-03 ff3147e
inst_pg_msgs: add BlockSeq type for nested repeat
