+++ mahos/msgs/inst_pg_msgs.py
@@ -420,9 +420,6 @@ class Blocks(UserList):
             blk.extend(b.total_pattern())
         return blk
 
-    def insert_pulse(self, blk_index: int, ptn_index: int, pulse: AcceptedPulse):
-        self.data[blk_index].insert(ptn_index, pulse)
-
     def repeat(self, num: int) -> Blocks[Block]:
         """Return new Blocks[Block] with `num`-times repeat.
 
@@ -579,6 +576,228 @@ class Blocks(UserList):
         return sum([b.total_pattern_num() for b in self.data])
 
     def scale(self, s: int) -> Blocks[Block]:
-        """Scale all the periods in the Blocks."""
+        """Scale all the durations in the Blocks."""
 
         return Blocks([b.scale(s) for b in self.data])
+
+
+class BlockSeq(Message):
+    """Nestable and named sequence of Blocks.
+
+    Methods compatible with Blocks are implemented.
+    Blocks[Block] and BlockSeq are almost equivalent if BlockSeq contains Block only.
+    Blocks[Block] may be a bit handier if you don't need the nested repeat
+    because it supports list-like operations.
+
+    """
+
+    def __init__(
+        self,
+        name: str,
+        data: list[Block | BlockSeq],
+        Nrep: int = 1,
+        trigger: bool = False,
+    ):
+        self.name = name
+        self.data = data
+        self.Nrep = Nrep
+        self.trigger = trigger
+
+    def _new(self, data) -> BlockSeq:
+        return BlockSeq(self.name, data, self.Nrep, self.trigger)
+
+    def collapse(self) -> Block:
+        """Collapse the contained Blocks | BlockSeq to single Block.
+
+        Resultant Block is almost equivalent to the original BlockSeq, but
+        the trigger attributes of non-first Blocks are silently discarded.
+
+        """
+
+        blk = self.data[0].collapse()
+        for b in self.data[1:]:
+            blk.extend(b.collapse())
+        return blk.repeat(self.Nrep).collapse()
+
+    def repeat(self, num: int) -> BlockSeq:
+        """Return new BlockSeq with `num`-times repeat."""
+
+        return BlockSeq(self.name, self.data, self.Nrep * num, self.trigger)
+
+    def total_sequence(self):
+        return self.data * self.Nrep
+
+    def unique_blocks(self) -> Blocks[Block]:
+        """Extract unique (in terms of their names) Blocks inside."""
+
+        ret = Blocks()
+        names = []
+        for bs in self.data:
+            if isinstance(bs, Block):
+                if bs.name in names:
+                    continue
+                ret.append(bs)
+                names.append(bs.name)
+            else:  # BlockSeq
+                blks = bs.unique_blocks()
+                ret.extend([b for b in blks if b.name not in names])
+                names.extend([b.name for b in blks])
+
+        return ret
+
+    def channels(self) -> list[str | int]:
+        """Get list of included channels.
+
+        If types of all channels are identical (all str or all int),
+        the result is sorted.
+        Otherwise the order is unpredictable (because mixed strs and ints cannot be sorted).
+
+        """
+
+        channels = set()
+        for blk_or_seq in self.data:
+            channels.update(blk_or_seq.channels())
+        try:
+            return list(sorted(channels))
+        except TypeError:
+            return list(channels)
+
+    def _decode(self, channel: str | int, max_len: int | None = None) -> list[bool]:
+        ptn = []
+        for blk_or_seq in self.total_sequence():
+            if isinstance(blk_or_seq, Block):
+                ptn.extend(blk_or_seq.decode(channel, max_len=max_len))
+            else:
+                ptn.extend(blk_or_seq._decode(channel, max_len=max_len))
+            if max_len is not None and len(ptn) > max_len:
[…]
