### 2023-10-04 3f87e50
inst.dtg: implement configure_blockseq
