+++ mahos/inst/dtg/dtg.py
@@ -8,8 +8,10 @@ Data Timing Generator module.
 """
 
 from __future__ import annotations
+import typing as T
 from os import path
 import time
+from dataclasses import dataclass
 
 import numpy as np
 
@@ -53,7 +55,7 @@ def gen_block(length, blockid, name):
 
 
 def gen_pattern(blockid, pulses):
-    """generate a python tree for a pattern command.
+    """generate a tree for a pattern command.
 
     pattern is to fill block 'blockid' with the binary sequence 'pulses'.
 
@@ -76,7 +78,7 @@ def gen_pattern(blockid, pulses):
 
 
 def gen_subsequence(subid, subname, steps):
-    """generate a python tree for a subsequence entry.
+    """generate a tree for a subsequence entry.
 
     subsequence entry is based on given id, name, list of subsequences.
 
@@ -122,7 +124,7 @@ def gen_subsequence(subid, subname, steps):
 
 
 def gen_sequence(label, subname, Nrep, goto, jumpto, trigger):
-    """generate a python tree for a sequence entry.
+    """generate a tree for a sequence entry.
 
     sequence entry is based on given label, name of the block/subsequence,
     number of repetitions 'Nrep' and goto label 'goto'.
@@ -166,6 +168,27 @@ def gen_sequence(label, subname, Nrep, goto, jumpto, trigger):
             ["DTG_GOTO_RECID", goto.nbytes, goto],
         ],
     ]
+
+
+@dataclass
+class Sequence:
+    name: str
+    Nrep: int = 1
+    label: str = ""
+    goto: str = ""
+    jumpto: str = ""
+    trigger: bool = False
+
+
+class Step(T.NamedTuple):
+    name: str
+    Nrep: int
+
+
+@dataclass
+class SubSequence:
+    name: str
+    steps: list[Step]
 
 
 class DTG5000(VisaInstrument):
@@ -246,36 +269,39 @@ class DTG5000(VisaInstrument):
 
         return bits
 
-    def get_total_block_length_seq(
-        self, blocks: Blocks[Block], subsequences, sequences
+    def _find_name(self, li: list[Block | SubSequence], name: str) -> Block | SubSequence | None:
+        for elem in li:
+            if elem.name == name:
+                return elem
+        return None
+
+    def _get_total_block_length_seq(
+        self, blocks: Blocks[Block], subsequences: list[SubSequence], sequences: list[Sequence]
     ) -> int | None:
         t = 0
         for i, seq in enumerate(sequences):
-            blks = [b for b in blocks if b.name == seq["name"]]
-            if len(blks) > 0:
-                t += blks[0].raw_length() * seq["Nrep"]
+            blk = self._find_name(blocks, seq.name)
+            if blk is not None:
+                t += blk.raw_length() * seq.Nrep
             else:
                 subt = 0
-                subseq = [s for s in subsequences if s["name"] == seq["name"]]
-                if len(subseq) == 0:
+                subseq = self._find_name(subsequences, seq.name)
+                if subseq is None:
                     self.logger.error(
-                        f"Sequence ({i:d}, {seq['label']}) refers an undef. block/subsequence:"
-                        + f" {seq['name']}"
+                        f"Sequence ({i:d}, {seq.label}) refers to undef block/subseq: {seq.name}"
                     )
                     return None
-
-                for j, step in enumerate(subseq[0]["steps"]):
-                    blks = [b for b in blocks if b.name == step["name"]]
-                    if len(blks) == 0:
+                for j, step in enumerate(subseq.steps):
+                    blk = self._find_name(blocks, step.name)
+                    if blk is None:
                         self.logger.error(
-                            "SubSequence ({:d}, {:s}) refers an undefined block ({:s}).".format(
-                                j, subseq[0]["name"], step["name"]
[…]
