### 2023-10-04 265db67
spodmr: use BlockSeq for pulse pattern
