### 2023-10-04 caa6b1f
spodmr: add unnested block generation
