### 2023-10-04 f915cd3
inst.dtg: add conf strict
