### 2023-10-05 a051227
InstrumentClient: add query for module/class names
