### 2023-10-05 8d5afeb
podmr/spodmr: add config preset
