### 2023-10-05 487b67b
podmr: use offset_base_inc() in generator kernel
- generated pattern can be changed when laser_width or
  widths of init or final is integer multiple of base_width.
- In this case, older code added base_width, but
  offset_base_inc() avoids adding base_width.
- This change makes it better when one wants to ignore base width
  and get exact duration.
