+++ mahos/meas/qdyne.py
@@ -51,21 +51,7 @@ class Qdyne(BasicMeasNode):
     def __init__(self, gconf: dict, name, context=None):
         """Pulse ODMR measurement with Qdyne protocol.
 
-        :param pulser.freq: pulse generator frequency (default: 2.0E9)
-        :type pulser.freq: float
-        :param pulser.reduce_start_divisor: the divisor on start of reducing frequency (default: 2)
-            reduce is done first by this value, and then repeated by 10.
-        :type pulser.reduce_start_divisor: int
-        :param pulser.split_fraction: fraction factor (F) to split the free period for MW phase
-            modulation (default: 4). the period (T) is split into (T // F, T - T // F) and MW phase
-            is switched at T // F. Thus, larger F results in "quicker start" of the phase
-            modulation (depending on hardware, but its response may be a bit slow).
-        :type pulser.split_fraction: int
-        :param pulser.minimum_block_length: minimum block length in generated blocks
-                                            (default: 1000)
-        :type pulser.minimum_block_length: int
-        :param pulser.block_base: block base granularity (default: 4)
-        :type pulser.block_base: int
+
         :param pulser.start_delay: (sec.) delay time before starting PG output. (default: 0.0)
         :type pulser.start_delay: float
 
@@ -73,6 +59,21 @@ class Qdyne(BasicMeasNode):
         :type pulser.raw_events_dir: str
         :param pulser.remove_raw_events: (default: True) Remove RawEvents file after loading it.
         :type pulser.remove_raw_events: bool
+
+        :param pulser.split_fraction: (default: 4) fraction factor (F) to split the free period
+            for MW phase modulation. the period (T) is split into (T // F, T - T // F) and MW phase
+            is switched at T // F. Thus, larger F results in "quicker start" of the phase
+            modulation (depending on hardware, but its response may be a bit slow).
+        :type pulser.split_fraction: int
+        :param pulser.pg_freq: pulse generator frequency (has preset)
+        :type pulser.pg_freq: float
+        :param pulser.reduce_start_divisor: (has preset) the divisor on start of reducing frequency
+            reduce is done first by this value, and then repeated by 10.
+        :type pulser.reduce_start_divisor: int
+        :param pulser.minimum_block_length: (has preset) minimum block length in generated blocks
+        :type pulser.minimum_block_length: int
+        :param pulser.block_base: (has preset) block base granularity of pulse generator.
+        :type pulser.block_base: int
 
         """
 
@@ -92,8 +93,7 @@ class Qdyne(BasicMeasNode):
         else:
             self.tweaker_cli = None
 
-        has_fg = "fg" in self.conf["target"]["servers"]
-        self.worker = Pulser(self.cli, self.logger, has_fg, self.conf.get("pulser", {}))
+        self.worker = Pulser(self.cli, self.logger, self.conf.get("pulser", {}))
         self.io = QdyneIO(self.logger)
         self.buffer: Buffer[tuple[str, QdyneData]] = Buffer()
         self._pub_interval = self.conf.get("pub_interval_sec", 1.0)
