### 2023-10-05 66fb5ad
spodmr: fix laser duty calculation
