### 2023-10-05 a431bf0
io: use param['dpi'] for image export
