+++ mahos/meas/spectroscopy_io.py
@@ -80,6 +80,8 @@ class SpectroscopyIO(object):
         :type params.figsize: tuple[float, float]
         :param params.fontsize: matplotlib fontsize
         :type params.fontsize: float
+        :param params.dpi: matplotlib dpi
+        :type params.dpi: float
         :param params.label: matplotlib labels
         :type params.label: list[str]
         :param params.color_fit: matplotlib colors for fit data
@@ -149,7 +151,7 @@ class SpectroscopyIO(object):
         color = cycle(params.get("color") or [f"C{i}" for i in range(10)])
         marker = cycle(params.get("marker") or ["o", "s", "^", "p", "*", "D", "h", "."])
 
-        fig = plt.figure(figsize=params.get("figsize", (12, 12)))
+        fig = plt.figure(figsize=params.get("figsize", (12, 12)), dpi=params.get("dpi", 100))
         ax = fig.add_subplot(111)
 
         last_n = params.get("last_n", 0)
