+++ mahos/meas/odmr_io.py
@@ -85,6 +85,8 @@ class ODMRIO(object):
         :type params.figsize: tuple[float, float]
         :param params.fontsize: matplotlib fontsize
         :type params.fontsize: float
+        :param params.dpi: matplotlib dpi
+        :type params.dpi: float
         :param params.label: matplotlib labels
         :type params.label: list[str]
         :param params.color_fit: matplotlib colors for fit data
@@ -158,7 +160,7 @@ class ODMRIO(object):
         plt.rcParams["font.size"] = params.get("fontsize", 28)
         normalize_n = params.get("normalize_n", 0)
 
-        fig = plt.figure(figsize=figsize)
+        fig = plt.figure(figsize=figsize, dpi=params.get("dpi", 100))
         ax = fig.add_subplot(111)
 
         ax.set_xlim(x_min * coeff, x_max * coeff)
