### 2023-10-05 b11cbe1
podmr/spodmr: fix definition of 'normalize'
- normalize of two signal is now 'difference divided by mean'.
- this seems to be used commonly in the literature.
