+++ mahos/meas/odmr_io.py
@@ -153,8 +153,12 @@ class ODMRIO(object):
         unit = "MHz"
 
         xs = [(data.params["start"], data.params["stop"]) for data in data_list]
-        x_max = params.get("xmax") or max([max(x) for x in xs])
-        x_min = params.get("xmin") or min([min(x) for x in xs])
+        x_max = params.get("xmax")
+        if x_max is None:
+            x_max = max([max(x) for x in xs])
+        x_min = params.get("xmin")
+        if x_min is None:
+            x_min = min([min(x) for x in xs])
 
         figsize = params.get("figsize", (12, 12))
         plt.rcParams["font.size"] = params.get("fontsize", 28)
