+++ mahos/meas/hbt_io.py
@@ -143,12 +143,10 @@ class HBTIO(object):
         xs = [data.get_xdata(normalize=normalize) for data in data_list]
         x_max = params.get("xmax")
         if x_max is None:
-            x_max = max([max(x) for x in xs])
+            x_max = max([max(x) for x in xs]) * coeff
         x_min = params.get("xmin")
         if x_min is None:
-            x_min = min([min(x) for x in xs])
-        x_max *= coeff
-        x_min *= coeff
+            x_min = min([min(x) for x in xs]) * coeff
 
         plt.rcParams["font.size"] = params.get("fontsize", 28)
 
