### 2023-10-05 bdfc99f
spodmr: remove unused BlocksBuilder
