+++ mahos/meas/spodmr_worker.py
@@ -135,218 +135,6 @@ class Bounds(object):
         self._fg = fg_bounds
 
 
-class BlocksBuilder(object):
-    """Build the PG Blocks for SPODMR from PODMR's Blocks."""
-
-    def __init__(self, trigger_width: float, trigger_channel: str):
-        self.trigger_width = trigger_width
-        self.trigger_channel = trigger_channel
-
-    def build_complementary(
-        self,
-        blocks: list[Blocks[Block]],
-        accum_window: int,
-        accum_rep: int,
-        drop_rep: int,
-        laser_width: int,
-        trigger_width: int,
-    ):
-        trigger0_blk = Block(
-            "trigger0",
-            [
-                (("sync", self.trigger_channel), trigger_width),
-                ("sync", accum_window - trigger_width),
-            ],
-        )
-        sync0_blk = Block("sync0", [("sync", accum_window)])
-        # drop1_blk = Block("drop1", [(None, accum_window)])
-        trigger1_blk = Block(
-            "trigger1",
-            [(self.trigger_channel, trigger_width), (None, accum_window - trigger_width)],
-        )
-        # rest1_blk = Block("rest1", [(None, accum_window)])
-
-        extended_blks = Blocks()
-        laser_duties = []
-        # markers just for PulseMonitor visualization
-        markers = [0]
-
-        for blks in blocks:
-            blks = blks.remove("sync")
-            assert len(blks) == 4
-            blks0: Blocks[Block] = blks[:2]
-            blks1: Blocks[Block] = blks[2:]
-            T = blks0.total_length()
-            assert T == blks1.total_length()
-            rep, residual = accum_window // T, accum_window % T
-            blk0 = blks0.repeat(rep).collapse()
-            blk1 = blks1.repeat(rep).collapse()
-            if residual:
-                blk0.insert(0, (blk0.pattern[0].channels, residual))
-                blk1.insert(0, (blk1.pattern[0].channels, residual))
-            extended_blks.extend(
-                [
-                    blk0.union(sync0_blk).suffix("_d").repeat(drop_rep),
-                    blk0.union(trigger0_blk).suffix("_t"),
-                    blk0.union(sync0_blk).repeat(accum_rep - 1),
-                    blk1.suffix("_d").repeat(drop_rep),  # no need to union(drop1_blk)
-                    blk1.union(trigger1_blk).suffix("_t"),
-                    blk1.repeat(accum_rep - 1),  # no need to union(rest1)
-                ]
-            )
-            # residual is dark: duty becomes slightly lower than laser_width / T
-            laser_duties.append(laser_width * rep / accum_window)
-            markers.append(extended_blks.total_length())
-
-        return extended_blks, laser_duties, markers
-
-    def build_partial(
-        self,
-        blocks: list[Blocks[Block]],
-        accum_window: int,
-        accum_rep: int,
-        drop_rep: int,
-        laser_width: int,
-        trigger_width: int,
-    ):
-        trigger_blk = Block(
-            "trigger",
-            [(self.trigger_channel, trigger_width), (None, accum_window - trigger_width)],
-        )
-
-        extended_blks = Blocks()
-        laser_duties = []
-        # markers just for PulseMonitor visualization
-        markers = [0]
-
-        for blks in blocks:
-            blks = blks.remove("sync")
-            assert len(blks) == 2
-            T = blks.total_length()
-            rep, residual = accum_window // T, accum_window % T
-            blk = blks.repeat(rep).collapse()
-            if residual:
-                blk.insert(0, (blk.pattern[0].channels, residual))
-            extended_blks.extend(
-                [
-                    blk.suffix("_d").repeat(drop_rep),
[…]
