### 2023-10-06 96b8182
docs: add docs of SPODMR
