+++ docs/_sources/inst/generated/mahos.inst.server.InstrumentClient.rst.txt
@@ -18,6 +18,7 @@
    
       ~InstrumentClient.__init__
       ~InstrumentClient.call
+      ~InstrumentClient.class_name
       ~InstrumentClient.configure
       ~InstrumentClient.get
       ~InstrumentClient.get_param_dict
@@ -25,6 +26,8 @@
       ~InstrumentClient.help
       ~InstrumentClient.is_locked
       ~InstrumentClient.lock
+      ~InstrumentClient.module_class_names
+      ~InstrumentClient.module_name
       ~InstrumentClient.pause
       ~InstrumentClient.release
       ~InstrumentClient.reset
