+++ docs/_sources/meas/generated/mahos.meas.spodmr.SPODMR.rst.txt
@@ -1,0 +1,37 @@
+﻿mahos.meas.spodmr.SPODMR
+========================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.meas.spodmr
+
+.. autoclass:: SPODMR
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~SPODMR.__init__
+      ~SPODMR.change_state
+      ~SPODMR.close_resources
+      ~SPODMR.export_data
+      ~SPODMR.get_param_dict
+      ~SPODMR.get_param_dict_labels
+      ~SPODMR.handle_req
+      ~SPODMR.load_data
+      ~SPODMR.main
+      ~SPODMR.save_data
+      ~SPODMR.update_plot_params
+      ~SPODMR.validate
+      ~SPODMR.wait
+   
+   
+
+   
+   
+   
