+++ docs/_sources/gui/generated/mahos.gui.spodmr.SPODMRGUI.rst.txt
@@ -1,0 +1,25 @@
+﻿mahos.gui.spodmr.SPODMRGUI
+==========================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.gui.spodmr
+
+.. autoclass:: SPODMRGUI
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~SPODMRGUI.init_widget
+   
+   
+
+   
+   
+   
