+++ docs/_sources/gui/generated/mahos.gui.spodmr.SPODMRMainWindow.rst.txt
@@ -1,0 +1,31 @@
+﻿mahos.gui.spodmr.SPODMRMainWindow
+=================================
+
+.. autosummary class template to exclude inherited_members
+   The other parts are same as default.
+
+.. currentmodule:: mahos.gui.spodmr
+
+.. autoclass:: SPODMRMainWindow
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~SPODMRMainWindow.__init__
+      ~SPODMRMainWindow.closeEvent
+   
+   
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
+   

[…]
