+++ mahos/inst/dtg/__init__.py
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-Data Timing Generator module.
-
-.. This file is a part of MAHOS project.
-
-"""
-
-from .dtg import DTG5000, DTG5078, DTG5274, DTG5334, DTG5274_mock
-
-__all__ = ["DTG5000", "DTG5078", "DTG5274", "DTG5334", "DTG5274_mock"]
