+++ docs/_sources/inst/generated/mahos.inst.pg.DTG5334.rst.txt
@@ -1,12 +1,12 @@
-﻿mahos.inst.dtg.DTG5078
-======================
+﻿mahos.inst.pg.DTG5334
+=====================
 
 .. autosummary class template to exclude inherited_members
    The other parts are same as default.
 
-.. currentmodule:: mahos.inst.dtg
+.. currentmodule:: mahos.inst.pg
 
-.. autoclass:: DTG5078
+.. autoclass:: DTG5334
 
    
    .. automethod:: __init__
@@ -16,7 +16,6 @@
 
    .. autosummary::
    
-      ~DTG5078.block_granularity
    
    
 
@@ -26,8 +25,8 @@
 
    .. autosummary::
    
-      ~DTG5078.MAX_BLOCK_LENGTH
-      ~DTG5078.MAX_FREQ
-      ~DTG5078.MIN_BLOCK_LENGTH
+      ~DTG5334.MAX_BLOCK_LENGTH
+      ~DTG5334.MAX_FREQ
+      ~DTG5334.MIN_BLOCK_LENGTH
    
    
