### 2023-10-06 b4f3be9
docs: fix broken docstring (param, not params)
