### 2023-10-09 e64fc8b
odmr: use ClockSource for analog PD sampling
