+++ mahos/meas/odmr_worker.py
@@ -18,6 +18,8 @@ from ..msgs.inst_pg_msgs import Block, Blocks, TriggerType
 from ..inst.sg_interface import SGInterface
 from ..inst.pg_interface import PGInterface
 from ..inst.pd_interface import PDInterface
+from ..inst.daq_interface import ClockSourceInterface
+from ..util.conf import PresetLoader
 from .common_worker import Worker
 
 
@@ -25,25 +27,52 @@ class Sweeper(Worker):
     """Worker for ODMR Sweep."""
 
     def __init__(self, cli, logger, conf: dict):
-        Worker.__init__(self, cli, logger)
+        Worker.__init__(self, cli, logger, conf)
+        self.load_conf_preset(cli)
+
         self.sg = SGInterface(cli, "sg")
         self.pg = PGInterface(cli, "pg")
-        self.pd_names = conf.get("pd_names", ["pd0", "pd1"])
+        self.pd_names = self.conf.get("pd_names", ["pd0", "pd1"])
         self.pds = [PDInterface(cli, n) for n in self.pd_names]
+        self._pd_analog = self.conf.get("pd_analog", False)
+        if self._pd_analog:
+            self.clock = ClockSourceInterface(cli, self.conf.get("clock_name", "clock"))
+        else:
+            self.clock = None
         self.add_instruments(self.sg, self.pg, *self.pds)
 
-        if "pd_clock" not in conf:
-            raise KeyError("sweeper.pd_clock must be given")
-        self._pd_clock = conf["pd_clock"]
-        self._minimum_block_length = conf.get("minimum_block_length", 1000)
-        self._start_delay = conf.get("start_delay", 0.0)
-        self._drop_first = conf.get("drop_first", 0)
-        self._sg_first = conf.get("sg_first", False)
-        self._pd_analog = conf.get("pd_analog", False)
+        self.check_required_conf(["pd_clock", "block_base", "minimum_block_length"])
+        self._pd_clock = self.conf["pd_clock"]
+        self._minimum_block_length = self.conf["minimum_block_length"]
+        self._block_base = self.conf["block_base"]
+        self._start_delay = self.conf.get("start_delay", 0.0)
+        self._drop_first = self.conf.get("drop_first", 0)
+        self._sg_first = self.conf.get("sg_first", False)
         self._continue_mw = False
-        self._conf = conf
 
         self.data = ODMRData()
+
+    def load_conf_preset(self, cli):
+        loader = PresetLoader(self.logger, PresetLoader.Mode.FORWARD)
+        loader.add_preset(
+            "DTG",
+            [
+                ("block_base", 4),
+                ("pg_freq_cw", 1.0e6),
+                ("pg_freq_pulse", 2.0e9),
+                ("minimum_block_length", 1000),
+            ],
+        )
+        loader.add_preset(
+            "PulseStreamer",
+            [
+                ("block_base", 1),
+                ("pg_freq_cw", 1.0e6),
+                ("pg_freq_pulse", 1.0e9),
+                ("minimum_block_length", 1),
+            ],
+        )
+        loader.load_preset(self.conf, cli.class_name("pg"))
 
     def get_param_dict_labels(self) -> list:
         return ["cw", "pulse"]
@@ -51,7 +80,7 @@ class Sweeper(Worker):
     def get_param_dict(self, label: str) -> P.ParamDict[str, P.PDValue] | None:
         if label == "cw":
             timing = P.ParamDict(
-                time_window=P.FloatParam(self._conf.get("time_window", 10e-3), 0.1e-3, 1.0)
+                time_window=P.FloatParam(self.conf.get("time_window", 10e-3), 0.1e-3, 1.0)
             )
[…]
