### 2023-10-10 #5720 @arfon
:wave: @aquilesC @mcbed @caldarolamartin @sidihamady – would any of you be willing to review this submission for JOSS? The submission under consideration is MAHOS: Measurement Automation Handling and Orchestration System, https://github.com/ToyotaCRDL/mahos

Based on your experience, and possibly related submissions by yourselves already published here, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
