### 2023-10-10 #5938 @editorialbot
[REVIEW]: MAHOS: Measurement Automation Handling and Orchestration System
**Submitting author:** @ktahar (<a href="http://orcid.org/0000-0002-9474-8970">Kosuke Tahara</a>)
**Repository:** https://github.com/ToyotaCRDL/mahos
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.3
