### 2023-10-12 681b6e3
add generic piezo class AnalogPiezo
