+++ mahos/inst/piezo.py
@@ -7,6 +7,8 @@ Piezo Stage module.
 
 available class(es):
     E727_3_USB - PI E727.3 Piezo controller using PIPython (GCS USB).
+    E727_3_USB_AO - PI E727.3 Piezo controller using PIPython and DAQ AnalogOut.
+    AnalogPiezo - Generic Piezo Driver using DAQ AnalogIn and AnalogOut.
 
 """
 
@@ -18,7 +20,7 @@ from pipython import GCSDevice
 
 from .instrument import Instrument
 from .exceptions import InstError
-from .daq import AnalogOut
+from .daq import AnalogIn, AnalogOut
 from ..msgs.confocal_msgs import Axis
 
 
@@ -41,7 +43,8 @@ class E727_3_USB(Instrument):
     def __init__(self, name, conf=None, prefix=None):
         Instrument.__init__(self, name, conf=conf, prefix=prefix)
         self.init_device()
-        self.init_param()
+        self.init_axes()
+        self.range = [self.limit[ax] for ax in self.AXIS_IDs]
         if self.conf.get("start_servo", True):
             self.start_servo()
 
@@ -56,7 +59,7 @@ class E727_3_USB(Instrument):
         self.dev.ConnectUSB(devs[0])
         self.logger.info(f"Connected to {devs[0]}.")
 
-    def init_param(self):
+    def init_axes(self):
         limit = self.conf.get("limit_um", ((0.0, 100.0), (0.0, 100.0), (0.0, 100.0)))
         for low, high in limit:
             if low < 0.0 or high > 200.0:
@@ -69,7 +72,6 @@ class E727_3_USB(Instrument):
         self.Axis_to_ID = {Axis.X: self.AXIS_ID_X, Axis.Y: self.AXIS_ID_Y, Axis.Z: self.AXIS_ID_Z}
         self.limit = {ax: r for ax, r in zip(self.AXIS_IDs, limit)}
         self.target = {ax: 0.0 for ax in self.AXIS_IDs}
-        self.range = [self.limit[ax] for ax in self.AXIS_IDs]
 
         self.logger.info(f"got AXIS_IDs: {self.AXIS_IDs}")
 
@@ -225,7 +227,46 @@ class E727_3_USB(Instrument):
             return None
 
 
-class E727_3_USB_AO(E727_3_USB):
+class TransformMixin(object):
+    def init_transform_matrix(self, limit: list[tuple[float, float]]):
+        """Init transform matrices and target space range: self.T, self.Tinv and self.range.
+
+        limit is cubic-range in command space.
+
+        When self.conf["transform"] is undefined,
+        This method sets identity transforms and self.range = limit.
+
+        Otherwise, self.T is set as self.conf["transform"], which is the linear transformation
+        from target space (x, y, z) to command space (X, Y, Z).
+        Also requires user defined cubic-range in target space: self.conf["range_um"].
+
+        """
+
+        if "transform" not in self.conf:
+            self.T = self.Tinv = np.eye(3)
+            self.range = limit
+            return
+
+        self.T = np.array(self.conf["transform"], dtype=np.float64)
+        self.Tinv = np.linalg.inv(self.T)
+        # self.conf["range_um"] defines cubic-range in target space,
+        # limit is cubic-range in command space
+        self.check_required_conf(["range_um"])
+        low, high = np.array(limit).T
+        # validate that the range after transform is inside the limit.
+        for corner_point in product(*self.conf["range_um"]):
+            p = self.T @ np.array(corner_point)
+            if any(p < low) or any(high < p):
+                msg = f"range corner point {p} goes outside command space limit {low}, {high}."
+                self.logger.error(msg)
+                raise ValueError(msg)
+            else:
+                self.logger.debug("corner point: {:.4f}, {:.4f}, {:.4f}".format(*p))
[…]
