### 2023-10-13 f85ad74
inst.piezo: add base class BasePiezo3Axes
