+++ mahos/inst/piezo.py
@@ -24,15 +24,163 @@ from .daq import AnalogIn, AnalogOut
 from ..msgs.confocal_msgs import Axis
 
 
-class E727_3_USB(Instrument):
+class BasePiezo3Axes(Instrument):
+    """Base Class for Piezo with 3 axes.
+
+    Listed parameters (confs) are common for all the child classes.
+
+    :param transform: (default: identity matrix) a 3x3 linear transformation matrix
+        from (real) target coordinate to command-value coordinate.
+        This is usable for tilt correction to compensate the cross-talks in piezo axes.
+    :type transform: list[list[float]]
+    :param limit_um: (default: [[0.0, 100.0], [0.0, 100.0], [0.0, 100.0)]) limits of positions
+        in command coordinate in the following format:
+        [[Xmin, Xmax], [Ymin, Ymax], [Ymin, Ymax]].
+    :param range_um: travel ranges in target coordinate in the following format:
+        [[xmin, xmax], [ymin, ymax], [zmin, zmax]].
+        This should be identical to limit_um if transform is default (identity matrix), i.e.,
+        the target corrdinate and the command-value coordinate are identical.
+    :type range_um: list[tuple[float, float]]
+
+    """
+
+    DEFAULT_LIMIT = ((0.0, 100.0), (0.0, 100.0), (0.0, 100.0))
+    LIMIT_BOUNDS = ((0.0, 200.0), (0.0, 200.0), (0.0, 200.0))
+
+    def __init__(self, name, conf, prefix=None):
+        Instrument.__init__(self, name, conf=conf, prefix=prefix)
+
+        # The cubic-limit in command space
+        limit = self.conf.get("limit_um", self.DEFAULT_LIMIT)
+        for (low, high), (lb, ub) in zip(limit, self.LIMIT_BOUNDS):
+            if low < lb or high > ub:
+                raise ValueError(f"limit_um is out of hard-bounds ({lb}, {ub})")
+        self.limit: dict[Axis, tuple[float, float]] = {ax: r for ax, r in zip(Axis, limit)}
+        self.target: dict[Axis, float] = {ax: 0.0 for ax in Axis}
+        self.init_transform_matrix()
+
+    def set_target_pos(self, positions: list[float], axes: list[Axis] | None = None) -> bool:
+        """update self.target and move toward it."""
+
+        if axes is None:
+            axes = list(Axis)
+        positions = list(positions)
+        for i, ax in enumerate(axes):
+            if ax not in Axis:
+                self.logger.error(f"Unknown Axis: {ax}")
+                return False
+            mn, mx = self.limit[ax]
+            if positions[i] < mn:
+                positions[i] = mn
+            if positions[i] > mx:
+                positions[i] = mx
+        for ax, pos in zip(axes, positions):
+            self.target[ax] = pos
+        return self.move(axes)
+
+    def move(self, axes) -> bool:
+        raise NotImplementedError("move() is not implemented")
+
+    def set_target_X(self, pos) -> bool:
+        return self.set_target_pos([pos], [Axis.X])
+
+    def set_target_Y(self, pos) -> bool:
+        return self.set_target_pos([pos], [Axis.Y])
+
+    def set_target_Z(self, pos) -> bool:
+        return self.set_target_pos([pos], [Axis.Z])
+
+    def set_target(self, value: dict | list[float]):
+        if isinstance(value, dict) and "ax" in value and "pos" in value:
+            ax, p = value["ax"], value["pos"]
+            if all([isinstance(v, (tuple, list)) for v in (ax, p)]) and len(ax) == len(p):
+                axes, pos = ax, p
+            elif all([not isinstance(v, (tuple, list)) for v in (ax, p)]):
+                axes, pos = [ax], [p]
+            else:
+                self.logger.error('set("target", value): invalid format for value in dict.')
[…]
