### 2023-10-13 78ddc5c
daq: set buffer_size for BufferedEdgeCounter
