### 2023-10-13 7d5a953
confocal.tracer: increase buffer
