### 2023-10-13 e4ad777
inst: remove samples_margin in input classes
- classes: BufferedEdgeCounter and AnalogIn
