### 2023-10-13 073fb24
inst.daq: load preset using device type
