### 2023-10-13 b6fd4c8
inst.piezo: change default write_and_start to True
- now default setting is same as that for USB-6363
