### 2023-10-13 c67ede1
BufferedEdgeCounter: double buffer_size in gate mode
