+++ mahos/inst/daq.py
@@ -541,31 +541,51 @@ class AnalogIn(ConfigurableTask):
 
     :param lines: list of strings to designate DAQ's physical channels.
     :type lines: list[str]
-    :param buffer_size: (default: 10000) Software buffer (queue) size.
-    :type buffer_size: int
+    :param queue_size: (default: 10000) Software buffer (queue) size for acquired data.
+    :type queue_size: int
 
     :param clock_mode: If True (False), configures as clock-mode (on-demand-mode).
     :type clock_mode: bool
     :param bounds: bounds (min, max) values of the expected input voltages per channels.
     :type bounds: list[tuple[float, float]]
-    :param finite: (default True) Switch if finite mode or infinite mode.
+
+    :param clock: DAQ terminal for clock.
+    :type clock: str
+    :param cb_samples: The number of samples for each callback.
+    :type cb_samples: int
+    :param samples: The number of samples. In infinite mode (finite=False), this is used for
+        automatic buffer allocation.
+    :type samples: int
+    :param buffer_size: The input buffer size. If not given, buffer size is not set manually
+        (automatic buffer allocation is used).
+    :type buffer_size: int | None
+
+    :param rate: (default: 10000.0) Estimated sampling rate.
+    :type rate: float
+    :param clock_dir: (default: True) True (False) for rising (falling) edge clock.
+    :type clock_dir: bool
+    :param finite: (default: True) Switch if finite mode or infinite mode.
     :type finite: bool
-    :param every: (default False) Switch if every1 mode or not.
+    :param every: (default: False) Switch if every1 mode or not.
     :type every: bool
-    :param oversample: (default 1) cb_samples and samples are multiplied by `oversample`,
-                       and everyN_handler is passed mean of `oversample` readings.
-    :type oversample: int
-    :param stamp: (default False) Attach timestamp for each samples.
+    :param stamp: (default: False) Attach timestamp for each samples.
     :type stamp: bool
-    :param drop_first: (default 0) drop the data on first N callbacks.
+    :param drop_first: (default: 0) drop the data on first N callbacks.
     :type drop_first: int
+
+    :param oversample: (default: 1) cb_samples, samples, and buffer_size are multiplied by
+                       `oversample`, and everyN_handler is passed mean of `oversample` readings.
+    :type oversample: int
 
     Finite or Infinite mode
     -----------------------
 
     In finite mode (finite=True), task is finished with `samples`.
     In inifinite mode (finite=False), task continues infinitely but it can be stopped by stop().
-    (`samples` is used to determine the buffer size.)
+
+    On some device, we cannot call EveryNCallBack at arbitrary `cb_samples` in infinite mode.
+    (For example, buffer size must be even integer multiple of cb_samples.)
+    Setting buffer_size explicitly, or every=True may resolve such a situation.
 
     Callback handlers
     -----------------
@@ -574,7 +594,7 @@ class AnalogIn(ConfigurableTask):
     If every1 mode (every=True), EveryNCallBack is called with N=1 regardress of `cb_samples`.
     And every1_handler is called at each callbacks.
     EveryNCallBack aquires data and call everyN_handler at each `cb_samples` callbacks.
-    If every1 mode is not active (every=False), EveryNCallBack is called
+    If every1 mode is inactive (every=False), EveryNCallBack is called
     after acquiring `cb_samples`, and everyN_handler is called for each callback.
     In this mode every1_handler is never called.
 
@@ -586,8 +606,8 @@ class AnalogIn(ConfigurableTask):
         self.check_required_conf(("lines",))
         self.lines = conf["lines"]
 
[…]
