+++ docs/_sources/inst/generated/mahos.inst.piezo.E727_3_USB.rst.txt
@@ -24,21 +24,15 @@
       ~E727_3_USB.get_error
       ~E727_3_USB.get_pos
       ~E727_3_USB.get_pos_ont
-      ~E727_3_USB.get_range
-      ~E727_3_USB.get_target
+      ~E727_3_USB.init_axes
       ~E727_3_USB.init_device
-      ~E727_3_USB.init_param
+      ~E727_3_USB.move
       ~E727_3_USB.query_on_target
       ~E727_3_USB.query_pos
       ~E727_3_USB.query_servo
       ~E727_3_USB.query_target_pos
       ~E727_3_USB.set
       ~E727_3_USB.set_servo
-      ~E727_3_USB.set_target
-      ~E727_3_USB.set_target_X
-      ~E727_3_USB.set_target_Y
-      ~E727_3_USB.set_target_Z
-      ~E727_3_USB.set_target_pos
       ~E727_3_USB.start
       ~E727_3_USB.start_servo
       ~E727_3_USB.stop
@@ -46,6 +40,11 @@
    
    
 
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
    
    
    
