+++ docs/_sources/inst/generated/mahos.inst.daq.AnalogOut.rst.txt
@@ -22,6 +22,7 @@
       ~AnalogOut.get
       ~AnalogOut.get_buffer_size
       ~AnalogOut.get_onboard_buffer_size
+      ~AnalogOut.load_conf_preset
       ~AnalogOut.set
       ~AnalogOut.set_output
       ~AnalogOut.set_output_once
