### 2023-10-16 c9e0420
podmr/spodmr: add new pattern T1
- pattern to measure T1 (spin-lattice relaxation time).
- pattern0 is laser - tau (... - laser) pattern.
- pattern1 is laser - 180pulse - tau pattern.
- duration of pattern0 and pattern1 is made same
  (180pulse duration is implied in pattern0) due to spodmr.
- already had laser - tau (... - laser) pattern in Rabi's pattern1
  for example, but it's much better to have complementary
  measurement protocol with 180 deg. pulse.
