+++ mahos/gui/ui/spodmr.py
@@ -132,6 +132,7 @@ class Ui_SPODMR(object):
         self.methodBox.addItem("")
         self.methodBox.addItem("")
         self.methodBox.addItem("")
+        self.methodBox.addItem("")
         self.gridLayout_2.addWidget(self.methodBox, 3, 1, 1, 1)
         self.partialBox = QtWidgets.QComboBox(parent=self.groupBox_2)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.MinimumExpanding, QtWidgets.QSizePolicy.Policy.Fixed)
@@ -605,21 +606,22 @@ class Ui_SPODMR(object):
         self.pd_lbBox.setPrefix(_translate("SPODMR", "lower: "))
         self.pd_lbBox.setSuffix(_translate("SPODMR", " V"))
         self.methodBox.setItemText(0, _translate("SPODMR", "Rabi"))
-        self.methodBox.setItemText(1, _translate("SPODMR", "FID"))
-        self.methodBox.setItemText(2, _translate("SPODMR", "Spin Echo (envelope)"))
-        self.methodBox.setItemText(3, _translate("SPODMR", "Spin Echo (different tau)"))
-        self.methodBox.setItemText(4, _translate("SPODMR", "CP"))
-        self.methodBox.setItemText(5, _translate("SPODMR", "CPMG"))
-        self.methodBox.setItemText(6, _translate("SPODMR", "XY4"))
-        self.methodBox.setItemText(7, _translate("SPODMR", "XY8"))
-        self.methodBox.setItemText(8, _translate("SPODMR", "XY16"))
-        self.methodBox.setItemText(9, _translate("SPODMR", "180 pulse train"))
-        self.methodBox.setItemText(10, _translate("SPODMR", "Spin Echo (90 pulse sweep)"))
-        self.methodBox.setItemText(11, _translate("SPODMR", "Spin Lock"))
-        self.methodBox.setItemText(12, _translate("SPODMR", "XY8 correlation"))
-        self.methodBox.setItemText(13, _translate("SPODMR", "XY8 correlation (1 flip)"))
-        self.methodBox.setItemText(14, _translate("SPODMR", "XY8 correlation (N flip)"))
-        self.methodBox.setItemText(15, _translate("SPODMR", "ddgate: XY8-CPMG-XY8"))
+        self.methodBox.setItemText(1, _translate("SPODMR", "T1"))
+        self.methodBox.setItemText(2, _translate("SPODMR", "FID"))
+        self.methodBox.setItemText(3, _translate("SPODMR", "Spin Echo (envelope)"))
+        self.methodBox.setItemText(4, _translate("SPODMR", "Spin Echo (different tau)"))
+        self.methodBox.setItemText(5, _translate("SPODMR", "CP"))
+        self.methodBox.setItemText(6, _translate("SPODMR", "CPMG"))
+        self.methodBox.setItemText(7, _translate("SPODMR", "XY4"))
+        self.methodBox.setItemText(8, _translate("SPODMR", "XY8"))
+        self.methodBox.setItemText(9, _translate("SPODMR", "XY16"))
+        self.methodBox.setItemText(10, _translate("SPODMR", "180 pulse train"))
+        self.methodBox.setItemText(11, _translate("SPODMR", "Spin Echo (90 pulse sweep)"))
+        self.methodBox.setItemText(12, _translate("SPODMR", "Spin Lock"))
+        self.methodBox.setItemText(13, _translate("SPODMR", "XY8 correlation"))
+        self.methodBox.setItemText(14, _translate("SPODMR", "XY8 correlation (1 flip)"))
[…]
