### 2023-10-16 dd4af21
podmr/spodmr: add new plot param flipY
- if flipY == True, pattern0 data and pattern1 data
  are flipped in get_ydata().
- this makes better plot in some cases with color centers
  other than diamond NVC.
