+++ mahos/gui/ui/podmr.py
@@ -547,6 +547,9 @@ class Ui_PODMR(object):
         self.gridLayout.addWidget(self.label_3, 3, 0, 1, 1)
         spacerItem4 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Minimum)
         self.gridLayout.addItem(spacerItem4, 2, 5, 1, 1)
+        self.flipYBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
+        self.flipYBox.setObjectName("flipYBox")
+        self.gridLayout.addWidget(self.flipYBox, 1, 4, 1, 1)
         self.verticalLayout.addWidget(self.groupBox_3)
         spacerItem5 = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Policy.Minimum, QtWidgets.QSizePolicy.Policy.Expanding)
         self.verticalLayout.addItem(spacerItem5)
@@ -691,7 +694,8 @@ class Ui_PODMR(object):
         PODMR.setTabOrder(self.iqdelayBox, self.plotenableBox)
         PODMR.setTabOrder(self.plotenableBox, self.plotmodeBox)
         PODMR.setTabOrder(self.plotmodeBox, self.taumodeBox)
-        PODMR.setTabOrder(self.taumodeBox, self.sigdelayBox)
+        PODMR.setTabOrder(self.taumodeBox, self.flipYBox)
+        PODMR.setTabOrder(self.flipYBox, self.sigdelayBox)
         PODMR.setTabOrder(self.sigdelayBox, self.sigwidthBox)
         PODMR.setTabOrder(self.sigwidthBox, self.refdelayBox)
         PODMR.setTabOrder(self.refdelayBox, self.refwidthBox)
@@ -701,13 +705,13 @@ class Ui_PODMR(object):
         PODMR.setTabOrder(self.autoadjustBox, self.xlogBox)
         PODMR.setTabOrder(self.xlogBox, self.ylogBox)
         PODMR.setTabOrder(self.ylogBox, self.fftBox)
-        PODMR.setTabOrder(self.fftBox, self.fg_waveBox)
+        PODMR.setTabOrder(self.fftBox, self.fg_disableButton)
+        PODMR.setTabOrder(self.fg_disableButton, self.fg_cwButton)
+        PODMR.setTabOrder(self.fg_cwButton, self.fg_gateButton)
+        PODMR.setTabOrder(self.fg_gateButton, self.fg_waveBox)
         PODMR.setTabOrder(self.fg_waveBox, self.fg_amplBox)
         PODMR.setTabOrder(self.fg_amplBox, self.fg_phaseBox)
         PODMR.setTabOrder(self.fg_phaseBox, self.fg_freqBox)
-        PODMR.setTabOrder(self.fg_freqBox, self.fg_disableButton)
-        PODMR.setTabOrder(self.fg_disableButton, self.fg_cwButton)
-        PODMR.setTabOrder(self.fg_cwButton, self.fg_gateButton)
 
     def retranslateUi(self, PODMR):
         _translate = QtCore.QCoreApplication.translate
@@ -847,6 +851,7 @@ class Ui_PODMR(object):
         self.label_8.setText(_translate("PODMR", "mode"))
         self.plotenableBox.setText(_translate("PODMR", "enable edit"))
         self.label_3.setText(_translate("PODMR", "referencing"))
+        self.flipYBox.setText(_translate("PODMR", "flip Y"))
         self.tabWidget.setTabText(self.tabWidget.indexOf(self.mainTab), _translate("PODMR", "Main"))
[…]
