+++ mahos/gui/ui/spodmr.py
@@ -367,6 +367,39 @@ class Ui_SPODMR(object):
         self.groupBox_3.setObjectName("groupBox_3")
         self.gridLayout = QtWidgets.QGridLayout(self.groupBox_3)
         self.gridLayout.setObjectName("gridLayout")
+        self.label_7 = QtWidgets.QLabel(parent=self.groupBox_3)
+        self.label_7.setObjectName("label_7")
+        self.gridLayout.addWidget(self.label_7, 4, 0, 1, 1)
+        self.plotenableBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
+        self.plotenableBox.setObjectName("plotenableBox")
+        self.gridLayout.addWidget(self.plotenableBox, 0, 0, 1, 1)
+        self.fftBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
+        self.fftBox.setObjectName("fftBox")
+        self.gridLayout.addWidget(self.fftBox, 3, 3, 1, 1)
+        self.taumodeBox = QtWidgets.QComboBox(parent=self.groupBox_3)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.MinimumExpanding, QtWidgets.QSizePolicy.Policy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.taumodeBox.sizePolicy().hasHeightForWidth())
+        self.taumodeBox.setSizePolicy(sizePolicy)
+        self.taumodeBox.setObjectName("taumodeBox")
+        self.taumodeBox.addItem("")
+        self.taumodeBox.addItem("")
+        self.taumodeBox.addItem("")
+        self.taumodeBox.addItem("")
+        self.gridLayout.addWidget(self.taumodeBox, 1, 3, 1, 1)
+        self.label_14 = QtWidgets.QLabel(parent=self.groupBox_3)
+        self.label_14.setObjectName("label_14")
+        self.gridLayout.addWidget(self.label_14, 3, 0, 1, 1)
+        self.complexBox = QtWidgets.QComboBox(parent=self.groupBox_3)
+        self.complexBox.setObjectName("complexBox")
+        self.complexBox.addItem("")
+        self.complexBox.addItem("")
+        self.complexBox.addItem("")
+        self.complexBox.addItem("")
+        self.gridLayout.addWidget(self.complexBox, 4, 1, 1, 1)
+        spacerItem2 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Minimum)
+        self.gridLayout.addItem(spacerItem2, 1, 6, 1, 1)
         self.plotmodeBox = QtWidgets.QComboBox(parent=self.groupBox_3)
         sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Policy.MinimumExpanding, QtWidgets.QSizePolicy.Policy.Fixed)
         sizePolicy.setHorizontalStretch(0)
@@ -383,55 +416,25 @@ class Ui_SPODMR(object):
         self.plotmodeBox.addItem("")
         self.plotmodeBox.addItem("")
         self.gridLayout.addWidget(self.plotmodeBox, 1, 1, 1, 1)
-        self.plotenableBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
-        self.plotenableBox.setObjectName("plotenableBox")
-        self.gridLayout.addWidget(self.plotenableBox, 0, 0, 1, 1)
-        self.ylogBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
[…]
