### 2023-10-16 d790db8
spodmr: add new plot param offset
