+++ mahos/gui/ui/spodmr.py
@@ -416,10 +416,6 @@ class Ui_SPODMR(object):
         self.plotmodeBox.addItem("")
         self.plotmodeBox.addItem("")
         self.gridLayout.addWidget(self.plotmodeBox, 1, 1, 1, 1)
-        self.normalizeBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
-        self.normalizeBox.setChecked(True)
-        self.normalizeBox.setObjectName("normalizeBox")
-        self.gridLayout.addWidget(self.normalizeBox, 1, 4, 1, 1)
         self.label_8 = QtWidgets.QLabel(parent=self.groupBox_3)
         self.label_8.setObjectName("label_8")
         self.gridLayout.addWidget(self.label_8, 1, 0, 1, 1)
@@ -432,9 +428,16 @@ class Ui_SPODMR(object):
         self.label_2 = QtWidgets.QLabel(parent=self.groupBox_3)
         self.label_2.setObjectName("label_2")
         self.gridLayout.addWidget(self.label_2, 1, 2, 1, 1)
+        self.normalizeBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
+        self.normalizeBox.setChecked(True)
+        self.normalizeBox.setObjectName("normalizeBox")
+        self.gridLayout.addWidget(self.normalizeBox, 2, 2, 1, 1)
         self.flipYBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
         self.flipYBox.setObjectName("flipYBox")
-        self.gridLayout.addWidget(self.flipYBox, 4, 2, 1, 1)
+        self.gridLayout.addWidget(self.flipYBox, 2, 1, 1, 1)
+        self.offsetBox = SpinBox(parent=self.groupBox_3)
+        self.offsetBox.setObjectName("offsetBox")
+        self.gridLayout.addWidget(self.offsetBox, 2, 3, 1, 1)
         self.verticalLayout.addWidget(self.groupBox_3)
         spacerItem3 = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Policy.Minimum, QtWidgets.QSizePolicy.Policy.Expanding)
         self.verticalLayout.addItem(spacerItem3)
@@ -571,13 +574,14 @@ class Ui_SPODMR(object):
         SPODMR.setTabOrder(self.iqdelayBox, self.plotenableBox)
         SPODMR.setTabOrder(self.plotenableBox, self.plotmodeBox)
         SPODMR.setTabOrder(self.plotmodeBox, self.taumodeBox)
-        SPODMR.setTabOrder(self.taumodeBox, self.normalizeBox)
-        SPODMR.setTabOrder(self.normalizeBox, self.xlogBox)
+        SPODMR.setTabOrder(self.taumodeBox, self.flipYBox)
+        SPODMR.setTabOrder(self.flipYBox, self.normalizeBox)
+        SPODMR.setTabOrder(self.normalizeBox, self.offsetBox)
+        SPODMR.setTabOrder(self.offsetBox, self.xlogBox)
         SPODMR.setTabOrder(self.xlogBox, self.ylogBox)
         SPODMR.setTabOrder(self.ylogBox, self.fftBox)
         SPODMR.setTabOrder(self.fftBox, self.complexBox)
-        SPODMR.setTabOrder(self.complexBox, self.flipYBox)
-        SPODMR.setTabOrder(self.flipYBox, self.fg_disableButton)
+        SPODMR.setTabOrder(self.complexBox, self.fg_disableButton)
         SPODMR.setTabOrder(self.fg_disableButton, self.fg_cwButton)
         SPODMR.setTabOrder(self.fg_cwButton, self.fg_gateButton)
[…]
