### 2023-10-17 bef7d77
podmr/spodmr: rename plot param xlogscale -> logX
- ylogscale -> logY too.
- data version updates: PODMRData to v3, SPODMRData to v1.
