+++ mahos/gui/ui/spodmr.py
@@ -419,12 +419,12 @@ class Ui_SPODMR(object):
         self.label_8 = QtWidgets.QLabel(parent=self.groupBox_3)
         self.label_8.setObjectName("label_8")
         self.gridLayout.addWidget(self.label_8, 1, 0, 1, 1)
-        self.ylogBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
-        self.ylogBox.setObjectName("ylogBox")
-        self.gridLayout.addWidget(self.ylogBox, 3, 2, 1, 1)
-        self.xlogBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
-        self.xlogBox.setObjectName("xlogBox")
-        self.gridLayout.addWidget(self.xlogBox, 3, 1, 1, 1)
+        self.logYBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
+        self.logYBox.setObjectName("logYBox")
+        self.gridLayout.addWidget(self.logYBox, 3, 2, 1, 1)
+        self.logXBox = QtWidgets.QCheckBox(parent=self.groupBox_3)
+        self.logXBox.setObjectName("logXBox")
+        self.gridLayout.addWidget(self.logXBox, 3, 1, 1, 1)
         self.label_2 = QtWidgets.QLabel(parent=self.groupBox_3)
         self.label_2.setObjectName("label_2")
         self.gridLayout.addWidget(self.label_2, 1, 2, 1, 1)
@@ -577,9 +577,9 @@ class Ui_SPODMR(object):
         SPODMR.setTabOrder(self.taumodeBox, self.flipYBox)
         SPODMR.setTabOrder(self.flipYBox, self.normalizeBox)
         SPODMR.setTabOrder(self.normalizeBox, self.offsetBox)
-        SPODMR.setTabOrder(self.offsetBox, self.xlogBox)
-        SPODMR.setTabOrder(self.xlogBox, self.ylogBox)
-        SPODMR.setTabOrder(self.ylogBox, self.fftBox)
+        SPODMR.setTabOrder(self.offsetBox, self.logXBox)
+        SPODMR.setTabOrder(self.logXBox, self.logYBox)
+        SPODMR.setTabOrder(self.logYBox, self.fftBox)
         SPODMR.setTabOrder(self.fftBox, self.complexBox)
         SPODMR.setTabOrder(self.complexBox, self.fg_disableButton)
         SPODMR.setTabOrder(self.fg_disableButton, self.fg_cwButton)
@@ -707,8 +707,8 @@ class Ui_SPODMR(object):
         self.plotmodeBox.setItemText(6, _translate("SPODMR", "normalize1"))
         self.plotmodeBox.setItemText(7, _translate("SPODMR", "concatenate"))
         self.label_8.setText(_translate("SPODMR", "mode"))
-        self.ylogBox.setText(_translate("SPODMR", "log Y"))
-        self.xlogBox.setText(_translate("SPODMR", "log X"))
+        self.logYBox.setText(_translate("SPODMR", "log Y"))
+        self.logXBox.setText(_translate("SPODMR", "log X"))
         self.label_2.setText(_translate("SPODMR", "tau mode (x axis)"))
         self.normalizeBox.setText(_translate("SPODMR", "normalize"))
         self.flipYBox.setText(_translate("SPODMR", "flip Y"))
