### 2023-10-17 4629b69
SPODMRData: add method measurement_time()
- store pulse pattern length for that.
