### 2023-10-17 15157bc
podmr_fitter: guess variable only if it varies
