### 2023-10-18 44eb46c
gui.podmr/spodmr: export png on save
