### 2023-10-18 #5938 @aquilesC
First, I want to congratulate @ktahar for a very good job. I myself have been working on these kind of approaches to creating distributed software control, but I think Mahos is an elegant solution, with clean code and well documented.

In going through the checklist, there are some items pertaining to the paper itself (and hence I don't know if creating an issue in the software repo is the best approach) that I think can be improved:

- [ ] Abstract: It is not clear what the software does. There is a generic description of how instruments can be controlled, but no reference to what MAHOS actually does.
- [ ] Statement of need: There is a description of the field and a good selection of other libraries on which MAHOS builds, but it is not clear why MAHOS is needed. What are the other libraris missing that this library addresses?
- [ ] And a reviewed statement of need can also be reflected in the documentation (or the other way around. I believe the documentation makes a clearer case than the paper in that regard).
