### 2023-10-19 6d8a8b7
docs: update paper to stress "Statement of need"
- Reflecting the paper review by @aquilesC
- Updated Overview of the docs accordingly
