+++ misc/paper/paper.bib
@@ -34,7 +34,7 @@
 }
 
 @article{wrachtrup2006,
-  title = {Processing Quantum Information in Diamond},
+  title = {{Processing Quantum Information in Diamond}},
   author = {Wrachtrup, J{\"o}rg and Jelezko, Fedor},
   year = {2006},
   month = may,
@@ -57,13 +57,13 @@
   number = {84},
   pages = {5304},
   author = {Hernán E. Grecco and Matthieu C. Dartiailh and Gregor Thalhammer-Thurner and Torsten Bronger and Florian Bauer},
-  title = {PyVISA: the Python instrumentation package},
+  title = {{PyVISA: the Python instrumentation package}},
   journal = {Journal of Open Source Software}
 }
 
 @article{qudi,
   author = {Jan M. Binder and Alexander Stark and Nikolas Tomek and Jochen Scheuer and Florian Frank and Kay D. Jahnke and Christoph Müller and Simon Schmitt and Mathias H. Metsch and Thomas Unden and Tobias Gehring and Alexander Huck and Ulrik L. Andersen and Lachlan J. Rogers and Fedor Jelezko},
-  title = {Qudi: A modular python suite for experiment control and data processing},
+  title = {{Qudi: A modular python suite for experiment control and data processing}},
   journal = {SoftwareX},
   year = 2017,
   volume = {6},
@@ -116,4 +116,19 @@
   year = 2020,
   version = {v0.8},
   doi = {10.5281/zenodo.3732545},
+}
+
+@misc{nspyre,
+  author = {Jacob Feder and
+            Alexandre Bourassa and
+            mtsolmn and
+            bensoloway and
+            jamendez99 and
+            mendezUC and
+            Aidan Jones},
+  title = {{nspyre-org/nspyre}},
+  publisher = {Zenodo},
+  year = 2023,
+  version = {v0.6.1},
+  doi = {10.5281/zenodo.8396120},
 }
