### 2023-10-19 #5938 @ktahar
Hi @aquilesC. Thank you so much for your review. I'm glad you liked MAHOS.

I tried to improve the paper according to your comments, by stressing reason why mahos is created / difference from existing libraries.
The key point is MAHOS provides a framework for distributed system but the existing libraries provide rather monolithic ones.
Please let me know if this point looks still unclear in the paper.

The overview section in the documentation is updated accordingly:
https://toyotacrdl.github.io/mahos/overview.html#what-is-this

May I regenerate the paper PDF here ? :
