### 2023-10-25 41ad7da
workflow: add python 3.11 in test matrix
