### 2023-10-25 c716cbd
requirements: update scipy-stack version
- QdyneAnalyzer is updated to avoid warnings
