+++ requirements.txt
@@ -1,9 +1,9 @@
 # scipy stack
-numpy ~= 1.24.4
-scipy ~= 1.10.1
+numpy ~= 1.26.1
+scipy ~= 1.11.3
 matplotlib ~= 3.7.2
-pandas ~= 1.3.5
-ipython ~= 8.12.2
+pandas ~= 2.1.1
+ipython ~= 8.16.1
 # data analysis
 scikit-learn ~= 1.3.0
 lmfit ~= 1.1.0
