+++ requirements.txt
@@ -1,8 +1,8 @@
 # scipy stack
-numpy ~= 1.26.1
-scipy ~= 1.11.3
-matplotlib ~= 3.7.2
-pandas ~= 2.1.1
+numpy ~= 1.24.4
+scipy ~= 1.10.1
+matplotlib ~= 3.7.3
+pandas ~= 2.0.3
 ipython ~= 8.16.1
 # data analysis
 scikit-learn ~= 1.3.0
