+++ requirements.txt
@@ -3,7 +3,7 @@ numpy ~= 1.24.4
 scipy ~= 1.10.1
 matplotlib ~= 3.7.3
 pandas ~= 2.0.3
-ipython ~= 8.16.1
+ipython ~= 8.12.3
 # data analysis
 scikit-learn ~= 1.3.0
 lmfit ~= 1.1.0
