### 2023-10-25 #5938 @aquilesC
Thanks for the update, @ktahar . I think the manuscript is moving in the proper direction, but there are some things that can still be clarified.

In the abstract, the sentence: `It would become very12 difficult to write debuggable, testable, and maintainable programs on a monolithic framework` is quite a strong opinion. Many frameworks are based on the *monolithic* approach and deliver testable, debuggable and extensible programs. I think it may be valuable to focus on the strengths of MAHOS and who exactly will benefit.

In the statement of need, the sentence: " and talk26 to each other" can be further expanded. How is this achieved? What decisions were made? For example, why selecting ZeroMQ over other message brokers.

I believe one of the strengths of this approach is that the data generated by a node is accessible by any other program implementing the same ZeroMQ framework. This allows to write a control software in Python and visualization software in any other language. It even allows to write measurements in a Jupyter notebook and keep them documented. However, I can't gather this behavior from the manuscript, and is hidden below the layer of the 'distributed' sentence.
