### 2023-10-26 e0b731d
podmr_fitter: add T1 fitter
