### 2023-10-27 b842a0b
docs: update the paper to increase "need" contents
- many thanks to @aquilesC and @sidihamady for review
