+++ misc/paper/paper.md
@@ -21,48 +21,76 @@ bibliography: paper.bib
 # Summary
 
 The experimental researchers often face challenges building up the systems
-to automate their measurement procedures
-in physical science and engineering, or related areas.
-While simple automation could be done by a single script,
-systematic orchestration is required to handle complex cases.
+to automate their measurement procedures in physical science and engineering, or related areas.
+While simple automation could be done by a single script, systematic orchestration is required to handle complex cases.
 The measurement automation typically requires several pieces of programs in diverse layers:
 1) low-level drivers to communicate with the instruments,
 2) high-level measurement logics or analysis algorithms, and
 3) graphical user interfaces (GUIs) or top-level automation scripts.
-The system may also involve multiple computers.
-It would become very difficult to write debuggable, testable, and maintainable programs
-on a monolithic framework.
-To solve this kind of problem in measurement automation,
-`MAHOS` provides a flexible framework to build distributed systems consisting of small independent programs.
+
+A `modular` framework is necessary to write maintainable programs for this purpose.
+On such a framework, the pieces of codes are organized within high-level abstraction mechanisms (module, class, etc.)
+and the interfaces are cleanly defined.
+
+The `distributed` messaging is one of the modern approaches for building extensible computing or control system.
+In this approach, individual programs run on multiple computers and they exchange the data by using messaging mechanism.
+Each program works in concert with the other to realize the system's functions.
+
+Although there are several `modular` frameworks for laboratory automation,
+the `distributed` messaging approach is not widely recognized and tested in this field.
+`MAHOS` tries to bring these two concepts together and
+provide a `modular` and `distributed` framework for measurement automation.
 
 # Statement of need
 
-`MAHOS` is a Python framework for building distributed measurement systems.
+`MAHOS` is a `modular` and `distributed` framework for building measurement systems in Python.
 Python has been a major language of choice for laboratory automation.
 The low-level binding libraries for instruments, such as PyVISA [@pyvisa] and PyDAQmx [@pydaqmx], are available for Python.
-There are also several projects aiming to provide a comprehensive software framework
+There are also several projects aiming to provide the comprehensive `modular` framework
 ranging from the instrument drivers to the GUIs [@pymeasure; @qudi; @nspyre].
+While we believe that the `distributed` messaging could result in more extensible and flexible systems,
+the existing libraries take rather centralized approaches.
+That is why we have built `MAHOS` to provide both `modular` and `distributed` framework for measurement automation.
 
-While we believe that a distributed framework would result in a more maintainable and flexible system,
-the existing libraries provide rather monolithic ones.
-That is why we have built `MAHOS` to provide a flexible framework for constructing distributed systems.
-On the `MAHOS`-based system, several pieces of the programs are run on different processes (or threads)
-and talk to each other.
-Due to this architecture, it is quite straightforward to set up multi-computer systems.
-Moving from GUI operation to scripted automation is made seamless as well;
-the GUI and top-level script (3rd layer) are seen as equivalent clients from a measurement logic (2nd layer).
-
-Several concepts are inspired by the ROS project [@ros].
-Each piece of the program is called a `node`.
-Two patterns of communication methods are provided: Request-Reply (synchronous remote procedure call)
+On the `MAHOS`, several programs run as different processes
+and exchange the data by using a distributed messaging system.
+Two distributed system concepts are inspired by the ROS project [@ros].
+First, each piece of the program is called a `node`.
+Second, two patterns of communication methods are provided for the nodes: Request-Reply (synchronous remote procedure call)
 and Publish-Subscribe (asynchronous data distribution).
-However, contrary to the ROS, the `MAHOS` library is kept rather small and simple.
[…]
