+++ misc/paper/paper.bib
@@ -1,10 +1,17 @@
 @misc{pydaqmx,
   author = {Pierre Cladé},
   title = {PyDAQmx},
-  year = {2021},
+  year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/clade/PyDAQmx}
+}
+
+@misc{zeromq,
+  author = {{The ZeroMQ authors}},
+  title = {{ZeroMQ: An open-source universal messaging library}},
+  year = {2023},
+  url = {https://zeromq.org/}
 }
 
 @article{gruber1997,
