+++ docs/_sources/arch_node.rst.txt
@@ -1,18 +1,34 @@
 Node, Client, and Communication
 ===============================
 
-MAHOS helps to build `distributed systems`, which consists of multiple programs (processes) communicating each other.
+MAHOS helps to build `distributed` measurement automation systems, which consist of multiple programs (processes) communicating with each other.
 We call these programs :term:`nodes <node>` (:class:`Node <mahos.node.node.Node>`).
 A node usually provides some services for other nodes.
-The :term:`node client` (:class:`NodeClient <mahos.node.client.NodeClient>`) provides the standard way to access node's functions.
-The core functionalities of node systems are implemented in :ref:`mahos.node` package.
+The :term:`node client` (:class:`NodeClient <mahos.node.client.NodeClient>`) provides the standard way to access the node's functions.
+The core functionalities of node systems are implemented in the :ref:`mahos.node` package.
 
-The node :class:`communication <mahos.node.comm.Context>` is currently based on `ZeroMQ <https://zeromq.org/>`_ library.
+Data transport and serialization
+--------------------------------
+
+The node :class:`communication <mahos.node.comm.Context>` is currently based on the `ZeroMQ <https://zeromq.org/>`_ library.
 The transport layer is abstracted by ZeroMQ, and we can choose different transports by changing a configuration (the endpoint) only.
-However, we assume TCP unless otherwise specified.
+We recommend using TCP as the default.
 By choosing TCP, multi-host (multi-computer) configuration is quite straightforward.
-ZeroMQ doesn't specify the data serialization format; it provides methods to send / receive byte array of arbitrary length.
-We adopt Python standard pickle for serialization.
+However, the transportation overhead can become an issue when one deals with very large data
+such as high-resolution images produced at a high rate.
+In such a case, the overhead can be reduced significantly by running the relevant nodes as threads
+in a single process and using intra-process transportation.
+
+ZeroMQ doesn't specify the data serialization format; it provides methods to send/receive byte arrays of arbitrary length.
+Currently, we define the message types (classes) in the :ref:`mahos.msgs` package and serialize the instances
+using Python standard pickle.
+This approach is adopted due to pickle's high compatibility with Python objects and moderate performance.
+
+.. note::
+
+   The pickle-based serialization practically limits the messaging within Python only.
+   This might be changed to another approach (e.g., using Protocol Buffers)
+   in the future to support other programming languages.
 
 Communication patterns
 ----------------------
@@ -24,30 +40,30 @@ Req-Rep
 
 :term:`Req-Rep` (Request-Reply) pattern is synchronous :term:`RPC`.
 First, the client sends a request to the server (node).
-The server node does something accordingly, and send back a reply to the client.
+The server node does something accordingly and sends back a reply to the client.
 
 Pub-Sub
 ^^^^^^^
 
 :term:`Pub-Sub` (Publish-Subscribe) pattern is asynchronous, one-to-many data distribution.
-The message regarding a topic is published by single node (Publisher).
+The message regarding a topic is published by a single node (Publisher).
 Here, the topic is a label to distinguish the data type.
-Publisher can publish multiple topics from single endpoint.
+A publisher can publish multiple topics from a single endpoint.
 The subscriber which `subscribes to` a topic receives the messages.
 The number of subscribers can be changed dynamically,
-and zero subscriber is a valid state too (message is just not sent to anywhere in this case).
+and zero subscribers is a valid state too (the message is just not sent anywhere in this case).
 
 Client usage
 ------------
 
 It can be said that a node `provides` functions via Rep and Pub;
-corresponding client uses them via Req and Sub.
[…]

[…]
