### 2023-10-27 #5938 @ktahar
Dear @sidihamady and @aquilesC ,

Thank you so much for your comments and suggestions.
I fixed the paper and [introduction of doc](https://toyotacrdl.github.io/mahos/overview.html) clarify the difference between "modular" and "distributed".
- "Modular" is a very important aspect that makes the software maintainable, and MAHOS also tries to be like this.
But, as @aquilesC pointed out, this point is also realized in the existing libraries too.
- "Distributed" approach makes MAHOS unique. I added explanations of why this could be benefitial: flexible multi-host configuration and data / service accessibility. The latter contains what @aquilesC has suggested.
- Added short discussion on possible downside of "distributed" approach: overhead of data transfer. Moving to intra-process configuration can reduce this. And the ability to do so is one of the reasons to select ZeroMQ.

I hope these changes could remove the ambiguities and answer your comments properly.

To @sidihamady :
> You are the only author of the manuscript and the only committer on github. I would like to know if others helped in the development process?

For now, I am only significant contributor to this code base.
(Some implementations of confocal microscope / ODMR experiments are based on that of the predecessor project developed at Tokyo Institute of Technology. But the most part of that was written by me. The professors at Tokyo Tech approved to publish MAHOS in current form as single-author project.)

>  Could you tell us if the code management was done previously using another version control system (public or private)?

Yes, privately. As commented in first commit https://github.com/ToyotaCRDL/mahos/commit/b24a2081751c976306519f6c565fd4398b9a7339 ,
I re-initialized the repository just before making this public, to avoid exposing internal information of the company.
We had 789 commits since 2022-01-19 in older repo.

> And if the software was used and then cited in published works?

It has been used for [this work](https://dx.doi.org/10.1063/5.0166745), but has not been cited there (because this paper had not been published yet).

To @aquilesC :

> This allows to write a control software in Python and visualization software in any other language.

This point is going to be interesting and I actually considered (multiple times) to enable this. In current version, it is a little hard to use the language other than Python because the data is serialized using Python standard pickle.
But, it will not be too hard to move on to other serialization library with multi-language support  (such as Protocol Buffers) in future version.
I have added a note in the documentation here:
https://toyotacrdl.github.io/mahos/arch_node.html#data-transport-and-serialization
