### 2023-10-29 #5938 @sidihamady
Dear @ktahar,

Thank you for these clarifications.

(i) You could perhaps add a header to the source files specifying the type of copyright (BSD) and referring to the LICENSE file.

(ii) The documentation is well done and installation is simple (at least under my systems: Debian and AlmaLinux). It might be necessary to add an install-mahos.sh installation script but users can also do that.
But there are 2 issues (that I added to github) : (i) user should update PATH for virtualenv ; (ii) install python C dev package. After these issues were resolved, testing was successful.

Regards,
Sidi
