### 2023-10-30 6b67ad1
spodmr: fix bug regarding block_base
- block_base was passed to generate_raw_blocks() of PODMR's
  generators. This cound result in short laser pattern at starting block.
  Then, the laser pattern was enlarged to fill the 'residual' period.
- Because We'd like to fill the 'residual' period with (almost) NOP,
  fix_block_base of 1 is passed to generate_raw_blocks() and
  tweaks of generated blocks for block_base are done in BlockSeqBuilder.
