### 2023-10-31 77cc4f8
gui/main_monitor: avoid masking 'name' out
