### 2023-10-31 66c0e55
docs: add notes on Python3 dev in installation
- On Linux, we have to install dev-package for Python3
  to compile C extension.
- fix #3. thanks @sidihamady.
