### 2023-10-31 d5bbac1
docs/install: add "add PATH" note for virtualenv
- fix #2. thanks @sidihamady for pointing this out.
