### 2023-10-31 c7f34ea
docs: add license notes at module docstring
- thanks @sidihamady for the suggestion.
