+++ mahos/cli/launch.py
@@ -1,5 +1,13 @@
 #!/usr/bin/env python3
 
+"""
+mahos launch command.
+
+.. This file is a part of MAHOS project, which is released under the 3-Clause BSD license.
+.. See included LICENSE file or https://github.com/ToyotaCRDL/mahos/blob/main/LICENSE for details.
+
+"""
+
 from __future__ import annotations
 import time
 import importlib
