+++ mahos/meas/iodmr_io.py
@@ -3,7 +3,8 @@
 """
 File I/O for Imaging ODMR.
 
-.. This file is a part of MAHOS project.
+.. This file is a part of MAHOS project, which is released under the 3-Clause BSD license.
+.. See included LICENSE file or https://github.com/ToyotaCRDL/mahos/blob/main/LICENSE for details.
 
 """
 
