+++ tests/test_log_broker.py
@@ -1,5 +1,13 @@
 #!/usr/bin/env python3
 
+"""
+Tests for LogBroker.
+
+.. This file is a part of MAHOS project, which is released under the 3-Clause BSD license.
+.. See included LICENSE file or https://github.com/ToyotaCRDL/mahos/blob/main/LICENSE for details.
+
+"""
+
 import time
 
 from fixtures import ctx, gconf, log_broker, dummy_logging
