+++ mahos/util/typing.py
@@ -1,5 +1,13 @@
 #!/usr/bin/env python3
 
+"""
+Type hints for MAHOS.
+
+.. This file is a part of MAHOS project, which is released under the 3-Clause BSD license.
+.. See included LICENSE file or https://github.com/ToyotaCRDL/mahos/blob/main/LICENSE for details.
+
+"""
+
 from ..msgs.common_msgs import Message, Request, Resp
 
 from typing import NewType, Callable, Union, Tuple
