### 2023-10-31 #5938 @ktahar
Dear @sidihamady ,

Thank you for good suggestions.
I added license notes at the header of source files, and fixed the installation document.

Thanks for the suggestion regarding the installation script too.
But I would prefer not adding it at the moment due to following reasons:

- Assuming the script would contain package installations (sudo apt install... or pip install...), it will be good for some users (who understands what's done by that), but it might be a little unsafe for the others. It can break their computer environment.
- Using virtualenv is my preference and recommendation but this is not a must. Different tools (anaconda etc.) can be used too. The scripted installation will enforce only one way.
- Maintaining the script working for all the possible environments would be rather difficult. (For me, it's going to be quite hard on Windows.)

Best regards,
Kosuke
