### 2023-10-31 #5938 @sidihamady
Dear @ktahar,

The end user can create an installation script himself and adapt it to his platform (Linux I mean); No problem :)

The code is particularly well written and even if the documentation in the code files is not exhaustive, its clarity makes it easy to understand.

Regards,
Sidi
