### 2023-11-01 cdb2079
fix minor bugs
