### 2023-11-01 76f4320
podmr,spodmr: add param flip_head for T1
