+++ mahos/gui/podmr.py
@@ -727,6 +727,7 @@ class PODMRWidget(ClientWidget, Ui_PODMR):
             ("invertY", self.invertYBox),
             ("readY", self.readYBox),
             ("reinitX", self.reinitXBox),
+            ("flip_head", self.flipheadBox),
         ]
         set_enabled(params, name_widgets)
 
@@ -913,6 +914,7 @@ class PODMRWidget(ClientWidget, Ui_PODMR):
         self.invertYBox.setChecked(p.get("invertY", False))
         self.reinitXBox.setChecked(p.get("reinitX", False))
         self.readYBox.setChecked(p.get("readY", False))
+        self.flipheadBox.setChecked(p.get("flip_head", False))
         self.nomwBox.setChecked(p.get("nomw", False))
         self.invertinitBox.setChecked(p.get("invertinit", False))
         self.reduceBox.setChecked(p.get("enable_reduce", False))
@@ -1018,6 +1020,7 @@ class PODMRWidget(ClientWidget, Ui_PODMR):
         params["invertY"] = self.invertYBox.isChecked()
         params["reinitX"] = self.reinitXBox.isChecked()
         params["readY"] = self.readYBox.isChecked()
+        params["flip_head"] = self.flipheadBox.isChecked()
         params["partial"] = self.partialBox.currentIndex() - 1
         params["nomw"] = self.nomwBox.isChecked()
         params["invertinit"] = self.invertinitBox.isChecked()
@@ -1233,6 +1236,7 @@ class PODMRWidget(ClientWidget, Ui_PODMR):
                 self.invertYBox,
                 self.reinitXBox,
                 self.readYBox,
+                self.flipheadBox,
                 self.invertinitBox,
                 self.iqdelayBox,
                 self.tpwidthBox,
