+++ tests/test_podmr.py
@@ -72,6 +72,7 @@ def test_podmr_pattern_divide():
         "readY": False,
         "invertinit": False,
         "reinitX": False,
+        "flip_head": True,
         "start": 100e-9,
         "num": 2,
         "step": 100e-9,
@@ -117,6 +118,7 @@ def test_podmr_patterns():
         "readY": False,
         "invertinit": False,
         "reinitX": False,
+        "flip_head": True,
         "start": 100e-9,
         "num": 2,
         "step": 100e-9,
