### 2023-11-02 db4e9f9
examples/cfm: fix renamed key
