### 2023-11-03 3372c88
gui.spodmr: fix spacer's SizePolicy
- don't know why this was set QSizePolicy::Minimum
