### 2023-11-03 f682628
spodmr: plot std dev
