### 2023-11-03 80fb884
odmr: plot std dev
