### 2023-11-04 #5938 @aquilesC
Dear @ktahar and @arfon , I am done with the review of the paper, I think is ready for publication. It is an interesting piece of software and I am curious to see how it is going to develop further. I have always been a fan of ZMQ for building flexible acquisition software, and MAHOS is a very good example of what can be achieved.
