### 2023-11-04 3b8fd80
allow custom serialization as Message methods
