### 2023-11-06 ae906e3
gui.spectroscopy: raise on init failure
