### 2023-11-06 06db23d
again, fix examples and add docs
