### 2023-11-06 05898cf
gui/confocal: put XY, XZ, YZ in individual layouts
