+++ mahos/gui/ui/confocal.py
@@ -12,7 +12,7 @@ from PyQt6 import QtCore, QtGui, QtWidgets
 class Ui_Confocal(object):
     def setupUi(self, Confocal):
         Confocal.setObjectName("Confocal")
-        Confocal.resize(1630, 639)
+        Confocal.resize(1630, 753)
         self.verticalLayout = QtWidgets.QVBoxLayout(Confocal)
         self.verticalLayout.setContentsMargins(0, 0, 0, 0)
         self.verticalLayout.setSpacing(0)
@@ -27,8 +27,12 @@ class Ui_Confocal(object):
         self.splitter.setObjectName("splitter")
         self.histo_XY = HistogramLUTWidget(parent=self.splitter)
         self.histo_XY.setObjectName("histo_XY")
-        self.graphicsView = GraphicsView(parent=self.splitter)
-        self.graphicsView.setObjectName("graphicsView")
+        self.graphicsView_XY = GraphicsView(parent=self.splitter)
+        self.graphicsView_XY.setObjectName("graphicsView_XY")
+        self.graphicsView_XZ = GraphicsView(parent=self.splitter)
+        self.graphicsView_XZ.setObjectName("graphicsView_XZ")
+        self.graphicsView_YZ = GraphicsView(parent=self.splitter)
+        self.graphicsView_YZ.setObjectName("graphicsView_YZ")
         self.histo_XZ = HistogramLUTWidget(parent=self.splitter)
         self.histo_XZ.setObjectName("histo_XZ")
         self.histo_YZ = HistogramLUTWidget(parent=self.splitter)
@@ -150,12 +154,12 @@ class Ui_Confocal(object):
         self.yzfitButton = QtWidgets.QPushButton(parent=Confocal)
         self.yzfitButton.setObjectName("yzfitButton")
         self.horizontalLayout_3.addWidget(self.yzfitButton)
-        self.yzclearButton = QtWidgets.QPushButton(parent=Confocal)
-        self.yzclearButton.setObjectName("yzclearButton")
-        self.horizontalLayout_3.addWidget(self.yzclearButton)
         self.yzmaxButton = QtWidgets.QPushButton(parent=Confocal)
         self.yzmaxButton.setObjectName("yzmaxButton")
         self.horizontalLayout_3.addWidget(self.yzmaxButton)
+        self.yzclearButton = QtWidgets.QPushButton(parent=Confocal)
+        self.yzclearButton.setObjectName("yzclearButton")
+        self.horizontalLayout_3.addWidget(self.yzclearButton)
         self.yzpopButton = QtWidgets.QPushButton(parent=Confocal)
         self.yzpopButton.setObjectName("yzpopButton")
         self.horizontalLayout_3.addWidget(self.yzpopButton)
@@ -200,8 +204,10 @@ class Ui_Confocal(object):
 
         self.retranslateUi(Confocal)
         QtCore.QMetaObject.connectSlotsByName(Confocal)
-        Confocal.setTabOrder(self.histo_XY, self.graphicsView)
-        Confocal.setTabOrder(self.graphicsView, self.histo_XZ)
+        Confocal.setTabOrder(self.histo_XY, self.graphicsView_XY)
+        Confocal.setTabOrder(self.graphicsView_XY, self.graphicsView_XZ)
+        Confocal.setTabOrder(self.graphicsView_XZ, self.graphicsView_YZ)
+        Confocal.setTabOrder(self.graphicsView_YZ, self.histo_XZ)
         Confocal.setTabOrder(self.histo_XZ, self.histo_YZ)
         Confocal.setTabOrder(self.histo_YZ, self.idleButton)
         Confocal.setTabOrder(self.idleButton, self.piezoButton)
@@ -231,9 +237,9 @@ class Ui_Confocal(object):
[…]
