### 2023-11-06 fe8ad08
make dirs for log etc. configurable
- default dir is changed too
