### 2023-11-06 18c46ff
gui/confocal: fix bug of tracker
