+++ mahos/gui/confocal.py
@@ -414,8 +414,6 @@ class trackDialog(QtWidgets.QDialog, Ui_trackDialog):
         apply_widgets(
             params,
             [
-                ("xnum", self.xnumBox),
-                ("ynum", self.ynumBox),
                 ("delay", self.delayBox, 1e3),
                 ("time_window", self.timeBox, 1e3),
                 ("dummy_samples", self.dummysampleBox),
