### 2023-11-06 f3d6dad
add examples using custom serializer
