### 2023-11-06 4a6fff9
docs: update docs for custom serializer
