+++ docs/_sources/meas/generated/mahos.meas.podmr.PODMR.rst.txt
@@ -35,4 +35,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
