+++ docs/_sources/msgs/common_msgs.rst.txt
@@ -8,6 +8,10 @@
 .. rubric:: Classes
 
 .. autoclass:: Message
+
+   .. automethod:: Message.serialize
+   .. automethod:: Message.deserialize
+
 .. autoclass:: Request
 .. autoclass:: Reply
 .. autoclass:: State
