+++ docs/_sources/meas/generated/mahos.meas.confocal.Confocal.rst.txt
@@ -41,4 +41,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
