+++ docs/_sources/meas/generated/mahos.meas.odmr.ODMR.rst.txt
@@ -33,4 +33,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
