+++ docs/_sources/meas/generated/mahos.meas.spodmr.SPODMR.rst.txt
@@ -34,4 +34,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
