+++ docs/_sources/meas/generated/mahos.meas.iodmr.IODMR.rst.txt
@@ -30,4 +30,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
