+++ docs/_sources/meas/generated/mahos.meas.confocal_tracker.ConfocalTracker.rst.txt
@@ -40,4 +40,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
