+++ docs/_sources/meas/generated/mahos.meas.state_manager.StateManager.rst.txt
@@ -29,4 +29,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
