+++ docs/_sources/node/generated/mahos.node.log_broker.LogBroker.rst.txt
@@ -24,4 +24,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
