+++ docs/_sources/node/generated/mahos.node.global_params.GlobalParams.rst.txt
@@ -25,4 +25,9 @@
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+   
    
