### 2023-11-06 efd98d6
docs: fix the paper a little
