+++ docs_src/conf.py
@@ -22,7 +22,7 @@ copyright = "2022, Toyota Central R&D Labs., Inc"
 author = "Kosuke Tahara"
 
 # The full version, including alpha/beta/rc tags
-release = "0.2.2"
+release = "0.2.3"
 
 
 # -- General configuration ---------------------------------------------------
