+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Changelog
 [Unreleased]
 ------------
 
+[0.2.3] - 2023-11-06
+--------------------
+
 Added
 ^^^^^
 
