### 2023-11-06 #5938 @ktahar
@aquilesC and @sidihamady : Thank you again for the reviews. Your comments and suggestions were really helpful.

@arfon : Thanks for quickly managing this.
I've fixed paper a little, made a release on github https://github.com/ToyotaCRDL/mahos/releases/tag/v0.2.3 ,
and archived it on zenodo https://zenodo.org/records/10074373 .

DOI: 10.5281/zenodo.10074373
