### 2023-11-06 #5938 @ktahar
@aquilesC - After discussion above, I've added a (rather experimental) capability to enable custom serializer for foreign language interfacing. Please check out [examples/custom_serializer](https://github.com/ToyotaCRDL/mahos/tree/main/examples/custom_serializer) if you're interested. The note [here](https://toyotacrdl.github.io/mahos/arch_node.html#data-transport-and-serialization) is updated too.
