### 2023-11-06 #5938 @aquilesC
@ktahar , that was very quick! The example is simple but it shows the power of interfacing across languages. I'm sure this pattern is very powerful in the long run (for instance to leverage µManager, or WebUI technologies.) Looking forward to seeing how this evolves!
