### 2023-11-07 4619124
InstrumentServer: try-except init of overlays
