### 2023-11-08 3509cc5
meas.confocal: limit image pub rate when finished
