### 2023-11-09 ac5ca8b
podmr_fitter.RabiFitter: guess negative A too
