### 2023-11-10 ce13616
daq.AnalogIn: set data transfer method
